Bug 383252: Cannot drag / drop URL or link onto tabbar, r=mconnor
When SeaMonkey switched to toolkit's nsDragAndDrop.js, it lost the dragDropSecurityCheck method, which for SM was on nsDragAndDrop, but for FF was on tabbrowser.

Moving that method from tabbrowser to toolkit's nsDragAndDrop.js, and cleaning it up a little.

UltraBlame original commit: 3c07e517696dddc1dada0a71e4ac657a7fd7c872