Bug 97954: Don't cite Makefile variables in the value for LIBXUL_DIST. r=bsmedberg
Background:

Thread-safe SpiderMonkey needs NSPR for threading support, so it uses
the values of NSPR_CFLAGS and NSPR_LIBS to compile and link.  The
configure script puts values for those variables in
$(DEPTH)/config/autoconf.mk.  They're recursively expanded Make
variables, and the configure script gives them values that cite the
value of LIBXUL_DIST.

LIBXUL_DIST is itself a recursively expanded variable which usually
refers to $(DIST), but refers to $(LIBXUL_SDK) if we're building
against a separate XUL tree.

Once SpiderMonkey has its own configure script, that script should
take options specifying values to use for NSPR_CFLAGS and NSPR_LIBS.
The values we pass to it shouldn't try to reference $(LIBXUL_DIST) or
any other variable from the surrounding makefiles: a stand-alone
Spidermonkey's configuration and build system shouldn't mention LIBXUL
at all.  SpiderMonkey doesn't depend on LIBXUL, only NSPR, which just
happens to be installed in the same dist directory as LIBXUL.

That means we need to compute fully-substituted values for NSPR_CFLAGS
and NSPR_LIBS, suitable for stand-alone use in the top-level configure
script, which we can pass down to the SpiderMonkey configure script
(once it has one).

The first step is to actually compute an independent value for
LIBXUL_DIST, one that doesn't refer to either $(DIST) or
$(LIBXUL_SDK).  This means we need to make that choice in the
configuration script, not in autoconf.mk.  Subsequent patches will use
this to compute fully-substituted values for NSPR_LIBS and NSPR_CFLAGS.

UltraBlame original commit: 4611778ba22a213f59020384f99a09cbcb2e0650