Bug 504920. Use cairo_set_can_convert_to_dib(TRUE) when creating optimized DDB image surfaces. r=vlad
This lets cairo convert and keep a DDB image into a DIB section when we paint
it to a canvas.  Avoiding the continuous conversion to an image surface
improves the performance of drawImage on to a canvas on Win32, in particular
the peacekeeper experimentalMovie benchmark.

UltraBlame original commit: ca46dd208b993575d10f2b62de4710dd33bb5cae