Correct handling of unitless zero in calc() expressions.  (Bug 595648)  r=bzbarsky  a=blocking2.0:betaN+
This ensures that we treat a unitless zero as a number rather than a
length inside of calc() expressions.  This causes both the acceptance of
previously-rejected expressions (such as -moz-calc(0 * 1em)) and the
rejection of previously-accepted ones (such as -moz-calc(0 + 1em)).

UltraBlame original commit: 97e0776740d37912dbcf49fd8e63038d2ad24b3a