Bug 625141 - Avoid overwriting used register in TypedArray IC (r=dvander)
There was an edge case when creating an IC to store to typed arrays, in which
we would overwrite a register which was a component of the address. Fix by
preserving the address in that case, and restoring it after the overwrite.

This also checks that we save/restore each register only once.

UltraBlame original commit: ea2b1b5e3837c0f98dd7a8ce381dbbf45f753b21