Bug 683802 - Force allocation for sized PWSTRINGs in XPCWrappedNative argument conversion. r=mrbkap
This is the last remaining place where we might pass a non-dipper with useAllocator==false to XPCConvert *and, as a consequence*, do something different within XPCConvert. This paves the way for the elimination of useAllocator for non-dipper types.

UltraBlame original commit: 30472afa1c400403b0dc8cc44a09d882fc1fb240