Bug 696672. Use a larger font size in bidi/613149-2* tests. r=smontagu
This currently fails on unaccelerated GDI because in the reference we render
all of the text in one call which causes the blending of the overlap between
the ':' and the 'M' to be different than if we render the strings in two steps.
Increasing the font size makes it so that the space between characters is large
enough to avoid this problem.

UltraBlame original commit: b455a7838d6d4cb4c5e6214b7ccd491a6c7a28e2