Bug 707033 - Kill the hashtables of WebGLUniformLocation's held by WebGLProgram's - r=jgilbert
Currently, WebGLProgram holds a hashtable of refptrs to all WebGLUniformLocation's created from it.

It's used for one thing: to ensure that multiple getUniformLocation() calls on same uniform return the same WebGLUniformLocation object.

However, in a discussion on the public_webgl mailing list this week, we agreed that this was not mandated by the spec and that on the contrary, to ensure consistent behavior across browsers, we're going to require that NOT to happen. So this becomes a conformance issue.

Removing this hashtable also simplifies code and ownership patterns.

UltraBlame original commit: 432d88a73914212f73ebf3fef49925e00aea1b78