Change css3-animations behavior for repeated keys in an keyframes rule:  do replacement on a per-property basis.  (Bug 738003)  r=bzbarsky
This implements my proposal in
http://lists.w3.org/Archives/Public/www-style/2011Apr/0381.html and
http://lists.w3.org/Archives/Public/www-style/2011Apr/0387.html .  I
think it was a serious mistake to implement what the spec says, and I'm
fixing that mistake so that we have a chance to change the spec.

In other words, when an keyframes rule has two key selectors at the
same time, the later one no longer overrides the entirety of the earlier
one.  The overriding is done for each property that's in the later rule.
(And the -moz-animation-timing-function is taken only from the keyframe
actually used for the given property; if there's no declaration there
then the computed value of the property is used.)

The test for keyframes cascade fails without the patch; the test for
keyframes cascade2 tests behavior that works both before and after the
patch.

UltraBlame original commit: e449da8cc3e1a0463712bf1b3e03262160578cf6