Bug 758415 - Implement expando object infrastructure for WN Xrays. r=mrbkap Note: This overloads the naming of some of the existing infrastructure, but the signatures etc are sufficient to disambiguate. The other infrastructure goes away in a subsequent patch.
Note: We tag sandbox expandos with their global to make sure that the expandos
are never shared between sandboxes. A consequence of this scheme is that an
expando from a sandbox to an object will _always_ result in a GC edge back to
the sandbox, meaning that the sandbox is always kept alive for the lifetime of
the expando target. This could happen before, but only if a non-primitive expando
was placed (since the value of the expando would live in the consumer's
compartment). We could avoid this edge by using a reference-counted Identity()
object instead, but I suspect it's not worth worrying about.

UltraBlame original commit: cfafb4e9cc349c6906569e8afd5e69c07463af4b