Bug 767488 - Cache the current process handle after reading it for the first time during stack walking; r=dbaron
The current process handle never changes, so it doesn't make a lot of sense
to read it over and over again.  Note that this doesn't close the process
handle explicitly but that's fine, since Windows will release all of the
open handles when the application is terminated.

UltraBlame original commit: 2cbf12886c8914b36127cdec5c93ea4884c7d6b1