Bug 776632 - Implement testing function getBuildConfiguration. r=luke
The fuzzers currently infer the build configuration based on the presence or absence of various flags, and gkw noted that this will not work with JSGC_ROOT_ANALYSIS, so I thought it'd be nice to add something more direct. This is a stub set for now.

UltraBlame original commit: 379789b07fb6d32ae26c289a6442e48b502d08fc