Implements Calls to C++ (bug 680315, r=dvander)
Add VMFunction structure to collect information about callable C functions.

Trampoline functions generate wrapper for the C functions which are cached
inside the IonCompartment and shared between the multiple insitances required by
the CodeGenerator.

The stack produced by the call to the wrapper of the C function is mapped by
IonCCallStack structure which can be used as an IonEnvironment to inspect the
top level IonFrame.

Additionnal instrumentation is added to ease manipulation of register and to
optimize call to C functions.  Additional functions are added to identify errors
from function which are returning a Value.

This currently support x86_64 and x86 architectures.

---

UltraBlame original commit: 691c944bb0258f5bab2942cba2b40a2ec6a9d3ae