Bug 792036 - Add a SpecialPowers API to provide access to the Components object in various forms. r=mrbkap,ted
Now that window.Components is no longer acccessible to page JS, we need a way
to access it in mochitests. So this patch provides SpecialPowers.Components,
which is the bonafide Components object for the window upon which SpecialPowers
is defined. It also provides SpecialPowers.{Cc,Ci,Cr,Cu}, which are SpecialPowers-wrapped
versions of their respective sub-objects (with the exception of Ci, which is unwrapped).

UltraBlame original commit: 1460d44ec8def5ee7d1d4ebb9534dcce652a0966