Bug 794051 - Read battery status via fgets. r=dhylands
The call to fscanf for reading the battery status used an incorrect pointer in its
arguments. The code only worked because the first byte of a character array is the
first element within the array. To fix this problem, fscanf has been replaced by
fgets, which is less error prone and checks the length of the supplied array.

Additionaly, we don't want to operate on an undefined result if fgets fails. In
this case we assume that the battery is not being charged.

UltraBlame original commit: 20169e3f33b350e251a8c228a140c5afa51f25e4