Bug 852565 - Don't expand the viewport for small pages. r=kats
Only expand the CSS viewport when a page reaches the screen size. If it's
smaller than the screen size, lock the dynamic toolbar and keep the same,
smaller CSS viewport. This 'fixes' sites that try to size themselves to the
size of the screen and get it wrong.

UltraBlame original commit: 74c06a7308d19bf5501dfcbf91f5c8093d52d1d6