Bug 860338 - remove MOZ_WEBVTT. r=ted
The webvtt dom implementation and parser are a small amount of code.
Given the tendency of build-time switches to bitrot, it's not worth
keeping this one.

The webvtt dom implementation and parser are a small amount of code.
Given the tendency of build-time switches to bitrot, it's not worth
keeping this one.

UltraBlame original commit: d757eeebcc43c6cb17b73daa5a3d8f55e873398c