Bug 860941 - Make dialogArguments default to |undefined| rather than |null|. r=jst
This is correct by my reading of the spec. Quoting:

The dialogArguments IDL attribute, on getting, must check whether its browsing
context's active document's origin is the same as the dialog arguments' origin.
If it is, then the browsing context's dialog arguments must be returned
unchanged. Otherwise, if the dialog arguments are an object, then the empty
string must be returned, and if the dialog arguments are not an object, then
the stringification of the dialog arguments must be returned.

UltraBlame original commit: af186224869402007bc54ad3aeeecdc0a9aa7fee