Bug 868130 - Move nsCxPusher into XPConnect. r=gabor
The JSContext stack is an XPConnect construction. In particular, there are
situations where we want to manipulate it outside the lifetime of nsContentUtils
but within the lifetime of the stack itself. In order to do this cleanly, it's
helpful to use private XPConnect APIs. So the first step here is to move this into
js/src/xpconnect, so that we can take advantage of the stuff in xpcprivate.h.

UltraBlame original commit: a8c7766cd507bcc30a332f4c58989d49b25d0419