bug 852129 - use HyperTextAccessible for invalid img r=surkov, bz
If the img is not valid then its children will be rendered, and the
sensible way to handle that is by giving the img a HyperTextAccessible
instead of an ImageAccessible.  Since the accessible name of such an img
should be the value of the alt attribute we add similar logic as
ImageAccessible::NativeName() to HyperTextAccessible::NativeName()
conditioned on the tag being img.

UltraBlame original commit: 5f243d3f45beae0ed65011357e2f4274a89dd164