Bug 884362 - Use AutoJSContext in nsScriptSecurityManager::LoadPolicy. r=bz
This cx is just used for rooting etc. As such, we really just want whatever
cx is on the stack, but want a default if the stack-top cx is null (which it
sometimes is, here). And that's exactly what AutoJSContext gives us.

UltraBlame original commit: 424742645e2e87e2465971025b4953a1967d5c8b