Bug 885621. Fix gfxPattern::GetMatrix() with azure and add GetInverseMatrix(). r=bas
Currently SetMatrix will take the Inverse of the passed in matrix but GetMatrix
doesn't not invert it on the way out. This fixes that and adds a
GetInverseMatrix() so that we don't double invert it when we need the inverse.

UltraBlame original commit: ec6f95743c5d777906bc4743d671500956eeacd2