Bug 889714 - Stash the XPCShellEnvironment instance on the global, rather than the cx. r=mrbkap
We need to stash it somewhere, because the shell-provided JSNatives need to
access it. This solution only works for script running in the scope of the
global we create. Butthat's fine here, because we only use it for load() and
quit(), which only exist in the scope of that global.

UltraBlame original commit: 35039a0a3316c1acb4f2cbdbf22b183ff83de06a