Bug 868896: Use REQUEST_GET_UNLOCK_RETRY_COUNT on emulator, r=vyang
With this patch, Gecko sends REQUEST_GET_UNLOCK_RETRY_COUNT to rild
when the retry counter for a lock is being queried on the emulator.

On devices other than the emulator the function returns the error code
REQUEST_NOT_SUPPORTED, which is precisely the behavior we currently
implement.

UltraBlame original commit: 01169ceebd7aff26a9b6eaa6ecba39061721ecbc