Bug 892603 - Remove unused MOZ_APP_ICON and LINUX_BRANDING_FILES from mobile/android/base/Makefile.in. r=mfinkle
MOZ_APP_ICON was a Makefile dependency not mentioned elsewhere in the
tree.

The text `LINUX_BRANDING` does not appear in the tree.  There are no
references to `fennec_48x48.png` or `fennec_72x72.png` outside of
mobile/android/base/Makefile.in, and that file explicitly copies the
referenced files (so does not depend on the deleted export rules).

This leaves the relevant Makefile.in files empty, so we can remove
them.

UltraBlame original commit: a70028dc9b0b5bffabd278d65de779ea350ca800