Bug 898877 - Ensure that viewport resizing behaviour due to dynamic toolbar is consistent. r=Cwiiis
In browser.js were two pieces of code that determined whether or not the CSS viewport
size should include the margins. These two pieces of code were inconsistent in that
one used rounding while the other used a fuzz. Also, one of them subtracted gViewportMargins
from gScreenHeight while the other added them. This patch makes the two pieces of code
consistent, and updates them to use a fuzz so that the CSS viewport is enlarged only
when dealing with pages that are equal to or larger than the screen (with the toolbar hidden).

UltraBlame original commit: f9d39959fc55d35e54047eedf86eb74d2fea0c46