Bug 880064 - Change HTMLTrackElement::Kind to a DOMString value. r=Ms2ger
This was originally in the spec as an enumerated value, but due to the current
problems with reflecting enumerated values it had to be changed to a DOMString.
This has been changed in the HTML5 spec as well.

UltraBlame original commit: c1bb3be506eeb27f6b09e0386cb02f4217cc7e17