Bug 879578 - Re-select last selected node when page is reloaded; r=paul
Done at inspector-panel level, this module is not reloaded and is
therefore used now to record the last node (unique css selector)
that has been selected as well as the URL of the page where it has
been selected.
On page reload, an attempt is made to re-select it, otherwise, the
logic implemented previously is executed.

UltraBlame original commit: bf671d77bdadb5428898870dca2381ba6a24ea9c