Bug 882142 - _postUpdateProcessing is not run after updates from Metro when Metro updates are disabled. r=rstrong
These updates can still be applied when the user manually clicks on check for update from
within Metro.  When _postUpdateProcessing is not called, then the last status in the update history XML
is not updated to succeeded.  This causes the Metro browser to show a "Restart to Update" button
because the last update's status is still applied.

UltraBlame original commit: a51560342d14604aa7505211e42c749eac77b9e4