Bug 905926 - Stop creating an ephemeral JSContext during XPConnect shutdown. r=billm
This interacts badly with our attempts to shut things down nicely, and happens
to be entirely unused. the primary impact of it ends up being that the cx
destruction triggers a GC, which we depend on. But we can do that manually.

UltraBlame original commit: 338b7a7e856116b4ccfe9898fcaf6cea229ce6e0