Bug 922617: Don't inherit BluetoothDBusService from RawDBusConnection, r=echou
BluetoothDBusService uses RawDBusConnection, but isn't one. With
this patch, the connection is instead maintained as a field of
the class.

This change also allows to delete the class instance when the
connection gets closed, which will allow for further cleanups
of the DBus connection handling within Bluetooth.

UltraBlame original commit: 32b6fc360a0b6884879b305419f38d4ed367f29a