Bug 773296 - Part 12: Record whether we are parsing an supports condition. r=dbaron
This stores on the nsCSSParser whether we are somewhere in the middle
of parsing an supports condition.  Because supports condition parsing
uses the scanner recording function (to save its conditionText), and
variable reference containing values also need it, we can't do both at
once.  Luckily, if we're parsing an supports condition, we don't really
need to store the token stream text; we only need to know if it was
valid, and we don't need its actual value later.  So we use this flag
later to see if we can skip turning on scanner recording while parsing
variable reference containing values.

UltraBlame original commit: 88fd50578a763adeaefbf1214e3e4c5cf75598ac