Bug 949304 - Print Makefile counts during config.status; r=glandium
We now capture and print the number of Makefile.in and the number of
generated Makefile as part of config.status. This should give us a nice,
easy to extract metric going forward.

Pushing on a CLOSED TREE because this patch didn't deserve to get backed
out.

UltraBlame original commit: 456d3ef07190dc92997f48be72f894ed293d575c