Bug 964559 - Run ogg mochitests unconditionally. r=doublec
Now that support is always enabled, we can remove the conditionals
on the ogg-dependent tests.

The canplay test previously run only if ogg was disabled is updated
to verify that the runtime pref works when set to a non-default
'false' value.

UltraBlame original commit: 662f7982dc8b81c8b945c70d3758868bee227c02