Bug 401564 - Test and fix for redirection losing Accept header. r=mcmanus
The test has a redirect where the initial GET request had Accept:
application/json, and ensures that this is also the type of the
redirected request.

The fix is done in the setup of the replicated HttpBaseChannel, where
the previous channel's Accept header is copied to the new one, if it
exists.

UltraBlame original commit: 3a737c64fa7433212f03ed69ff18d02f007d1a83