Bug 977340 - Clean up cx usage in Promises. r=bz
None of these are hazards, because we already make sure to push the JSContext
in the cases where we do anything meaningful in JSAPI. But the current setup
trips the new assertions, and is ugly to boot. Let's fix it.

UltraBlame original commit: 219a096d2d1e7b22ce2ed5fc07d3d219bd7d6c88