Bug 979913: Conditionally define PAGE_SIZE, r=khuey
With this patch the PAGE_SIZE value from the system is used if
possible. The code still assumes that the PAGE_SIZE is 4096, but
emits a warning if the actual value differs. If page size is not
defined by the system, it is set to 4096 bytes and a warning is
emitted as well.

UltraBlame original commit: 3e9182f7e25d1a006e9113410e70295f63895960