Bug 979913: Conditionally define PAGE_SIZE, r=khuey
With this patch the PAGE_SIZE value from the system is used if
possible. If page size is not defined by the system, it is set
to 4096 bytes and a warning is emitted. PAGE_ALIGN_MASK is now
derived from PAGE_SIZE. With these changes, the code should be
more robust on systems with non-4KiB page size.

UltraBlame original commit: 6659e34090c966de7352c07c877ce605b7760357