Bug 977861. Avoid putting the scroll thumb at subpixel positions. r=mstange
The solution for fixing this properly is not clear and could be a lot of work.
Matt Woodrow agreed that this is probably the right thing to do for now.

This is b2g only because it was causing reftest failures on OS X that I do not
understand and is pretty important for performance.

UltraBlame original commit: 96c7ae85ff76b7f7bf1a3aec1a616debe356e5fe