Bug 988437 - Part 2: Make Firefox Account Android Account type unique per package. r=rnewman
This bumps every package's account type so that any issues shake out on
Nightly (which otherwise would not have experienced an account type
change).

The trivial change to the AndroidManifest is to ensure that the new
Account type is picked up by the build system, in the hopes of avoiding
a clobber.

UltraBlame original commit: 8c483fd63cadc80a8ac9b21acac8030757225bbc