Bug 996148 - With tabs in titlebar disabled, the private browsing mask can be incorrectly positioned in certain modes. r=MattN.
The positioning of the private browsing mask depends on us correctly setting the fullscreen placeholder
width. Unfortunately, we only ever did that when kicking off the tabs-in-titlebar code. This patch makes
it so that we set the width regardless of whether or not tabs in titlebar is enabled.

UltraBlame original commit: 47845138fae1269dfbb27538c5703f671d39ee05