Bug 990353 - Make the decision to discard source entirely per-global, rather than per-script. r=luke
This is effectively a policy decision based on the kind of code we expect to be
running somewhere. This is in contrast to lazy source, which is often a practical
per-script consideration of whether or not we can retrieve the source if requested.

More importantly, tracking this information on the global is much easier to
get right than tracking it on the script.

UltraBlame original commit: 487e6f72fcf74e590b6b475958abc88315bd05b0