bug 998959 -  remove xpt headers from system headers list r=bsmedberg
These headers were added with the initial support for hidden visibility
in bug 273336.  At that time this code probably lived in libxpcom and needed to
be callable outside of that, but that's not true anymore.  At this point not
linking this code into libxul seems like madness, so we
might as well stop calling them system headers so we stop exporting them
from libxul.

UltraBlame original commit: 79cfe78d34cb6ef080a23bda026da29b5e694053