Bug 996061 - part 3 - try to shim param information in xptiInterfaceEntry when possible; r=bholley
This patch does all the interesting work: we check for shims whenever we get
asked about parameter information for methods.  If there's a shim for what we
want, we ask the shim to provide the information.  If not, we fail as usual.

UltraBlame original commit: 2c7acfdd623bdb725c036a1b1fbb2cbe15f392b6