Bug 1026344 - Part 1: Don't use memcpy to copy style structs. r=dbaron
This is in preparation for nsStyleCoords becoming non-POD types.  The
use of memcpy is a bit of a trap when modifying the style structs, and
it likely has marginal value now that anything with an nsStyleCoord
can't be memcpy-initialized.

UltraBlame original commit: 7ca674bb56877e0cdaa9a85842a3a6bc37cb9c8e