Bug 1046342 - bring xpcshell's usage message a little closer to reality; r=bsmedberg
The -n, -j, and -P options appear to be no longer supported.

I rearranged the options list so that the options appear in the order of
processing in ProcessArgs.  I also took the opportunity to move the -m'
option out of the options claimed to be handled in
ProcessArgsForCompartment, since it no longer is.

UltraBlame original commit: 1163a780cb23591366133ed0b7246806c36f2fdd