Bug 1053071 - Simplify EXTERNAL_SOURCE_DIR handling. r=gps
As handled in configure.in, EXTERNAL_SOURCE_DIR is a single directory for
external sources, and more specifically, comm-central. There is no need
to handle it as a list in the moz.build emitter.

At the same time, avoid checking the external sources for every moz.build,
and bind it to a config environment.

UltraBlame original commit: 0ae7262449d8d7e4dcf0520c2b2ea539afc71a8d