Bug 1053185 - Fix preferences installation and packaging for Mulet. r=mshal
Due to the way preferences are read, we need to have firefox.js and
b2g.js to be living together in the same directory, as exposed in bug
1053185 comment 7. Doing so will ensure that the proper precedence is
given to the pref and in the end we ensure that the preferences we
redefine in b2g.js indeed overwrite those defined by firefox.js. To do
so, we need to define DIST_SUBDIR to browser for Mulet so that rules.mk
will pick up the proper preferences directory. We also have to make sure
that the B2G-specifics preferences are properly packaged otherwise the
redistribuable tarball will not boot Gaia at all.

UltraBlame original commit: a846839447908749948b3d114c17089c9a97140e