Bug 1057376 - Clear the string read buffer between deserializations of CSP policies. r=geekboy
In case when multiple CSP policies are applied it is needed to clear the
temporary string buffer that holds the string of each CSP policy that is
deserialized from the CSP policies stream.
Failing to clear the buffer will result in concatenations of the CSP policies
strings for each consecutive CSP policy that is being deserialized, which will
end in generation and application of invalid CSP policy for all except the
first CSP policy that was set for the document.

UltraBlame original commit: b70adcc6738fd6039675a8c61857c25c3de16a18