Bug 1065881 - [e10s] Crash on child process shutdown in ShadowLayerForwarder::InWorkerThread. r=nical
Had caused a crash in the test_streams_element_capture mochitest on Windows.  The issue was that the audio decoder thread was destroying stale media frames by releasing a reference to them.  This was causing the release of an Image on that thread, but that thread does not process a MessageLoop so the current MessageLoop is null.  This is not a problem.  Obviously, any thread that has no MessageLoop can't be the worker thread.

UltraBlame original commit: 96aefa3026881ae5647664b13fc92f7f9171ad6c