Bug 1086705: Disable -Werror=switch, r=froydnj, a=RyanVM
Gecko, for better or worse, allows non-m-c apps to define custom error codes and
use them in nsresult. This error breaks the ability to switch on those custom
error codes when an error happens. For this reason, it's not reasonable to make
this an error at present.

UltraBlame original commit: f7ecd7ba118f32dd250ab39e20513342a4444a3b