Bug 1039902 - Minify the Fennec JS code in automation. r=glandium
This patch adds a SpiderMonkey jsshell binary via tooltool and arranges
for it to be used in automation (more specifically, whenever JS_BINARY
is defined).

We'd prefer to build a host SpiderMonkey jsshell, but that's not trivial
right now; and we'd prefer to use a minifier better than the existing
Python jsmin (possibly written in JavaScript), but one step at a time.

UltraBlame original commit: 35a4be7fb0a861dc177273928176ab871d1392fc