Bug 1099430, part 二: Define WIN_TOP_SRC using MOZILLA_DIR, r=glandium.
win_srcdir is used as the Windows path for the srcdir, and so it needs to use
comm-central's topsrcdir to work properly. In contrast, WIN_TOP_SRC is really
only used to define BUILD_TOOLS, which refers to files in mozilla-central/build,
and so that can and should remain the mozilla-central topsrcdir.

UltraBlame original commit: 358aa39360d5cc1159d9f0a28d38e131e23d286b