Bug 1106324 - Call removeLock in case where the call was missing. r=qdot
While I was debugging bug 1105511, khuey noticed that one case was
missing a removeLock call.  We should add it; it seems like if we ever
hit this error case, the settings lock queue would get stuck.  It's not
clear whether this is actually the cause of bug 1105511, so I'm putting
it in a separate bug.

UltraBlame original commit: d920332c6ce6dadcd2a9a28956568064d2cdec8d