Bug 1108059 - Enable mp4, not webm, for MediaSource on MacOS X. r=ajones
This makes the setting the same as on Windows. This will reduce
the amount of testing we get for WebM, since only Linux will be
left with that enabled, but reduces the exposed code surface while
we're trying to stabilize things enough to ship support.

UltraBlame original commit: 8878e674973db4a6b8ae046dda700b7a520675cf