Bug 941749 - Add support for returning both known tabs and chrome windows from the marionette server.;r=automatedtester
This splits the api for getWindowHandles and getChromeWindowHandles for those
managing windows and tabs separately in tests. getWindowHandles is
re-implemented to return tabs when appropriate, and getChromeWindowHandles is
added to correspond to individual os windows (the prior behavior of
getWindowHandles). current_window_handle now corresponds to a tab on desktop,
while chrome_window_handle is provided for the current OS window.
switch_to_window will accept either a chrome window handle or tab handle.
In the case of a tab, switching will select the tab. In the
case of a chrome window handle, the prior behavior is retained.

UltraBlame original commit: 9011ae0f6be886db06a88c0858bbf088ae24babd