Bug 1115999 - Call ApplyRelativePositioning properly for some table cells that we choose not to reflow.  r=roc
I confirmed that the reftest passes with the patch and fails without the
patch.

This is the crazy codepath where nsTableRowFrame sets doReflowChild to
true, and then tests some additional conditions, and based on those
conditions decides to do part (but not all) of the work of reflowing the
child.  This is making relative and sticky positioning work correctly in
that case, which we should really just get rid of, but which I don't
want to deal with now.

UltraBlame original commit: c7dde66a57bb3bec566659631c3d7ea6a027a151