Bug 1123002 - Convert XPCStringConvert::ZoneStringCache::mBuffer to void*; r=mccr8
This variable is only used for pointer comparison in order to compute
a cache match.  This patch converts the type of the variable to void*
so that it can still be used for the pointer comparison, but not for
any other purpose, including dereferencing, since the pointer may
potentially be dangling if the string buffer dies.

UltraBlame original commit: d8c0fee0015a705ddd5298ce9f9494c36a740466