Bug 1112243 - add sizetopref to menulist-popuponly binding content. r=dao.
The rest of the menulist bindings have this, and is used to properly size
the menupopup when the menulist is opened. Without it, for long menulists
with a scrollbar, and short labels, the labels will get squeezed out during
layout and not be visible.

UltraBlame original commit: a465bbe65e90202d1c22ca7945304f56c3ed3a18