Bug 1136700 - enable tweaking of the maximum number of timeouts for mochitests; r=jmaher
When running tests locally, it's occasionally useful to be able to
increase the number of tests permitted to timeout before declaring the
test run a failure.  This patch adds the necessary bits to SimpleTest
and the appropriate amount of plumbing to runtests.py and mach to make
that so.

UltraBlame original commit: ab9f361648bd428a55366703b1bf5ee5049c7ec9