Bug 1132229: Return in Dispatch method if Bluetooth runnable is nullptr, r=shuang
With this patch, the dispatch method of Bluetooth runnables simply returns
if the supplied runnable is nullptr. This makes is easy to handle cases
where the caller of a Bluetooth interface method is not interested in the
result.

UltraBlame original commit: 15fc5f2edcab0c01c7fc5ed7d5ed48b35687cac2