Bug 1138934: Make sure the white texture of a component alpha pair is fully uploaded when new. r=nical
This is slightly wasteful since this will also always upload the full black texture. The wastage here is so rare and generally so few though that I don't think it's worth adding plumming for the texture host to know if this is the texture on white of on black.

UltraBlame original commit: 713a86aa0c37f1c6841017109d200a1222e6d2c3