Bug 1109592 - Only open connection to nfcd while NFC is switched on. r=allstars.chh
With this patch Gecko only creates a connection to nfcd while NFC is
switched on, and closes the connection after NFC has been switched off.

While the connection is being opened, commands to nfcd are held in a
waiting queue. Gecko flushes the queue after it connected successfully
to nfcd.

As a side effect of this patch, the NFC thread only exists while NFC
is switched on.

UltraBlame original commit: 6c311023dcabbea2d89401f9a11dce3c1e4e8fd0