Bug 1145017.  Use the new proto setup for custom element prototypes when possible.  r=wchen,bholley
Note that with this patch I'm also changing the priority order of the given
proto and the custom proto; the former takes priority.  This makes sense to me:
if the caller is doing |new Foo| they really should get something with
Foo.prototype as its proto.  This should not be a problem for custom elements
in general so far, because nodes/elements are mostly not constructible anyway.
For now.  When they become that way, I think this is the behavior we'll want.

UltraBlame original commit: 6030e30ae43d368edcb74ee57867c964be41bf58