Bug 1137771 - Intermittent browser_animation_play_pause_button.js; r=miker
Instead of waiting for the next auto-refresh event only (which could be a
response to an older request, therefore not having the expected state yet),
wait until the state changes to what we expect in the test.
This means that if the play/pause button doesn't work anymore, then the test
will timeout, but at least it won't intermittently fail as it was doing until
now.

UltraBlame original commit: 6da23147f765f26bfdb44602dfcf7ee0700e8781