Bug 1148390 - Dynamically add padding to share icon on GB devices. r=wesj
The ideal solution is to remove the whitespace from the action bar drawables
(bug 1122752) but since I'm not sure where else these images might be used and
how removing the padding will affect all of the different devices and
configurations, it would be easier to uplift a solution where we *only* change
the padding on the share icon and *only* on the affected devices.

However, this solution is dangerous because we hard-code a filename: if we
rename ab_share, things will break. Also, unnecessary special casing makes it
ugly.

UltraBlame original commit: 73daa8f0f5f0f2b057ebddd89af6f76132d5e16f