Bug 1156847 - Part 4: When storing a service worker script to the cache, store its security info in the cache as well; r=khuey
Before this patch, we would call Cache.put() before opening the channel,
which means that we have no way of knowing the security info for the
channel in order to set it on the Response object that we synthesize.
This patch adds an nsIRequestObserver to the tee created for piping the
body of the response to the cache, and delays calling Cache.put() until
we receive the nsIRequestObserver::OnStartRequest() notification, at
which point we set the obtained security info on the Response object to
be stored in the cache.

UltraBlame original commit: 343858d83d73777dfe1b6cd3924fb82f62e0ea96