Bug 921102 - 3 - Open/copy markup-view attribute links; r=bgrins
This part adds contextual menu items that become enabled when
the user right clicks on an attribute that has a link.
Depending on the nature of the link, a new tab will be opened or a node
selected.
The user can also choose to copy the link in the clipboard.

UltraBlame original commit: 8fe899f2537eab56aa541b79d5f4a0422ff1942c