Bug 1155653 - Preview animation target nodes in animationinspector panel; r=bgrins
This adds a generic walker method that cna be used to retrieve AnimationPlayerActor's
target DOM nodes in a way that they can be highlighted/selected later in the inspector.
Making the AnimationPlayerActor return the nodeActor directly wouldn't fully work, as
the walker would have missed the missing parent nodeFronts updates.
On the UI, this adds a new component in each PlayerWidget that displays a preview of the
target DOM nodes, highlight them on hover, and select them on click.

UltraBlame original commit: f88d5fb92a29339ae00e569b4058ccd646ddcdf2