Bug 1160720 - Stop dragging the color selector around if no buttons are pressed down on mousemove. r=bgrins
The color picker stops following the mouse when it receives a mouseup event.
However, in some cases when the mouse button is released outside the color
picker frame, the picker won't know about it and will continue to move the
selection when mouse comes back over the picker frame.

As the MouseEvent.buttons property specifies which buttons are pressed when
the event occurs the picker can use said property to detect if it should stop
following the mouse. When .buttons === 0 on a mousemove event no buttons are
pressed down and it should stop following the mouse.

UltraBlame original commit: 019c10ffbe1607f2618d216b86b028d41bceea74