bug 1162543 - make DocAccessibleChild::IDToAccessible correctly get the document r=davidb
DocAccessible::mAccessibleCache doesn't contain an entry for the document
itself so we need to handle it separately.  The document always has ID 0, so we
can just check if that is the ID we are getting and if so just return the
document.

UltraBlame original commit: b34c538f36804b7030a32404e67c859d52340a9d