bug 1161903 ensure pending DrainComplete is not run after Flush() r=cpearce
Using a Decode task already in the queue had the potential for out of order
flushing.

This is similar in behavior to the implementation prior to f550eb7809b6, but
keeps mDecoder->Flush() on the platform decoder's task queue, in the hope of
avoiding any ordering problems from calling in the middle of decoding input,
or races with MFTDecoder::mDiscontinuity.

The contract is not clear on whether DrainComplete() should be run during
Flush(), but I've kept it in this version.

UltraBlame original commit: 7b11fc3efb145d63e7d2ab7931a0f0282ae54063