Bug 1103348 - Part 1: Correctly reset the direction of an ancestor that is still in the tree when a text node is removed; r=smontagu
This is essentially a better fix for bug 894137.  Relying on the aNullParent
argument to UnbindFromTree doesn't actually tell us if the dir=auto node
is still in the tree or not.  Instead, we can check the parent of the said
node and only reset the direction if the parent exists, which means that
the node is still in the tree.

UltraBlame original commit: a54020a2280821292a46833c39f67c41a46d2bde