Bug 1158634 - Restore support for editing previews in Font Inspector. r=pbrosset
An input box similar to those in rule and computed views (without the search
symbol) is added to the font-inspector panel which can be used to control the
preview text. The updates are throttled to avoid unnecessary updates while the
text is being edited.

Since the PageStyleActor has allowed clients to define the preview text for a
while now, this feature is also enabled for older targets too.

UltraBlame original commit: 73027d05f9cad26f46cb498ff03bd2578f54463b