Bug 1170332 - Fix |mach robocop SINGLE_TEST|. r=ahal,gbrown
This patch declares robocop.ini an instrumentation manifest.  It's not
currently possible to declare tests that don't correspond to files, so
we include the .java extension.  (This could be revisited.)

In |mach robocop|, we use the generic test resolving infrastructure to
select the 'instrumentation'/'robocop' flavor/subsuite tests.  In
|runtestsremote.py|, we fall back to robocop.ini, as we always have.

UltraBlame original commit: 0d93a77bb15326125395894d9f36c61f71fb5338