Bug 1170477 - 2 - Fix undefined inspector.strings accessor in breadcrumbs; r=tromey
The only user of the l10s inspector strings bundle is a string that this
change removes.
The "Siblings" string is supposed to tell users that the section of the
menu that follows contains sibling nodes, but it's disabled (and looks like
an item you should normally be able to click on), and it's separated from
the list of siblings by a horizontal separator (making it look like even
less related to the next section).

Also fixed the fact that the inspector context menu was regenerated for
every sibling found on the current node.

Also fixed the fact that the inspector context menu would not appear when
right clicking on the documentElement breadcrumb node.

UltraBlame original commit: 9ec6079da48815fbc9338ff231e4657720217b13