Bug 949355 - Wait for iframes to be loaded before looking at their style sheets. r=bgrins
Previously the getStyleSheets method was only waiting for the topmost frame to
load before listing the style sheets. If a frame wasn't loaded in time, it
caused some sheets to be missing causing the test to timeout.

These changes make the actor to wait for a frame to be loaded in _addStyleSheets
before doing the actual adding. This allows the code of _addAllStyleSheets to be
moved directly to getStyleSheets as it no longer needs to wait for the load to
happen.

UltraBlame original commit: aebbbe2efe1d3b11e129a096bfb9650d56d961d8