bug 1170595 - switch to sending __delete__ from the parent instead of the child r=davidb, smaug
After the child sends the __delete__ message to the parent there is a period of
time in which the actor is registered, but the parent hasn't yet processed the
__delete__ message.  During that time the parent can still try and send
messages to the child, but that will crash the child process.  Fix this race by
making the child send a shutdown message to the parent, and have the parent
send __delete__ when it handles that message.

UltraBlame original commit: 78fc82c0601a4b90d76e2974ff4d64580c24f22b