Bug 1172536 - Stop using for each loops in layout; r=roc These are SpiderMonkey-proprietary legacy feature which has been deprecated and is expected to be removed in due course.
This commit also fixes a number of bugs in test_bug708874.xul. In particular,
because of the semicolon after the for head, the (alleged) loop body was only
executed for the final element of the array ({}), and because each of the
functions under test threw an exception, only the first call was executed.

I do not know of a way to test the changes in frame-verify.js, so I can't
guarantee they actually work.

UltraBlame original commit: fe94dced5e595db6b1e8aaa824583a31e6b05355