Bug 938188 - Make highlighter capable of highlighting only one region, fading out others. r=bgrins
Introduced a new highlighter option that makes each region <path> element in the highlighter only
cover the actual area of the corresponding region, excluding the area of nested regions.
This is useful when used with the existing showOnly region because it lets users see exactly where
a given region is.
This patch makes the layout-view use this new option, so that when users hover over the various
regions in the layout-view, only the corresponding regions are highlighted.

UltraBlame original commit: 975d6d156f8e86cceff703e85a17f9c8f15a5ed2