Bug 1175637 - Make sure browserWin.BrowserViewSourceOfDocument exists before attempting to call it. r=jryans
This handles the case when there is a syntax error in browser.js that shows up in the
Browser Console, and you try to click the link to the source.  Without this, an exception
is thrown but with it the source opens in a new window.

UltraBlame original commit: 5e376b895220eeb611705a7dbce4f4fe53d1aa78