Bug 1182677 - Touch last check file before main command invocation; r=smacleod
While not common, there are a few cases where `mach mercurial-setup`
fails. When checking the last time this command was executed, it is
sufficient to record the intent to run the command, not the fact that it
executed to completion.

UltraBlame original commit: 82d949cc91ddedf8ba1c3d9ba5765f4a927ab7b1