Bug 1186257 - Use default dtor in ContainerParser. r=jya
This fixes a non-unified build error. Defining and empty
dtor allows callers to inline (if they can prove it's not
a pointer to a derived class?) the dtor, but means the compiler
needs access to the details on the member variable wrapped
in smart pointers so it can call there dtors in turn.

I think marking it default instead of empty in the header
moves generation to the class implementation, which does
include SourceBufferResource.h.

UltraBlame original commit: f9e936bbe02a9443451251548ded844a2c14a3d3