bug 1179660 - define 'now' as the first second of the current year for pycert r=Cykesiopka
This is to avoid a dependency on the buildid so we don't have to
regenerate all of the test certificate with every ./mach build.
This can cause problems very near midnight on New Year's Eve.
If this happens, kick off a new build and get back to the party.

UltraBlame original commit: eb0a49e8322aac5212c01dc0f67dde0b2552be63