Bug 1191959 - Make sure that pinned tabs are still clickable after unuting a tab that is not playing; r=jaws
This bug happens becuase when toggleMuteAudio() is called from the click
handler for the tab, we remove the muted attribute during unmuting,
which makes the element display:none.  Therefore, when the mouse pointer
leaves that region, there is no element to receive the mouseout event
and as a result, the _overPlayingIcon variable stays true, which means
we stop tab switching in the mousedown handler.

UltraBlame original commit: f58ce99d0d99256512a5e5f7f648a98b9d4b5da7