Bug 1184604 - Lazily override printer output format on Linux if not already set. r=karlt
GTK versions prior to 2.24 incorrectly advertise themselves as able to print PDFs,
even when they can't. We were wholesale setting the output format to PostScript for
these older GTK versions, without giving the user the opportunity to override
it. We now lazily determine whether or not the output should be in PostScript,
which should give the user the opportunity to override.

UltraBlame original commit: f294f9e327aced29545eec980445b6ae2ad4aca6