Bug 1196970 - Part 5: Copy preprocessed code from objdir to build directory during Gradle build. r=sebastian
This means we don't require the directory in the object directory at
gradle-install time.  We're not concerned if the source files are
missing, since we have code to ensure they're fresh already; and if
they are missing, we'll quickly fail as we try to compile with missing
sources.

UltraBlame original commit: 0351b18996b97ecc3599510987056da547ceeee5