Bug 1196558: Don't assert should output task not be dispatched. r=edwin
We are in a racy situation where the task queue could start flushing after we've tested if it was being flushed. We do not care if the task isn't going to be dispatched as we want the frame to be dropped when that happens anyway.

UltraBlame original commit: 4a61d57d6d87a7d941418d50b25e31ee422f08c8