Bug 1202652 - Remove redundant and incorrect check when setting displayport margins. r=snorp
The intent of this check is to avoid setting the same margins more than once.
However this is redundant because the code in nsLayoutUtils::SetDisplayPortMargins
already has an equivalent check. Further, this code is wrong because it stores
the old margins per-tab, and so once a new document is loaded the margins may be
the same as "before" but they apply to a different element. In order to be correct
the check would have to track the target element as well as the margin values,
but it's easier to just get rid of this and let nsLayoutUtils handle it.

UltraBlame original commit: b337feb71c38797fbdfe4ff24ba34c5a74040c20