Bug 1205706 - Part 1: Pass-through white-listed intent extras to about:accounts. r=sebastian
Pretty straight-forward.  It's possible that getStringExtra is not
safe in the face of malicious data; the worst we might expect is a
crash on the Java side; a large memory allocation on the JS side; or
significant URL data transfer.  The first is valuable because we'd see
abusers in crash-stats; and the second two are already possible when
opening any URL.

UltraBlame original commit: 70f459b00b8821ef13678ec1dfd3493bf8e8914c