Bug 1205706 - Part 2: Add entrypoint parameter to selected Firefox Account activities. r=sebastian
This list is not meant to be exhaustive.  There are many ways to get
to the Get Started activity; not all of them are interesting.  We're
just trying to capture the significant entrypoints, like firstrun and
preferences.

UltraBlame original commit: 49dda13aabe2a5e74b0409db132b6e4778b9fa05