Bug 1195180 part 10 - Remove refresh driver observing machinery from CommonAnimationManager; r=heycam
This patch removes a lot of code involved with observing the refresh driver from
nsAnimationManager and nsTransitionManager now that we no longer need to do
this.

The one piece it does not remove, however, is
AnimationCollection::mNeedsRefreshes since this flag actually serves a secondary
purpose in telling us when the animation style has not changed and so does
not need to be updated. A subsequent patch in this series will rename this
and update the code that makes use of it.

UltraBlame original commit: 19e81f18c957819b098b6a47781c5df929e427b3