Bug 1195288 - consider using python webserver for production talos. r=jmaher
Always use an in-process webserver, removing the need for apache - and
hopefuly providing better accuracy for numbers.

This means that we know have to copy the pagesets in the talos dir on
harness.

On windows, some pagesets paths were too long due to that, so the
solution is to replace "page_load_test" with "tests".

UltraBlame original commit: bdbdceb94fb17bca9a2d8c5fe750441f2a655f5a