Bug 1170958 - Remove ProcessedMediaStream::ForwardTrackEnabled. r=roc,jesup TrackUnionStream guarantees that TrackIDs are maintained if no tracks have claimed them before.
In the gUM case, we have a SourceMediaStream which we wholly own (the
DOMMediaStream's Input stream), piped into a TrackUnionStream which
no-one external is able to add tracks to (the DOMMediaStream's Owned
stream) - addTrack()ed tracks are added to the DOMMediaStream's Playback
stream.

The MediaStreamTracks being enabled/disable refer to a TrackID in the
DOMMediaStream's Owned stream.

Alas, we don't need to forward a track's enabled state, we can just do
it on the source.

UltraBlame original commit: 03a29506e206fb105de375fc03c0e92dec7bc557