Bug 1202713 - Fix the CSS viewport dimensions for desktop mode. r=snorp
The CSS viewport computation for desktop mode was incorrect. The end result was
that the right amount of CSS pixels was visible on the screen, but the layout
properties were incorrect (so elements that were width:100% would use a different
width than intended). This caused content to be truncated and otherwise
mis-rendered in desktop mode.

UltraBlame original commit: f89a4c296fb1dfdad6e4580ad36ba339f34f3c15