Bug 1182727 - Part 2: Run cmake through ninja to enable resumable builds; r=rail
Since CMake generated build systems can run cmake if necessary, this
will make it possible to pick up changes from the source directory if
any and resume as much of the build as possible.

This builds the foundation for removing the need to blow away any of the
work done by the previous runs of the script.

UltraBlame original commit: 5fb1820cd5df284a1a8760375cc10f7c2169e856