Bug 1205705 - Part 1: Always resolve or reject the LOADED message promise. r=markh,mfinkle
This is hygiene that completes the set of paths through this part of
the code.  If we wrapper.{init,retry}, we are guaranteed to have a new
promise; and now that promise will always be fulfilled.  It is
technically possible, but not anticipated, for an in-flight promise to
be replaced.  Such a situation should not occur, but if it does, the
obsolete promise will still exist but never be fulfilled (since
loading or errors only touch the most recent promise).  Eventually it
will be safely garbage collected.

UltraBlame original commit: 1d8058a6dfc237db11e88ffcdb20362fa0dd9df0