Bug 1208534 - Part 2: Fix test. r=mfinkle
It's quite challenging to both wait for "load", and wait for something
to happen in the DOM, since the DOM isn't prepared until after "load"
has fired.  This test therefore has a small race window: it is
possible that we could wait for the mutation only after the logins
have been loaded and the 'logins-list' DOM element is inserted.  The
logging should be good enough to identify this case; and in practice,
this is very unlikely.

Since I was here, I converted this to use SpawnTask.js.

UltraBlame original commit: 49d88d6dfa50e56d03096204783a844b970335aa