Bug 799937 - Make createElement()'s result always in the HTML namespace (except for XUL documents); r=bz
The new expected failures in web-platform are in a test that doesn't
match the spec, which has already been synced upstream and will be
synced shortly.  For the same test in imptests, I chose to modify the
test instead, because it's easier and imptests is going to be removed.

UltraBlame original commit: f6759ed53f46fa84963218e2a3f3e15e698239bf