Bug 1209603 patch 11 - Assert that PeekStyle* results don't change during difference computation.  r=heycam
This assertion catches the condition that led to the bug.

I confirmed that without patch 10 on this bug, the assert fires on the
reftest added in patch 4, but does not fire on slight modifications of
that testcase that don't show the bug.

UltraBlame original commit: bb679653777502d9099a403520ed8b17da936e3f