Bug 1217261 - Update mp4parse-rust to v0.1.2. r=kinetik
This switches to #[derive(Debug)] for trace println!s, and
updates the C api wrapper to thread through the new MediaContext
as an opaque struct. The idea is to be able to query it later.

'extern crate byteorder' converted to 'mod byteorder'.

UltraBlame original commit: 2ca3997df07f3f46dd97fe15bbfdec8e51913693