Bug 1216431 patch 6 - Back out bug 1209603 patch 8.  r=heycam
This requires a little bit of gymnastics since it has to add the inverse
of tests, since the is-a-reset-struct tests originally added in patch 8
were made unconditional in patch 9, and with this backout we now want to
execute the code only for inherited structs.

This also doesn't back out the cleanup to use NS_STYLE_INHERIT_BIT() for
constants rather than nsCachedStyleData::GetBitForSID.

This backs out the part of bug 1209603 whose speed I was concerned about.

UltraBlame original commit: 063bf9fb0cd344462c570a4cc05c2f537ecf52c8