Bug 1214366 - Part 2: Add getLastVisited equivalent to Fennec. r=rnewman
The approach is the same as that used on Desktop.  The function name
calls out the "prePath" and "milliseconds" since the first is likely
to be overlooked by Fennec consumers, and the second is notably
different than Desktop (where the database timestamps are microseconds).

UltraBlame original commit: 662e4aff03d4b102da29c5d500a072346be44bda