Bug 1129662: sdk/page-worker should use a remote page. r=krizsa
This makes page-worker load its pages in the remote process. It does so by
creating a single frame in the hidden window used to ensure we have a remote
process when necessary and then a module in the remote process is used to
create windowless browsers to load the pages.

This does break one API, getActiveView, but I don't think we should be
maintaining that and it has been unstable since its inception anyway.

Once downside, the l10n module now has to use the observer service to detect
documents rather than the DOM event, this might be causing more CPOW traffic
since that observer notification is shimmed so we may need to use the shim
waiver there.

UltraBlame original commit: 5cdf66dfef92364072336ce276f9f2c70d421718