Bug 1156238. Skip setting async scroll clips if we aren't painting to the window because they are useless then. r=mstange
Displayports only get acted upon when painting to the window, and the async scroll clips only get computed when we use a displayport.

In addition we change an assert because if we are painting to the window then our root reference frame is either a root frame, or a popup frame. In either case we should not be able to get to out of flows outside of the frame subtree rooted at the root reference frame by following placeholders.

UltraBlame original commit: cb0f1211b033725af6bb253715f18fccede02f18