Bug 1141884 - Rename ResetInputState and make it only apply to touch events. r=botond
In particular, we want to prevent requesting a snap in the scenario that a wheel
block gets prevent-defaulted. In general the function is really only relevant
for touch input and there's no need to run it for other types of input.

UltraBlame original commit: 21ac71c3b9338acbb71cf97327b6ad9ab19a0b7d