Bug 1201540 - Add support for multiple OTA/FOTA types. r=catlee
B2G updates can be of multiple types:
 - OTA, applied without rebooting the device,
 - FOTA with only Gecko/Gaia,
 - FOTA with whole system partition files,
 - FOTA dumping partitions images.

Each type of updates has its advantages and drawbacks. There is an
extensive documentation maintained on MDN about each and the options:
https://developer.mozilla.org/en-US/Firefox_OS/Building_and_installing_Firefox_OS/Firefox_OS_update_packages

All those updates are being packaged as a MAR file that gets injected
into the classical Firefox update mechanism, submitted to Balrog and
downloaded by the client. The content of the MAR will however depend on
the type of update: an OTA update will packate a Gecko and Gaia set of
files to update those parts; while any FOTA package is just an
update.zip that will get applied in recovery mode on the device.

So one fundamental difference is that OTA will not reboot your device
(just Gecko) while FOTA requires a working recovery mode and will reboot
your device. But OTA needs more system partition space to get applied,
and it can only update files that are within the /system/b2g/ directory.
FOTA on the other hand can update anything since the payload will
contain an update script written in Edify (Android recovery update
scripting language).

For each device we might need to produce several types of updates that
will be pushed to users depending on the context: for some users we want
to push just a Gecko/Gaia update, for some we know that we need to
update more content and thus we need to send some partitions.
Previously, the b2g_build.py script would only allow one kind of update
payload to be produced for each device available: we would need to have
a device "flame-kk-ota" and "flame-kk-fota" just to produce the OTA and
FOTA packages for the same device, thus resulting in a waste of
computing power and storage.

This commit introduces a new field "update_types" that can take an array
of values:
 - ota, to produce an OTA package as before
 - fota, to produce a FOTA package with only Gecko/Gaia
 - fota:full, to produce a FOTA package of all files of the system
   partition
 - fota:fullimg, to produce a FOTA package dumping partitions

The old "update_type" will be used in the absence of "update_types". And
if none are present, we will keep defaulting to generating OTA as
previously.

UltraBlame original commit: 6b39a6ec3f92a56f1a1a77f00c699fe28c84443a