Bug 1037056 - Producing FOTA packages for Flame. r=wcosta
The Flame device updates have been broken for a while because of updates
starting to be too big. Because of the way applying OTA works, there is
no good solution except switching to applying Gecko/Gaia updates in
recovery mode. This was done in bug 1037056 on the Buildbot instances,
but we need to support this on TaskCluster also.

UltraBlame original commit: 5a3d69c1b2ad78f2b86237bdeecfe077d8da56ef