Bug 1222167 - fix initialization order in GeckoChildProcessHost; r=jld
Member fields are supposed to be initialized in the order they are
declared, but the constructor of GeckoChildProcessHost initialized
mDelegate prior to mSandboxLevel.  This is probably harmless, but it
does cause a warning on clang-cl, so let's fix it.

UltraBlame original commit: 239a8e8c626b2186f2fae3b546a6d56e3117be35