bug 1221551 - unbind child documents before rebinding them r=davidb
It seems possible that a child document could be bound to a parent document
without first being unbound.  Some evidence for this being possible is that
OuterDocAccessible::InsertChildAt() explicitly checks for this and does
something similar if it already has a child.

UltraBlame original commit: 13be87aa74c909369a32ce0af574660a17306404