Bug 1222170 - fix initialization order in ScopedWLANObject; r=jimm
Member fields are supposed to be initialized in the order they are
declared, but the constructor of ScopedWLANObject initialized mObject
prior to mLibrary.  This is probably harmless, but it does cause a
warning on clang-cl, so let's fix it.

UltraBlame original commit: 851dc8327772d5f6491b1bed591fc328baea50b7