Bug 1199588 - Add --workspace option to Marionette runner; r=automatedtester
The default behaviour is unchanged: gecko.log and logcat are saved to the
cwd, a temporary Firefox profile is cloned or created in TMP.

When a path is specified via --workspace, gecko.log, logcat and the Firefox
profile directory are all saved there instead. A profile saved in the
workspace does not get deleted at the end of a session unless it's a clone.

UltraBlame original commit: 38257da4c79ed69becbf75418361b9c2c4dbd625