Bug 1214305 - Part 4: Use the secure upgraded channel URI in ServiceWorkerManager::PrepareFetchEvent; r=jdm
This is needed to ensure that the ServiceWorkerManager uses the
correct URI for non-subresource requests.  Note that we're relying
on the property that non-secure non-subresource requests can never
be intercepted, so we don't need to check the request type explicitly.

UltraBlame original commit: 2e3e4f256d4656acd316a6af9a5a9b1e57acce68