Bug 1171708 - Use localName when detecting <xul:browsers> for XULFrameLoaderCreated. r=billm
We were using tagName before, which is fine for the dynamically created browsers
in new tabs, but not fine for the initial browser tab, which has a tagName of
"xul:browser" instead of "browser". Using localName makes sure that we don't
get the XML namespace included with the node name. We check the XUL namespace
separately be checking the namespaceURI.

UltraBlame original commit: 7a6196e9c10c106f24433ff55b23ebb276b1da60