Bug 1221270 - Let AddonUpdateService use AddonManagerPrivate to check for updates; r=mfinkle The Android-specific AddonUpdateService has a bit of redundant code because AddonManagerPrivate has a backgroundUpdateCheck method that does a lot of the same thing. This patch makes AddonUpdateService call that method so there's less code and more consistency.
This requires flipping the "extensions.update.enabled" pref, which was
disabled in bug 528588 for showing the XUL addon update dialog. I don't
think this is relevant anymore in native Fennec and with the later
rewrite of AddonManager, so I'm fairly certain it's okay to flip that
pref.

The patch also disables the AddonManager update timer because we have
our own update timer on Android.

UltraBlame original commit: d19f1416a81e7b17e023eb682679adcd000f1071