Bug 1210554 - Wire up branchDeadlines_ partially. No Asm callbacks yet. r=nbp
AssemblerBufferWithConstantPools geta a branchDeadlines_ member which keeps
track of forward branch to unbound labels.

Add a hasSpaceForInsts() method which collects the logic for checking for
available space in one place. Insert a constant pool both when constant pool
loads are about to go out of range, and when short-range branch deadlines are
about to expire.

Add registerBranchDeadline() and unregisterBranchDeadline() methods that the
assembler will use to add and remove branches to be tracked.

UltraBlame original commit: 774b04222cba7c92ee16db7e9cd2625fcbf21650