Bug 1222880. Followup to the followup to handle the case when aStopAtAnimatedGeometryRoot isn't an ancestor of aAnimatedGeometryRoot the same way we did before the patches of this bug. r=mattwoordrow (more or less)
Which is to return an empty scroll clip (instead of whatever scroll clip was accumulated in the for loop).

This patch is actually even slightly better than what we did before in that we can stop when we hit the root AGR (which is also the root reference frame) instead of the root of the frame tree. These would be different when painting a popup frame, or when not painting to the window.

UltraBlame original commit: 5e8bd912ef98989ff16a39953231a0e096e68efa