Bug 1228033. When not painting to a window (and hence not retained) only have one AGR, the root AGR. r=mattwoodrow
The document structure of when printing is as follows:

ViewportFrame
  nsSimplePageSequenceFrame
    nsPageFrame
      (the content of the page)
    nsPageFrame
      (the content of the page)
    (further pages)

The nsSimplePageSequenceFrame calls nsLayoutUtils::PaintFrame with each nsPageFrame as the frame to paint.

nsPageFrame::BuildDisplayList will build display lists inside sibling nsPageFrames if it needs to descend into those sibling page frames to find placeholder frames for out of flow frames that are on the actual current page frame. Since these are sibling page frames they are not in the original page frame subtree, and hence we can't find an AGR for them that is in the original page frames subtree.

However, when printing we don't retain painted content, so there is no point in having more than one AGR. So if we aren't retaining (IsPaintingToWindow is necessary for that) then just have the root AGR.

This patch also fixes two other bugs.

1) We should always set mAnimatedGeometryRootForChildren for nsDisplayTransform items.

2) If the AGR choosen for a nsDisplayTransform is already the root AGR, then that is it's AGR. Don't try to set the AGR to the parent (which is null).

UltraBlame original commit: 2fa7dd926187b890c08f8737de58c5b22e219da6