Bug 1229780 - Remove TextureClientPool helper functions from ClientLayerManager, r=mwoodrow
The fix for bug 1209801 made it so that TextureClientPools are
distinguished by a TextureFlags parameter as well as a
SurfaceFormat. The pools will create TextureClients satisfying both the
requested flags and format parameters.

The helper functions ReturnTextureClientDeferred, ReturnTextureClient,
and ReportClientLost were updated to use the TextureFlags of the client
to guess which pool the client came from. It is possible, however, for a
TextureClient to have additional flags to those that the pool requested
it have. This could then cause the clients to be returned to incorrect
pools.

Fortunately these helper functions were not being used anywhere, so
remove them.

UltraBlame original commit: 3e4476c099f9b629663fe0d7a8f2d82414b840ce