Bug 1233549. Disallow ES3 compressed texture formats. r=jgilbert
This fixes assertion failures on platforms that don't have
these formats. This is a temporary fix until we iron out
this issue in the spec. It's easier to add this later than
it is to take it away.

UltraBlame original commit: e1e0539dd11ea094f97be928bc3108b44cbd2f8e