Bug 1200169 - Making the slow add-on watcher more tolerant;r=Felipe
Showing alerts more than once is annoying for the user and basically
useless. We therefore change a bit our strategy:

- if an add-on has behaved correctly for the last 5 minutes, reset our counter of offences;
- don't display alerts for an add-on more than once per 6 hours.

The only exception is if the add-on freezes the browser (i.e. causes
it to stop for more than 5 seconds at a time), in which case we
display the alert regardless of past offences, up to once per 10
minutes.

UltraBlame original commit: e43b68cffc6bab2d379308086b6a9bc6b7c581ce