Bug 1239366 - Remove leading 'v' from detected versions. r=mshal
The leading 'v' was intended to distinguish the detected from
the literal version string, but this is non-obvious, and could
be confused with the git tag convention.

Instead, make the strings match, but put the detected version
first since it's what configure is likely to act on.

UltraBlame original commit: 4f1c44b5377e8df5374052f7d0c9176cd4d1f052