Bug 1239603 - don't reject format 12 cmap with odd encoding. r=jfkthame
This fixes an issue affecting the use of Big Caslon under OSX which uses an improperly labeled format12 cmap with platform=0/encoding=1. Technically, a Unicode 1.1 encoding implies a font that's not entirely compatible with more recent versions of Unicode (e.g. Hangul range is different). But since this seems to be simply a font error and not a real problem for format12 cmaps, ignore this inconsistency.

UltraBlame original commit: 6b46cf41e92fea85e575d1e63dbb6e0343dbaa53