Bug 1239210 - Don't process WebIDL files when not compiling; r=glandium
The output of WebIDL codegen is a bunch of .h and .cpp files. These
aren't relevant when not compiling.

This change appears to shave ~3-4s off the "export" tier time for
--disable-compile-environment builds on my i7-6700K because WebIDL
codegen is currently a long pole in that tier. After this patch,
artifact clobber builds are ~43.5s.

UltraBlame original commit: cf0e0332832e2a8c774776c0f301751470c4e375