Bug 1240056 - Disable font anti-aliasing on desktop-test image. DONTBUILD. r=dustin
In bug 1239301 we noticed that the fonts on the releng Linux hosts versus the
docker image were slightly different and caused some test failures.
The difference is that on the releng hosts we have anti-aliasing disabled and
many tests (reftests mainly) had adapted for this configuration, hence, failing if enabled.

In this bug we're going to disable anti-aliasing for the desktop-test image in
order to pass the tests.

See the difference between the two pushes:
https://treeherder.mozilla.org/#/jobs?repo=try&group_state=expanded&revision=3c9a54d68c95,1d3078c86e52

We went from failing over 200 tests across jobs to less than 10-20 tests in total.

In the future we will have to consider if we want to fix this configuration.

UltraBlame original commit: fefe66d4bc837c652d44b137cad7b420d51500ad