Bug 1241697 - Upload experiments with core telemetry ping. r=mfinkle
We generate the experiments in the TelemetryPingGenerator, rather than passing
them into the Intent, meaning that if the UploadSerivce gets killed and the
Intent gets redelivered, the experiments will be re-generated and there is a
chance the uploaded document won't be identical to the one that would have been
uploaded (or was uploaded, if the service is killed after the upload occurs).
The way to fix this is to pass the experiments into the Intent so it gets
redelivered identically.

However, it's not worth making the documents identical for the moment. It will
be fixed when we implement the functionality to reupload already-generated
documents which we believe to have failed to upload (which is necessary for us
to specify "counts since last ping").

UltraBlame original commit: 77c38dd031d41fb5176dd6ce8d5c6e79c5d7510d