Bug 1243234 - Move mp4parse-rust initialization into constructor and clean up try_rust.  r=giles
Initializing in the constructor better matches libstagefright's behaviour
(and avoids reading and copying the stream contents every time
GetNumberTracks() is called).

Also restricts the size of the buffer to 1MB.  This will be handled in the
future by passing the parser a DataSource-like interface for reading from
the stream.

UltraBlame original commit: e3212c974c62fd43bb01332702b8885830d43ed6