Bug 1239208 - Fix intermittent browser_rules_edit-property-computed.js by waiting for the correct ruleview-changed event. r=pbro
The preview also causes a ruleview-changed event to be emitted. If the frame
script happened to detect the change fast enough, the test would continue to
commit the value and start to wait for a ruleview-changed event. If the event
triggered by the preview had not been emitted, the test would continue
immediately once the event triggered by the preview was received but before
the value was actually committed and reflected in the computed rule list.

The fix is to wait for the ruleview-changed triggered by the preview before
committing the change to avoid mixing these two events.

UltraBlame original commit: eb673021691c3c3257546995edf4731a1a7c165c