Bug 1234629 - Post: Make bouncer APK Java package org.mozilla.gecko, like the main APK. r=me
This small change is a follow-up to Bug 1242213, which did the same
for the main package.  This is a nod to the future and Gradle, which
cleanly splits the internal Java package (org.mozilla.gecko) from the
external Android package (org.mozilla.fennec and friends).

UltraBlame original commit: 400c030d3c01d8818fe07d4547e0c145e19f099e