Bug 1247979. Optimize ScaleToOutsidePixels. r=kats
This does a union of all the rectangles at once, instead of one-by-one. It
also cuts down a lot on allocation churn.

It would be possible to avoid an additional allocation/copy if we change
the interface of this function to mutate the existing region instead of
returning a new one.

UltraBlame original commit: 537dbc2844087425e41da1cc43517aaad0b42d16