Bug 1250307 - Add the `by: "bucket"` breakdown option; r=jimb
This commit introduces `{ by: "bucket" }` breakdown type. Similar to `by:
"count"`, this breakdown type is a leaf and doesn't further categorize or filter
nodes that it counts. Instead, it accumulates them in a bucket and reports them
as a array of node IDs.

MozReview-Commit-ID: EeyWlrj9ujf

UltraBlame original commit: 745ce98cba7d5fb201996762527bfe868d58b345