bug 1250818 - remove certificate issuer organization to common name fallback r=Cykesiopka
Before this change, if a certificate's issuer DN did not have an organization
component, nsIX509Cert.issuerOrganization would fall back to using the issuer
common name. This was never a good idea, because this gave misleading
information to consumers of this interface. Furthermore, it appears that all
consumers of this interface already do such a fallback (for display purposes)
when they've determined that it's a reasonable thing to do.

MozReview-Commit-ID: p2gmSP0nZW

UltraBlame original commit: 7cfb28e22ec2641951d8b7a414fdf6f1bd09ca32