Bug 1251755 - markupview: add z-index 1 to pseudoclass ind.r=pbrosset
With Bug 1195120, the ".tag-line .tag-state" has a z-index of 0 instead
of -1. This element is used to display the colored background of a selected
row.

Adding z-index:1 to the pseudoclass indicator ensures it is displayed above
the .tag-state element.

MozReview-Commit-ID: K4aWyJXevWf

UltraBlame original commit: 5ee60bdfe6b8ec27522ee6dfaf43dc6d6ba99a1d