Bug 1253244 - Correct x,y offset type checks and add tests; r=automatedtester
The xOffset and yOffset arguments to element.coordinates previously
checked loose falsiness.  If a number <= 0 would be given as an argument,
they would not be used and it would return the centre of the element.

This change adds more checks, documentation, and tests.

MozReview-Commit-ID: KmsiVFipzWx

UltraBlame original commit: a62776417e00d7589b7540eef788a69c57b55ba2