Bug 1254873 - Make --disable-js-shell the default for non-js-standalone builds. r=chmanchester
The implementation is a bit circumvoluted, but we do need to share
options between the top-level and js/src configures, possibly with
different defaults, and to properly pass things down from one to
the other. Until we are further down the road and can actually merge
both configures, this is a necessary evil.

UltraBlame original commit: 72e0643586509c61b8090e5f93f7b4185976d8dc