Bug 1255173 - Fix onZoomReset() in test_wheel_default_action.html to work with e10s. r=masayuki, r=jmaher
"test.events.async.enabled" must be set to true so we generate key
events in the parent process, which is what triggers the zoom reset.

In addition, some SpecialPowers infrastructure must be set up to
listen for "browser-fullZoom:zoomReset" in the parent and send it to
an observer in the child.

UltraBlame original commit: 3e48de1a3b6cadf976c4f57707cea1553844e944