Bug 1144096 part 18 - [css-grid] Fix a couple of bugs in how we handle child existing continuations when pushing/pulling children.  r=dholbert
There were two problems in the existing code (which was exposed by tests
that dynamically insert/remove items).  First, the situation when
we have some pushed items two or more fragments away and then pull up
those.  This creates a "hole" in the child next-in-flow chain like so:
grid-container-frag-0
  child1-frag-0
  ...
grid-container-frag-1
  ...
grid-container-frag-2
  child1-frag-1

After we reflow grid-container-frag-0 and it's still incomplete we will
reflow its NIF, grid-container-frag-1, but it will "stall" since it
doesn't have a continuation for child1.  We need to make sure to
always pull up a fragment for that child.  That's what the first hunk
is about in the patch.

Second problem is the opposite problem of pushing a child into a NIF
container that already has an OC child continuation, like so:
grid-container-frag-0
  OverflowList = { child1-frag-0 }
grid-container-frag-1
  OverflowContinuationsList = { child1-frag-1 }

When we reflow grid-container-frag-1 we'll pull in child1-frag-0
like so:
grid-container-frag-0
 ...
grid-container-frag-1
  PrincipalList = { child1-frag-0 }
  OverflowContinuationsList = { child1-frag-1 }

This is bad since we'll consume BSize twice here.  The fix is
to move it our ExcessOverflowContinuationsList instead, like so:
grid-container-frag-0
 ...
grid-container-frag-1
  PrincipalList = { child1-frag-0 }
  ExcessOverflowContinuationsList = { child1-frag-1 }

That's what the second hunk in this patch does.

UltraBlame original commit: 47452506ec5632ec4e535d6618552a8a7d804991