Bug 1257153 - Cancel implicit wait timers on rejection; r=automatedtester
This fixes a leak where timers were not cancelled upon promise rejection.
When the function passed to implicitlyWaitFor is rejected for not finding
elements after the timeout has elapsed, the elementSearch callback is
no longer called because the timer is cancelled when it calls reject().

MozReview-Commit-ID: FR2iA8s1NjX

UltraBlame original commit: 1fdf06c430110676ac5feaf35a0462239b805cff