Bug 1258153 Initialize mozilla::widget::Native key with actual inputting character if it's created for WM_CHAR r=m_kato
When user types a Unicode character with Alt + Numpad keys, recevied messages are:

1. WM_SYSKEYDOWN for Alt (VK_MENU)
2. WM_SYSKEYDOWN and WM_SYSKEYUP of Numpad keys
3. WM_KEYUP for Alt (VK_MENU)
4. WM_CHAR for the Unicode character

The WM_CHAR indicates Alt key with its scancode.  Then, we shouldn't set the key value of the keypress event as "Alt".  Instead of that, we should set key value to the inputted Unicode character.  Then, the keypress event won't be supprressed by WidgetKeyboardEvent::ShouldCauseKeypressEvents() in TextEventDispatcher (this is what the cause of bug 1253153 is.

MozReview-Commit-ID: 8KjOWvYcKMs

UltraBlame original commit: 8786bbb9702f808a01ae45a5153440a7c44699fb