Bug 1245424 Part 1 - Fix assert for a display:none summary on debug build. r=bz
If the main summary element has 'display: none' style, it won't
generates a summary frame as the first child of the details.

However, if a details element have two summaries and the first summary
has 'display: none', the second summary still generates a SummaryFrame
event if it isn't the main summary. So instead of checking on the
SummaryFrame as before, I check the content tree for the main summary by
using the idea in bug 1245424 comment 8. Another reason might be the
potential removal of SummaryFrame in bug 1258657.

MozReview-Commit-ID: H0evZ17zj5k

UltraBlame original commit: 01e5078734f559dd1bc8503196c8fa5af33c7385