Bug 1249288 - Add default search engine to core ping. r=rnewman
The default search engine attribute may be null in the core ping if we haven't
been able to retrieve the value yet. It's unclear when this might be, but the
possibility is in the javadoc of `SearchEngineManager.getEngine`.

MozReview-Commit-ID: IrJB6GyjyTO

UltraBlame original commit: 6ecd4b60ef8ad9fe5517eff40ba4810ca6069a52