Bug 1251150. Add crash annotations if image visibility is re-entering. r=mats
The previous annotations only checked if the re-entrancy happened via DecrementVisibleCount.

The check in RebuildApproximateFrameVisibility is not needed because we add a check in DoUpdateApproximateFrameVisibility.

The check in ClearApproximatelyVisibleFramesList is not needed because we add checks in DoUpdateApproximateFrameVisibility and Destroy. The other caller, ClearApproximateFrameVisibilityVisited, is covered because it is only called from DoUpdateApproximateFrameVisibility.

UltraBlame original commit: 6bf6edf222a15276ca24855fd9f6bae559595524