Bug 1259428 - part 5 - convert Message subclasses to constructor functions; r=jld
All we use our Message subclasses for nowadays is the constructor, so we
might as well strip the class away and just have functions that perform
the construction for us.  This change eliminates unnecessary vtables as
well as making the included headers somewhat smaller, which is always
nice.

UltraBlame original commit: cb6dee617908f71f565946ca2ee78e446b01948c