Bug 1262729 - Allow to more conveniently build toolchains. r=dustin
Currently, we have in-tree scripts that allow to build our toolchains.
We also have a taskcluster script for clang that can be used in with
manually created taskcluster tasks. I wrote a similar script for gcc a
while ago, for the same usage.

This change hooks all these up such that one can do a try push with
`-j linux64-clang` or `-j linux64-gcc` and get a toolchain built as a
result.

It also hooks things up such that changes to the toolchain build scripts
trigger those jobs as well, so as to ensure they are not broken by the
changes.

UltraBlame original commit: bba8a0264e24d8c5231d12ecfd518a2ede9d255b