Bug 1266047 - Fix crash in mozilla::AudioBufferAddWithScale_SSE r=padenot
This adds alignment checks to fallback to scalar operations if the
received buffers are not properly aligned. Bug 1266112 is the follow on
to either fix the alignment problem or add a vector path for the aligned
portion of the buffers.

MozReview-Commit-ID: 5HCXzipXlqD

UltraBlame original commit: 1d6411584d7b5515be877eb65324e3deefc8b1be