Bug 1263571 - Pre: onTabChanged data is always a String, declare it as such r=liuche
We're always using this as a String, we might as well make this explicit to avoid having to cast
anywhere. (We very rarely use the parameter, but some new code in the main part of Bug 1263571
would have to cast this to a String. We can avoid that if we just use the correct type.)

MozReview-Commit-ID: H8JdMzQtmRI

UltraBlame original commit: 332bbe886f1159709a787266fc26d3e3c1384a86