Bug 1263472. Limit the height of svg embed in image/test/mochitest/test_svg_animatedGIF.html. r=xidorn
The extra height of the test element (the embed) versus the reference element (a div with height 40px) can cause the iframe the test is run in to overflow, thus causing a scrollbar in the test but not the reference. But we are only interested in testing the frame of the contained animated gif.

UltraBlame original commit: c18a17e91f35ee8e4d4448d563932c688c7f9235