Bug 1263910 - Make AudioBufferAddWithScale handle unaligned buffers; r=padenot
ReverbAccumulationBuffer often produces unaligned buffers due to the way
it wraps around results. This modifies AudioBufferAddWithScale on SSE2
platforms to handle unaligned buffers by performing scalar operations
until both the input and output buffers are aligned to 16 bytes. It then
does as many vector operations as possible and switches back to scalar
operations for anything that is left over.

This could also be done within the ReverbAccumulationBuffer code but doing
it directly within the AudioNodeEngine code makes it available to other
callers in the future, at the cost of a few extra branches in the case
where everything was aligned anyway.

MozReview-Commit-ID: Ky0uIe5LMVq

UltraBlame original commit: 24e883121fcd199ab32d7148ab9e8bd74ffc50ed