Bug 1270317 - Upgrade to Mercurial 3.7.3 in the mozharness test environment; r=jlund
We had a test environment running on Python 2.6 and an ancient version
of Mercurial. AFAICT we run Python 2.7 everywhere, so this environment
can be dropped.

We also upgrade to Mercurial 3.7.3, as that is what automation now runs.

MozReview-Commit-ID: 7WTyD3CUjtj

UltraBlame original commit: 0e3c3535be0264138fb4fd3d0e85a6b42661e472