Bug 1270317 - Use robustcheckout extension for checking out Mercurial repos; r=jlund
Functionality for doing an optimal clone/pull+share+purge+update is now
implemented in the robustcheckout extension so it can be implemented in one
place and used by all the various tools needing to perform a "robust"
checkout using optimal practices.

This commit switches the MercurialVCS to use it.

Functionality for interfacing with shared repos and associated tests have
been removed because this is all implemented and tested in robustcheckout.

Various other tests have also been removed because they are redundant with
tests in the robustcheckout extension.

MozReview-Commit-ID: FGvmSHKM5e0

UltraBlame original commit: 7ca64e0cf6f1250637f659781a782fa9bedc8224