Bug 1270870 - Create specified profile directory if necessary; r=snorp
When getting a profile with a specific profile directory, we used to
require the directory to already exist. This patch makes sure the
directory is created if necessary. It lets us treat the guest profile as
an ordinary custom profile at a specific location.

UltraBlame original commit: 556334657e54db71e7f40ea7f92daec926d56bd0