Bug 1270870 - Check profile when delivering pending push message; r=nalexander
For pending push messages, we used to check for correct profile when the
message is received, but it's better to check for correct profile when
the message is delivered to Gecko.

This fixes a crash in headless mode that the previous patches
introduced.

UltraBlame original commit: feb13c8b708aade4acf07e520682c2a9f9328544