Bug 1274655 - Set MOZ_SOURCE_REPO in mozharness to avoid variance from shared repos; r=jlund
Recent changes to mozharness in bug 1270317 started using pooled shared
storage for Mercurial repos. This means the "default" path in Mercurial
repos is variable depending on which repo was the first to be built on
a machine.

By default, the Firefox build system resolves the source repository
from `hg paths default`. This is now incorrect default behavior in
automation.

We fix the regression by setting MOZ_SOURCE_REPO in the environment to
path to the repository that mozharness is currently building.

MozReview-Commit-ID: 34IPf7PJfuA

UltraBlame original commit: a2a4c06d9736850782d8cc52802e0207ca1bf27a