Bug 1275783 - Don't use HgtoolVCS to query pushlog; r=jlund
The normal MercurialVCS now supports querying pushlog. Use it.

This isn't really relevant to the bug summary and other commits in this
series. But I already wrote this commit and was too lazy to create a
new bug for it.

MozReview-Commit-ID: C97Zgox3xKB

UltraBlame original commit: 266983e523325fea16101e27030c3d7540d8a5be