bug 1272503 -  make alerting for changes in num_constructors more agressive r=froydnj f=wlach
There shouldn't be any noise in this number, either a patch adds or removes
constructors or it doesn't.  Choosing 0.25% is kind of arbitrary, but it should
be good enough since it will make us alert for an increase of one constructor
so long as we have less than 400 constructors, and currently we have 98.

UltraBlame original commit: a0ecfb9c274386956d390493efe732d184a922dc