Bug 1276886 - [mozprocess] Fix hang in output readers when process is in a new process group. r=ahal
MozReview-Commit-ID: 9e6HNAaF0Yo

In case of in-process restarts it can happen that the new process gets forked into a new process group.
When that happens we loose the capability to kill the process. To prevent a hang when joining the output
reader threads in wait(), we simply skip that call by passing-through the IO error.

UltraBlame original commit: dd6fb8e96ae4b9b2c2150e0515919ca8985b672b