Bug 1277407 - Add Adjust sandbox keyfile to tree. r=sebastian
Additionally, we add this file to the tree so it can be used by bug 1277553.

The following commit will add the docs explaining how to use the file I
added.

The added file contents are slurped directly into
AdjustConstants.MOZ_INSTALL_TRACKING_ADJUST_SDK_APP_TOKEN. We get the sandbox
token from the previously working value in [1].

An alternative solution would be to remove the assertion that
`--with-adjust-sdk-keyfile` is specified and provide a default value if the key
is not specified. However, after investigating bug 1277553, I dislike this
approach because we lose the compile-time checks that a key file is specified,
which is dangerous for release/beta builds where the key file may not be valid,
we push a release, and we don't get the data we expect.

Ideally, we specify `--with-adjust-sdk-keyfile` for non-MOZILLA_OFFICIAL
builds, but I don't know how to do that.

[1]: https://hg.mozilla.org/mozilla-central/rev/ac75f8f4c01d#l1.26

MozReview-Commit-ID: HPNgiwfzU5p

UltraBlame original commit: 6dde3c79db7dc15038b66c6cffdcbf7c1678712e