Bug 1251362 - Part 1 - Increase SwipeRefreshLayout weight. r=liuche
Otherwise, depending on the device's display dimensions, a second smart folder would be overlapped by the empty view message that is displayed when no history is present for the history panel to display.

Once we've updated to a newer version of the support library (see bug 1267884), we should revisit this and see if using "wrap_content" is working properly instead.

MozReview-Commit-ID: 1xBCeiST9n4

UltraBlame original commit: 35e71e69db49f4672512db7b26dc2ad517d7cf81