Bug 1266450 - part2: remove iframe container for HTML tooltip;r=bgrins
In order to have tooltips with a variable height, the tooltip container
should be allowed to resize itself on the fly, which cannot be achieved
with an iframe.

This changeset makes the HTMLTooltip rely on a HTML container inserted
in the XUL document directly. This allows to go back to a synchronous
API which also simplifies the implementation.

MozReview-Commit-ID: EDcsnVSKmeU

UltraBlame original commit: 4e4b15f80b8085a7dc7d60387af1d4b1888a1f05