Bug 1277644 - Make sure browser_Addons_sample test add-on tests CPOWs when it is supposed to. r=Yoric
browser_Addons_sample is used by the browser_AddonWatcher.js test to make sure
we can properly detect when an add-on consumes a bunch of CPU or does a lot of
CPOW traffic.

There's a race condition in the add-on where what is supposed to be a CPOW
might not always be a CPOW, so when we try to cause a bunch of CPOW
traffic, we don't get the expected performance warnings.

This makes sure that when we try to simulate CPOW usage, we do it
with an actual CPOW.

Additionally, this commit also includes the unpacked source of the
add-on, which before wasn't in the tree. I've also taken the liberty
of bumping the add-on version and signing it.

MozReview-Commit-ID: GICLYpi8Kon

UltraBlame original commit: 7e1dacbe0001749e96c1c4fca6db06d5dadc6e6f