Bug 1264948 part 1 - Register if the LifoAlloc is supposed to be infallible or not. r=jonco,h4writer
This patch adds a new flag to the LifoAlloc structure, which would be used to
assert when we attempt to allocate a new chunk for the LifoAlloc.  This ensure
that we assert (in debug builds) if we attempt to allocate beyong the reserved
space of the ballast.

UltraBlame original commit: a3f3f077a1ec0414343598d5f548ab2161286285