bug 1257718 look for new events as time advances r=padenot
|bailout| is reset for each aTime, so that the appropriate events for that
time can be found.

The |eventIndex| loop is adjusted so that, when it is re-entered, it keeps
the current set of events if they are appropriate (instead of advancing
every time it is entered).

|previous| and |next| are now advanced even when passing the last event,
removing the special case when past all events.

MozReview-Commit-ID: 8ZSIzKKGQbd

UltraBlame original commit: 48d91469d4bc6943d28c90096dbe733c4197d04b