Bug 1283229 - Thread `AutoLockHelperThreadState&` parameters through various `HelperThread`-related functions; r=terrence
This is a pre-requisite for moving the `HelperThread` infrastructure over from
using `PRLock` and `PRCondVar` to using `js::Mutex` and `js::ConditionVariable`.
Its needed because waiting on a `js::ConditionVariable` takes the `LockGuard`
for the condvar's mutex as a parameter, so we need to have that accessible. When
we do the conversion, `AutoLockHelperThreadState` will become derived from
`LockGuard` and specialized for the helper thread state lock, and so the guard
will therefore be accessible.

UltraBlame original commit: e5ad05d9cad4ac105f5f580887202a5041e197b4