Bug 1286993 - Add build/pymake to the SEARCH_PATHS in mach_bootstrap. r=gps
It used to be there, but was removed in bug 1151124 because what I
thought was the only use of pymake in mach was removed in that bug.

Which, it turns out, was not the case. So restore it, so that mach
empty-makefiles works again.

UltraBlame original commit: 37b4f4a716a4ebdecb754546a4550e6df8855ca6