Bug 1287011 - Remove preference gfx.filter.nearest.force-enabled.  r=gw280
This preference was added in bug 747274, and the need for it (i.e.,
having it default to a different value on Android) was removed a few
weeks later in bug 750598.

Whether we honor the preference depends on which drawing function we go
through, which bug 720531 will make even more inconsistent than it
already is (by causing it to be honored for border-image-repeat: space,
but not other values of border-image-repeat).

Rather than trying to clean it up, it seems best to remove it, since its
initial purpose appears to have been having a different default on
Android, which we longer need.

MozReview-Commit-ID: IFW2Ukwx4fH

UltraBlame original commit: 0bb4cf8a575985a1d013097836f1fb8281a7d6a2