Bug 1164027 - Do not call glTexImage2D if likely to fail on Adreno 3xx devices; r=jrmuizel,snorp
There is a bug in the graphics driver for Adreno 3xx GPUs on android
versions up to 4.4 which segfaults if a texture allocation fails due to
virtual memory exhaustion.

On affected devices, guess whether an allocation would succeed and avoid
calling glTexImage2D where we believe it would fail. As the texture will
not be allocated this will likely cause problems such as black
rectangles appearing on the screen, but that is preferable to crashing.

MozReview-Commit-ID: 7xVI22pAdwb

UltraBlame original commit: fff0c4030b81ceb8b98d4db74807b746f403770c