Bug 1288103 - Add experimental MOZ_ANDROID_ACTIVITY_STREAM build flag r=chmanchester
This will be used to enable the activity stream panel in place of the HomePager. We are
likely to migrate this to a switchboard flag in future once the new panel becomes
shippable (we are still investigating other distribution mechanisms, so it is entirely
possible this will completely change in future).

MozReview-Commit-ID: I9VSliO0IXE

UltraBlame original commit: 268d2ffd7a0541073df31abd121e0e6002cb73c4