Bug 1275142 - Allow a defect of restyle when the target element of SMIL animation is associated with an nsIFrame. r=birtles
Actually the check of restyle count after associating with the nsIFrame is
not a problem in this test.  The purpose of this test is to check no restyles
when the target element has no associated nsIFrame.
We will fix the defect in bug 866411.

MozReview-Commit-ID: BFRJsbTGdkh

UltraBlame original commit: c83fe95ad2784208327687793eb4766d83afaa7b