Bug 1293186: Do not assume that currentTime is still at seeking position when seeking event is fired. r=jwwang
The seeking attribute is changed synchronously, however, the seeking event is fired asynchronously. It is possible that by the time the seeking event is fired, the seeking operation has already completed and that the "time marches on", causing currentTime to be past the seeking position.

MozReview-Commit-ID: IB83w6zKCLF

UltraBlame original commit: 4fa9a49c57aa0fdc4c3506c6aae736e3bb2594d3