Bug 1293957 part.2 TSFTextStore::GetTextExt() shouldn't return TS_E_NOLAYOUT after hacking the query range r=m_kato
When TSFTextStore::GetTextExt() needs to hack the queried range, it means that the method should not return TS_E_NOLAYOUT since the hacks are implemented for that.  However, currently, if hacked offset is still at modified offset, it returns TS_E_NOLAYOUT.  This causes flicking IME windows.

For avoiding that, this patch adjust the hacked offset to the last unmodified offset or offset of the start of composition string. Even if there is no modified character in the composition string, we can use the offset since even if there is no character at the offset, query content event returns caret rect at the offset and if there are old characters, it returns its rect which the user still see.  So, the rect at the composition start is useful until layout is modified actually.

MozReview-Commit-ID: 4zo9mF3m9u9

UltraBlame original commit: beb5314cb9d546942bdc8137bf5ac9b2c8a5d8ec