Bug 1289805 - Don't pass in --exclude to flake8 paths that contain custom configuration, r=smacleod
This fixes a bug in flake8, where if you pass in --exclude to a path with a custom .flake8 file, that
configuration will be ignored. I'm not sure why this happens.

Prior to this commit series, the 'exclude' paths weren't being passed on to the flake8 linter properly.
This is why the problems hasn't surfaced until now. This is a band-aid fix until a proper (likely
upstream) fix can be landed.

MozReview-Commit-ID: KYhC6SEySC3

UltraBlame original commit: 98876f70367e0ab16ab8cd239d019efbf80f15be