Bug 1279650 - kill off useless warning for timed out favicon requests that have already been cancelled when the window closes, r=mikedeboer
As it is, when favicon requests time out, the timeout handler doesn't remove the loadData and request from gFaviconLoadDataMap after cancelling the request,
which means that when the tab or the parent window is closed we try and cancel the same thing again, which leads to errors in the console. This patch
addresses that issue by making timeouts also remove the loadData from gFaviconLoadDataMap.

MozReview-Commit-ID: FCeRFXjeXdJ

UltraBlame original commit: fbc71dcb173ac4f83d24653e692b2b2599205d48