Bug 1295588 - Add createEvent() telemetry; r=smaug, data-review=bsmedberg
The createEvent() method supports only a fixed list of legacy event
types, and we would like to converge with other browsers on a list of
events to support.  To this end, we want to know if any supported types
are unused so we can drop them.

MozReview-Commit-ID: 8uB1vepBCpi

UltraBlame original commit: f0f12701aeb78e99fe3fc19c608dedb41293409a