Bug 1297043 - Fix TableWidget selection when wrapTextInElements is true. r=jsnajdr
We were using the onMousedown event target to select a row, based on a data-id attribute of the cell node.
But when the wrapTextInElements is true, there is an extra span element inside the cell. And in this case,
we were trying to retrieve the data-id attribute on this span, which led to calling the selection function with a
null id, and thus, not selecting the row.
We fix this by retrieving the attribute from the correct node.
A test case was added in order to make sure the bug is fixed

MozReview-Commit-ID: Jcf42AZb2uk

UltraBlame original commit: a7ec26fab9e2cc7ef3fd99a919f09dc884b25f0f