Bug 1293595 - Make doorhanger overlay visible by default, and retrieve it during onCreate r=sebastian
Retrieving the doorhanger during onAttachedToWindow seems to be unreliable, resulting in NPE's when
it doesn't work (the only logical explanation would be that onCreate hasn't been called yet, meaning
we don't have a layout yet). We should instead findViewById() it during onCreate, where the layout
is guaranteed to exist.

Since we show the overlay on all devices now, we can make it visible by default, whereas in the past
we had a condition to only enable it on > 2.3 devices.

MozReview-Commit-ID: 1DGVAPSX0mi

UltraBlame original commit: fcdac0d4b2d59e303f6de50915f5420581dda5c7