Bug 1302907 - Run mozharness from source checkout if available; r=dustin
Currently, build tasks produce a zip file artifact containing the
content of testing/mozharness.  This zip file is downloaded and applied
by any test using mozharness (which is most of them).

Now that some tests have a source checkout, we don't need to download
a mozharness zip because we can just use the source checkout.

This commit teaches test-ubuntu.sh to accept a custom path to
mozharness files. If that path is defined, we skip downloading the
mozharness zip.

The taskgraph has been updated to pass the path of mozharness in the
checkout to test-ubuntu.sh when a source checkout is available.

MozReview-Commit-ID: 2P17Wx1oytI

UltraBlame original commit: 21fdf73bbb17e34cfe00e372695c4f21e4ba3e6a