Bug 1303707 - Fix collecting of formats-at-issues - r=Gijs
Array.filter was given the wrong information, so we were only adding
existing name, instead of adding new ones.
This is not critical, as this is only useful to better find when issues
are fixed, that's why I'm fixing it in a fly-by here.

MozReview-Commit-ID: CGkN3BbSTge

UltraBlame original commit: c71904aaef6a13bb92159f650e806919460d0887