Bug 1304145 - 3. Move GeckoService and notification package to Fennec; r=nalexander
GeckoService and the notification package have some interdependencies,
so if we want to move the notification package, we have to move
GeckoService also. With that said, it's good to move GeckoService in any
case, because it's a Fennec component just like GeckoApp.

UltraBlame original commit: d676a611d3ac68dfb398527fae4026629f1b0775