Bug 1303288 - Pass the window object instead of current scope to prevent wrapper issues when using the devtools reload addon. r=jlongster
If we happen to pass the current scope (`this`), the browser-loader codebase fails
when passing it as sandboxPrototype. It only works when passing an xray wrapper.
The document scope happen to not be an xray when using the addon, whereas `window`
always is no matter if we are using the addon or not.

MozReview-Commit-ID: GjYHkaCGBDd

UltraBlame original commit: 1aaa9010a9a18ca16ee3dfedfc8aa0f085fb4510