Bug 1247056 - More flexible SUMO link - r=gijs
Instead of always pointing to the same fix-video-audio-problems-firefox-windows
SUMO page, make it possible to point at others depending on the issue.

Also we won't show a "Learn How" button for WinXP anymore, as the linked
page is not relevant to XP, and CDMs are not available on XP anyway.

MozReview-Commit-ID: 7YR43Wa0IRe

UltraBlame original commit: 03553484a062c5361a8eed0a3c38ba8a3ed5f404