Bug 1247056 - More flexible SUMO link - r=Gijs
Instead of always pointing to the same fix-video-audio-problems-firefox-windows
SUMO page, make it possible to point at others depending on the issue.

Also we won't show a "Learn How" button for WinXP anymore, as the linked
page is not relevant to XP, and CDMs are not available on XP anyway.

MozReview-Commit-ID: 7YR43Wa0IRe

UltraBlame original commit: 93c70da4ed72a9aa11824144beed3280db74cb8b