Bug 1305775 - Output more metadata about files fetched to increase debuggability. r=ted
In bug 1305752 we discovered that we download a zip file into memory without any issues, however,
when we tried to unzip we discovered that we have an invalid zip file.

The information in the logs is not sufficient to determine what could be the root issue.

MozReview-Commit-ID: DKwDUCmUhFF

UltraBlame original commit: e709712a33e7bdfd8f8e8a6cb82b200f95b8b285