Bug 1305157 - Handle the build/ directory's GENERATED_FILES; r=gps
This is a little tricky since tup expects rules in a single Tupfile to
be topologically sorted, and the emitter emits the GeneratedFile object
before it emits the FinalTargetPreprocessedFile object. We work around
this by delaying when we process the application.ini.h GeneratedFile
object, but it's a hack.

MozReview-Commit-ID: 8nVm76nZOKb

UltraBlame original commit: 694718beec802d895597ce5e3d25655e33276a5d