Bug 1305836, part 1 - Remove CCAnalyzer leak checker. r=jgriffin
This leak checker may be triggering a shutdown leak on Windows,
doesn't work with e10s, and should not be needed now that ttaubert
fixed the ++DOMWINDOW leak detector to work.

The additional GCs and CCs this patch adds used to be run as part of
cc-analyzer.js, and are needed to avoid window leaks in tests.

MozReview-Commit-ID: IzZI6h2SCr2

UltraBlame original commit: 89a22098629e7e18383d794b78ee0700a7b28d96