Bug 1299694 - Ensure we don't enable Widevine unintentionally. r=glandium
On FreeBSD the target.kernel etc checks in enable_eme are failing,
but we're still falling through to |return value|, and so Widevine
is being enabled. If we remove the |return value| from enable_eme
we at least make Widevine disabled where it's not supposed to be
enabled.


MozReview-Commit-ID: D1h0IUidxhv

UltraBlame original commit: 7f7c7197e8ea6311df402e2bf32506f530f59399