Bug 1289249 - Make version control interaction generic; r=dustin
Previously, we assumed we only could have a single version control
checkout: Gecko/Firefox. The code reflected this by not passing
arguments to the vcs_checkout function.

Upcoming commits will introduce the need to perform a checkout of the
build/tools repository. In preparation for this, refactor the VCS
functionality so it is generic and can work on any repo.

MozReview-Commit-ID: B0Act9fz2Ee

UltraBlame original commit: db1b85faa8b06612fbe7fc536ac81a8090446ced