Bug 1289249 - Use run-task for mozharness jobs on docker-worker; r=dustin
This transform appears to always be executed on desktop-build.
build.sh always calls checkout-sources.sh and checkout-sources.sh
always performs a Gecko checkout. It /may/ perform a build/tools
checkout unless an environment variable is set.

Now that desktop-build is ready to use run-task, we switch this
transform to run-task. A gecko and tools checkout is *always*
performed. There is room to make the tools checkout optional.
But it completes nearly instantly. So I'm inclined to do this
as a follow-up enhancement, if needed.

Ideally, the checkout would be on a cache so subsequent tasks on
the worker only need to perform a purge+update as opposed to a
full checkout. However, many tasks make assumptions about the
VCS behavior and modifying the checkout path is a non-trivial
amount of work. This can be deferred to a follow-up bug.

MozReview-Commit-ID: 5Auerl4OL9x

UltraBlame original commit: 5df2521786e17703c106f4d7253a3e701454515d