Bug 1307016 - Have puppeteer __str__ encode fields to ascii. r=maja_zf
At least some of the environments the tests are running in do not appear to play
nice with non-ascii characters. Some of the jenkins runs are currently unhappy
due to this. This patch sees that all fields are encoded as ascii with
non-compatible characters replaced.

MozReview-Commit-ID: 6qSCyUujMLE

UltraBlame original commit: c0817cce76da48fb6bfb3ff4111bed9f5de4603a