Bug 1307730 - Disallow CORS fetches when we have an expanded principal; r=bzbarsky
It's not possible to construct a useful Origin header when we have
an expanded principal and are about to perform a CORS fetch.  Therefore,
instead of sending a CORS fetch with an |Origin: null| header, we
must fail the request.

UltraBlame original commit: 997a19a765c8ad2f0b7dd07f4c8ff4126bfa352b