Bug 1308287 part 1.  Change [NeedsSubjectPrincipal] to only do the Maybe thing for interfaces that can be exposed to workers.  r=baku
The idea is to not make consumers think about whether the principal exists or
not when the caller knows for sure that it does.

The substantive changes are in dom/bindings, nsHTMLDocument::SetDesignMode, and
around the CanUseStorage bits.  Everything else is pretty mechanical.

UltraBlame original commit: 3e0c2cad052c8752f84ef746c2b54371971afb1f