Bug 1304950 - Reduce timeslice to avoid races with the source ending. r=jwwang
Media element capture pushes data to MSG ahead of currentTime which together
with the direct listeners that we use in MediaRecorder we end up finishing
the recording in less than the 250ms that this test uses as the recording
timeslice.

Lowering the timeslice here seems to fix this. I'm using 1 here since it's the
minimum valid number.

MozReview-Commit-ID: KAlRoHWHPSV

UltraBlame original commit: 7480f3d08b18ee0d347a944020509c2d57bb5cf0