Bug 1275314 - Add an API to allow flushing out in-progress checkerboard reports. r=botond,ehsan
This is useful for talos tests that record checkerboarding. In those tests, the
page might still be in a checkerboard state at the end of the test, so it may be
necessary to flush out the report for measurement.

MozReview-Commit-ID: CtafG4NAGHN

UltraBlame original commit: 77054c32f5dfb00945a85ea175af677a918c2120