Bug 1290372 - _wait_for_connection() has to return early if instance is not running. r=ato
In case when the instance is not running, the method would check for the port again
and again until the timeout is reached. This extra time can be spent if the process
status is checked.

MozReview-Commit-ID: C2WAWNC5CWE

UltraBlame original commit: 198c4bf0c8df9765e9eba5f12eb1f7a899c277e5