Bug 1311439 - Skip _notifyBackgroundTab for hidden tabs. r=Gijs
If a hidden tab is added at the end of a tab bar with enough tabs to show the
scrollbutton-down arrow, it would cause the arrow to highlight incorrectly,
suggesting there an additional tab to find, but in reality there isn't since it
is hidden.  This can occur with DevTools Responsive Design Mode which makes use
of hidden tabs.

By skipping `_notifyBackgroundTab` for hidden tabs, we avoid this issue because
we no longer attempt to highlight the arrow.

MozReview-Commit-ID: 2FoJ7UouJCL

UltraBlame original commit: a39915ca6ab440c37e382686dd4c8c1980008f5a