Bug 1288104 part 1 - Move XDR buffer to the caller. r=luke
Add a typedef on top of mozilla::Vector to define the TranscodeBuffer which owns
the encoded content of a Script / Function.

This modification renames JS_EncodeScript, into JS::EncodeScript, and change its
prototype to have a JSContext, a TranscodeBuffer, and a Handle on a script that
we want to encode.

Similar modifications are made to JS_EncodeInterpretedFunction, and the Decode
variant of these.

UltraBlame original commit: 4a8c5061f3b753cb34028b59c9ccc654027ba021