Bug 1287007 - Refactor test_ext_schemas_api_injection.js r=billm
When the background API move to a separate implementation,
then the schema APIs will be generated (and ChildAPIManager
will just delegate the implementation to the parent process).

The purpose of the test is to verify that nested namespaces and
null-feturn values work, so just use the base classes instead of a
concrete implementation for BaseContext / SchemaAPIManager instead of
setting up a full extension.

MozReview-Commit-ID: CB5s7Ae24zS

UltraBlame original commit: b91a39e959ffb29b8da8d41684181d17f92fef80