Bug 1198381 - Extend setTimeout handling in nsGlobalWindow, r=smaug
The requestIdleCallback feature behaves in many ways as setTimeout
since it takes an optional timout when the idle callback will be
called regardless of the user agent being idle or not. This means that
the same mechanisms controlling setTimeout are needed for
requestIdleCallback.

MozReview-Commit-ID: 9mESsJnUexf

UltraBlame original commit: ffaf3d371130200ed62ffb462386a191901d1421