bug 1312778 - Ensure histogram recording is enabled before remote accumulation r=gfritzsche
There is a mechanism to selectively turn hgram recording off and on. This is
presently only used to throttle an otherwise-devastatingly-chatty hgram to only
contain accumulations from specific times (sync scroll).

This was only checked on the parent. This checks it on the child as well.

MozReview-Commit-ID: 4f0VXMHLaPW

UltraBlame original commit: 443c4ea33338dd374ba3019d8ca039dacddb3b05