Bug 1312540 - Simplify behaviour of GMP::GetPluginVersionForAPI. r=gerald
We only have one version of a GMP installed at once anyway. This version code
is to support Adobe Primetime, and that's disabled.

Making the behaviour of this code simpler will make it easier to have the
same behaviour of the GetGMPVersion code in e10s and non-e10s mode.

I will be removing this code soon, but I will do that in a later patch, so as
to not complicate the uplift.

MozReview-Commit-ID: 3cn7GhihWzm

UltraBlame original commit: 8d38f007c9111c5fea264ac79914d1296a2992a3