Bug 1313731 - Append /v1 to fxaccount's autoconfig's auth_server_base_url if not present r=rfkelly
Currently the fxaccounts content server passes this through directly from its
configuration files, and accepts this as either containing the trailing /v1 or
not. On stage, it contains it, but on dev it does not.

In the near future, it should be possible to append the trailing /v1
unconditionally, as the FxA content server is being changed to never send it
down (which will be consistent with the other prefs), however it's done
conditionally as to not break autoconfig against stage in the mean time.

MozReview-Commit-ID: AStTm2hHVHQ

UltraBlame original commit: 1459baf15694fadda842beb9a10b42689bc7498c