Bug 1316040 - Change target to non-target in test_caret_not_appear_when_typing_in_scrollable_content. r=mtseng
In this test, we assume the caret does not appear at the end of the input.
However in this case, self.actions.flick() will select all text within the
input, and then send_keys() will replace it with '!'. As a result, the final
content might varies on different platforms. Test the final content match
tho target_content might not be reliable.

To fix this, we instead assert that final content does *not* match the
non_target_content as the bug occurs.

MozReview-Commit-ID: 2YRN8W3PUXG

UltraBlame original commit: b737f14975c2a83bf6c3753528641e0f51745778