bug 1316163 - remove eCoalesceMutationTextChange event rule r=davidb
The only difference between it and eCoalesceReorder is that for
eCoalesceReorder we assert the target is one of a outer doc, application
accessible or xul tree.  However we never put events that have
eCoalesceMutationTextChange as there event rule in the queue, so there is no
reason for separate event rules, and we can remove eCoalesceMutationTextChange
and check that queued eCoalesceReorder events are in fact reorder events.

UltraBlame original commit: 313605cfc6e21eb33abffd8d1e2262a6295f0e22