bug 1270916 - add methods to maintain a queue of mutation events to coalesce or fire r=davidb
Mutation events are kept in a queue before firing.   The queue is only
coalesced when necessary, at present this is when queueing a hide event, or
just before firing all the events.  It may be possible to do without the
former, but that is left as future work.  The state of what types of events an
accessible is a target of is stored in the accessible.  Combining that with a
map from accessible and type pairs to events we can quickly remove unnecessary
events from the queue when we need to coalesce.

UltraBlame original commit: 0919d58c4e9e09a5aab6bce2c703cecd4dcb92c5