bug 1270916 - use accessible->IndexInParent() in DocAccessibleChildBase::ShowEvent() r=davidb
Since we will fire events in the correct order we can use the index of the
event target at the time the event is fired.  This protects from weird cases
where the target is inserted, and then children before the target are removed.

UltraBlame original commit: 43708cd293faa60fc2855a437602cb82d8bb9715