Bug 1314573 - Uses Reps in long mode in the console; r=bgrins
In the MessageBody component, we already have the "mode" props we use
to pass it to the Reps. We only default it to "long" so most of the
message will use it.
We only need to specify to use "short" mode in ConsoleTable cells.

MozReview-Commit-ID: 6KyZkN86bHJ

UltraBlame original commit: ab533a828db674a270fdf113477d1ebbb0ebdacf