Bug 1301055 - part3 : modify test prefs. r=jwwang
On fennec we have the special workaround that is doesn't dispatch the "error" event when the error is
"MEDIA_ERR_SRC_NOT_SUPPORTED" because we will use an external app to open it.

But we don't want this behavior affect the tests we already have, so close the pref.

MozReview-Commit-ID: 9aoy1vnekvY

UltraBlame original commit: cb8ff0bdb0f8a3a4e648f552f30c1ff52cb4cae7