Bug 1103196 - Add ability to ignore invalid TLS certificates; r=automatedtester,keeler,mossop
When the `acceptInsecureCerts` capability is set to true on creating
a new Marionette session, a `nsICertOverrideService` override service
is installed that causes all invalid TLS certificates to be ignored.
This is in line with the expectations of the WebDriver specification.

It is worth noting that this is a potential security risk and that this
feature is only available in Gecko when the Marionette server is enabled.

MozReview-Commit-ID: BXrQw17TgDy

UltraBlame original commit: d207581536c0ddb8189c643324ffccf491dbbaf0