Bug 1313497 - Storages can be provided to InvokeAsync - r=froydnj
InvokeAsync' Storages work like NewRunnableMethod, where template
parameters must be given, to specify the storage of arguments that are passed
to the target method.
This is especially useful when target methods take references (or pointers),
and there is a choice between passing the reference as-is because the object is
long-lived, or by taking a copy otherwise.

If no Storages are provided, InvokeAsync will store (non-reference/pointer)
objects and move them into the target method parameters.

MozReview-Commit-ID: 9qTQf84bOMv

UltraBlame original commit: b6d2c739857fc05296877c8f0ddf4d6748ccebb9