Bug 1320362 - Move indexedDb storage type in the storage inspector into a new column r=jdescottes
Ideally we would have used the GUID to join fields to build the compound key but because we show the uniqueId in the tree that would lead to an ugly user experience. As an
alternative we simply use e.g. db1 (default) as the uniqueId. This works as a unique key, is informative for the user and means minimal changes to tests.

MozReview-Commit-ID: I0XxTIoV33p

UltraBlame original commit: 65c73bad7dc1e2f74873928bc2e954dade385c41