Bug 1321495 part 1: Add needed includes/namepsaces to headers within layout/style, to preemptively fix unified-build bustage. r=TYLin
In CSSLexer.h and ServoCSSRuleList.h, this patch is providing the type for the
header's main class to inherit from.
In nsCSSPseudoClasses.h, this patch is providing an enum type (CSSEnabledState)
that's used in this header file.

MozReview-Commit-ID: 1YUVyi47jmA

UltraBlame original commit: 99c9d215e0b8d6ea51ba156a64138e014338fb11