Bug 1321127 - Add a assertion_count action to mozlog, r=ahal
This is intended as a structured replacement for the assertion checks
that previously used unstructured logs. It adds a log action
assertion_count, which takes the actual number of assertions observed
during a test, the minimum expeced number and the maximum expected
number. It also updates the reftest harness to use this logging.

MozReview-Commit-ID: JgjLlaYuvSG

UltraBlame original commit: 2be3e937ecd92ed4cd0404cfaa71fa451805e8fd