Bug 1322382 - Add EffectCompositor::GetBaseStyle for nsIFrame. r=birtles
Currently this function is used when an nsIFrame is generated by reframing. In
reframing, we call EffectCompositor::UpdateEffectProperties, as a result, we
have to clear the base style that we resolved once. So, for such cases, we need
to re-resolve the base style with the style context that is associated with the
nsIFrame.

MozReview-Commit-ID: F90OuF44SPI

UltraBlame original commit: 35209805acb4e4fc08876c829f78c233f5c5d7b8