Bug 1319245 - Track rich telemetry data for Activity Stream r=sebastian
General concept is to populate the "extras" field with a stringified JSON object
which contains bunch of additional data which, in aggregate, might give insight
into user's actions.

A builder is used in order to make constructing the extras json string easier.
This builder has a concept of a "global state" which lets us easily include some information
with every A-S telemetry ping.
Currently this is used to track whether or not user has a firefox account.

An instance of a builder is passed around, augmented as more information becomes known,
and is materialized into an "extras" string whenever an action occurs and telemetry
event needs to be sent.

MozReview-Commit-ID: GDmxkWChnnA

UltraBlame original commit: 7e6bdb4d5fab56795d1ee46540126e05139a00f1