Bug 1311620 - Part 3: Incorporate null_t in Animatable. r=birtles
We couldn't tell the difference between a null StyleAnimationValue and
transform:none on the compositor. This was not a problem before, since we didn't
need the null StyleAnimationValue on the compositor, because the null
StyleAnimationValue have to be passed with composite:add flag, and in the case of
composite:add we just used the underlying value, i.e. we didn't use the null
value at all.
But for normal additive animations, we have to check the null StyleAnimationValue
to tell whether we are processing a missing keyframe or not.
So in this patch, Animatable can be null_t to represent the null
StyleAnimationValue, and as a result of this change, we can drop
BaseAnimationStyle.

MozReview-Commit-ID: Au41ujHgPpU

UltraBlame original commit: d36e5d3ad6d2dcfd043d7e3aec195f61ae0ac3af