Bug 1323703 - implement a method to create DXVA2 decoder. r=mattwoodrow
When we want to decode with DXVA2 aPI directly instead of using it by MFT, we need to take responsibility for
creating a decoder and handle the related decoding operations by ourself. So implement a method to create and
hold a ref to DXVA2 decoder for DXVA2Manager.

MozReview-Commit-ID: 4EyrsjzEyYK

UltraBlame original commit: 5a11af0e0fcfb45c248b46602a06a94e3530e618