Bug 1326496 - testing:mochitest: Use find_library to get libc filename. r=ahal
On architectures like alpha and ia64, the glibc does not use the
canonical ABI version number 6 but 6.1 and therefore the filename
of the C library is not libc.so.6 but libc.so.6.1. We're therefore
making the Python code more flexible and use find_library from
ctypes.util to determine the filename from the environment instead
of hard-coding it.

UltraBlame original commit: 156ce27e0e11b30f7545fd79f5488b5f16bbe54a