Bug 1326496 - testing:mozbase: Use find_library to get libc filename. r=ahal
On architectures like alpha and ia64, the glibc does not use the
canonical ABI version number 6 but 6.1 and therefore the filename
of the C library is not libc.so.6 but libc.so.6.1. We're therefore
making the Python code more flexible and use find_library from
ctypes.util to determine the filename from the environment instead
of hard-coding it.

UltraBlame original commit: ef5bd9f2e0cb44e27c5c233b19a266ca646c86e8