Bug 1147371 - Implement IteratorClose for array destructuring. (r=arai)
Since result.done is always needed now, always emit the code that pushes
it on the stack.

For throwing, like for-of, IteratorClose is only called when non-iterator
code throws. Unlike for-of, both the iterator object and the done
boolean value are on the stack for the trynote. IteratorClose is only
called when !done.

UltraBlame original commit: ce293b3c0a8b245a98592a0de2b62bbc0177269d