Bug 1331601 - Copy tab listener state flags in RDM. r=ochameau
When starting and stopping RDM, we need want to ensure tab listener state flags
are preserved for the tab involved, which is a bit tricky with the tab dance
RDM is doing.

The state flags ensure the browser will call the correct handlers when switching
tabs to update the primary browser UI.  For example, this is needed to correctly
update the enabled state of the view source command for the current tab.

MozReview-Commit-ID: 7lKY0DKxgJH

UltraBlame original commit: 0de2f57ed353e10176053d1e83f95d42f031c1be