Bug 1317079 - Avoid spurious exception when snackbar is dismissed. r=jchen
When a snackbar with a button callback is dismissed, this translates to a rejected promise from sendRequestForResult(). We need to catch this in order to avoid a spurious 'JavaScript Error: "uncaught exception: undefined"' message appearing in the console and possibly causing confusion.

MozReview-Commit-ID: 7hsAOAMTeDP

UltraBlame original commit: ebc9b049b54da0cb5120de12e96e1bdf617289e5