Bug 1334488 - [manifestparser] Improve error message when using an invalid comment character at the start of a manifest, r=jmaher
This refactors some of the error logic into a new IniParseError exception. In addition to checking for a 'name' property, we also
check for a 'path' property. This is because some file-like objects coming out of the build system use a 'path' attribute instead
of the standard 'name'.

MozReview-Commit-ID: EXXl9gt1MQk

UltraBlame original commit: a15be64c84409012527a846e13f5ff36b9478b59