Bug 900784 part 0.4 - Move aScriptFromHead to the ScriptLoadRequest. r=bkelly
Moving aScriptFromHead to a field of the ScriptLoadRequest ensure that all the
state to call StartLoad is either in the ScriptLoader or in the
ScriptLoadRequest.

Thus, we can later add new calls to StartLoad, if we have to restart a request
because of some decoding issues.

UltraBlame original commit: 3df44b58188df34f0f6fd661d65c33128ed6fe10