servo: Merge #11406 - Fix Android APK building (packaging) (from aneeshusa:add-missing-build-apk-semicolon); r=mbrubeck
Thank you for contributing to Servo! Please replace each `[ ]` by `[X]` when the step is complete, and replace `__` with appropriate data:
- [x] `./mach build -d` does not report any errors
- [x] `./mach test-tidy --faster` does not report any errors
- [ ] These changes help with #10339.

Either:
- [x] There are tests for these changes (android-nightly builder) OR
- [ ] These changes do not require tests because _____

Pull requests that do not address these steps are welcome, but they will require additional verification as part of the review process.

I missed this syntax error during my review, and we don't check that
packaging succeeds in regular build suite, but only as a Nightly build.

larsbergstrom should we run the packaging as part of the regular build? I think it is important that this continue to stay working, especially with #9918. In that case, the only difference with the nightly builder is whether or not to upload to S3.

Source-Repo: https://github.com/servo/servo
Source-Revision: 1efacd393843a761ee91c3a1765ef4801ccf18cc

UltraBlame original commit: 77e96730242d9191a03ab87ad2ced38d0a84d379