Bug 1336243 - Fix a regression from bug 1324617 in launching the browser post-install. r=agashlin
Bug 1324617 changed the name of the registry path that was being used to find
the new installation to launch, but forgot to update that function. This patch
switches to a more direct method of getting the right path to launch.

MozReview-Commit-ID: Hexhj9y4ixc

UltraBlame original commit: e6d624b8326b6a9152f30a203bb41c3c66ac0d52