Bug 1334093 - Test jobs should always extract the mach script and mozinfo.json from common.tests.zip. r=aki
Because  the mach script can be used to trigger a couple of different test jobs, it would be great
to extract it by default. The same applies to mozinfo.json, which contains build information to be
used in eg. the manifestparser to conditionally skip builds.

MozReview-Commit-ID: ESwi0EtRnO2

UltraBlame original commit: c0a4ffb1eb40cb21388ade2283f134f2d2edf8a6