Bug 1317774 - Add gfx code needed to use WebRender in gecko. r=gfx
The overall architecture here is that we add a new layers type, LAYERS_WR,
which can be used in place of client layers. The WebRenderLayerManager, in
the EndTransaction call, paints content into images and ships them over the
PWebRenderBridge to the compositor thread. The WebRenderBridgeParent code on
the compositor side talks to WebRender via the API in webrender.h.

MozReview-Commit-ID: JKLTLJWVXiN

UltraBlame original commit: b4e83ae3d878b1552bbc142d848151db100ceba6