Bug 1319521 - Properly tear down the LayerTreeState for content. r=gfx?
We were never actually removing the LayerTreeState entry for content layer trees.
In non-webrender codepaths, this happens when a PLayerTransaction belonging to
a CrossProcessCompositorBridge is torn down (see CrossProcessCompositorBridge::
DeallocPLayerTransactionParent). The equivalent of that in the webrender world
is in DeallocPWebRenderBridgeParent, so we should do the same.

UltraBlame original commit: 7aaf9790febbf438b8c745ac7e692885e391a827