Bug 1319237 - Lower timeout elapse evaluation count; r=me a=followup-to-prevent-back out
Although low, 150 evaluations appears to cause intermittents with on
slow try environments, such as Android emulators.

Since the purpose of this test is to check that the timeout elapses
properly, we are not reducing quality by testing for a lower number
of evaluations.

MozReview-Commit-ID: KqOTARg1AHP

UltraBlame original commit: 60b6ab14650a3818f2d11c640026780e28988735