Bug 1336431: P5. Re-add ConfigurationChanged API. r=JamesCheng
We need to inform the MediaDataDecoder that a new sample configuration has been detected.
To do so, and to ensure that samples are easily matched to a given configuration, we first drain the decoder prior feeding a new frame.

MozReview-Commit-ID: Hye251CF21i

UltraBlame original commit: ad29077e3c801ce3b673f5ff44cf216ce1133a90