Bug 1335281 - Use frame loader for same process browsers. r=mrbkap
This changes the `relatedBrowser` property which held a <xul:browser> to the
more explicit `sameProcessAsFrameLoader` which takes an nsIFrameLoader.

This clarifies the purpose of the property and also (by switching to the frame
loader) makes it easier to set in some contexts.

MozReview-Commit-ID: LnEvSP8zkto

UltraBlame original commit: a273874c22052691f8cec93a82d374ad0ec52454