Bug 1339924 - Remove unneeded [queue] lambda capture - r=dholbert
'queue' is not used inside that lambda, so we can drop it.

Another option would have been to simulate its use through `Unused <<`.
But I think that it's a useful test, *not* to capture the queue here, to verify
that it is correctly captured in the enclosing MozPromise::Then.

MozReview-Commit-ID: LUQGPJzDcTb

UltraBlame original commit: 9137f214b99408af568316c91305131b7c6d8c0e