Bug 1342215 - Fix pageError stub generation. r=jdescottes
We set a listener for "pageError" but we don't remove it.
There is a `once: true` property passed as an option for addListener,
but this isn't a standard addEventListener call it does not do anything.
Reverting to remove the listener fix the issue.

MozReview-Commit-ID: CrpS1QY5jNo

UltraBlame original commit: 647cbb760050200a9e0c810acf23541c98f0f5ca