Bug 1343446 NativeKey::GetFollowingCharMessage() should ignore found message if PeekMessage(PM_REMOVE) retrieves different char message but the found odd char message was odd r=m_kato
According to crash reports, we may find WM_CHAR whose wParam is 0 and scancode is 0xFF with a call of PeekMessage(PM_NOREMOVE) but we'll remove usual char message with a call of PeekMessage(PM_REMOVE).

In such case, we should ignore the found odd message and take the usual char message which was removed from the queue actually.

MozReview-Commit-ID: Gw8LvCXxul

UltraBlame original commit: c3563703a34d21d9442654503b15a97795868e42