Bug 1282306 - More correctly describe APZ touch support in about:support. r=botond
Instead of just reporting the value of the pref, also make sure that about:support
incorporates the runtime check of whether or not a touch device was discovered
in the widget code. The code to do this already exists in TouchEvent::PrefEnabled,
so we can just reuse that.

MozReview-Commit-ID: DN7FSlsDwD1

UltraBlame original commit: aa2376407a4d130e54e2b1949991d5c7de65706d