Bug 1337056 - Part 6: Disable nsPermissionManager::GetEnumerator in the content process, r=kitcambridge
This property allows enumerating over all permissions. As the content
process no longer has access to all permissions, this method cannot be
correctly implemented in the content process anymore.

Because of that, we now error with NS_ERROR_NOT_AVALIABLE when it is
accessed in the content process.

MozReview-Commit-ID: BLNeYYcZhIi

UltraBlame original commit: 33447647d4e869e656e015f47b416e87bbaa61c3