Bug 1316318 - Disable APZ scrolling in subframes if the root document has visible anonymous content. r=mstange
Various pieces of browser chrome, such as the accessible carets or new
find-in-page highlighter, overlay the content with anonymous content elements.
These overlays are positioned with respect to the root document. This means that
if the overlay really pertains to a scrollable subframe (e.g. carets for a text
selection inside an iframe), then scrolling the subframe would result in the
anonymous contents being "mispositioned" from the user's point of view. The
browser chrome code that creates the overlay is responsible for moving it in
response to user scrolling, so that it appears correct. However, the only
mechanism the browser chrome code has to detect user scrolling are the main-thread
events, which are asynchronous with respect to APZ.

Therefore, in order to improve the user experience, we need to disable APZ in
cases where this happens. We specifically isolate the case where the root content
document has visible anonymous content elements, and disable APZ on scrollable
subframe elements inside that document. Note that APZ remains enabled on the
root scroller itself, because we expect the browser chrome code to use absolute-
positioning on their anonymous contents, so that even with APZ scrolling they
will stay where they are supposed to.

MozReview-Commit-ID: 770edaGWMLi

UltraBlame original commit: c20e214d6486e9c22080e649f2a048efec47fc85