Bug 1316990 - part 2b - add initial set of approved licenses for Rust packages; r=gerv
This commit adds the initial whitelist of (mostly) SPDX license
identifiers and the initial whitelist of separately-specified license
files for those packages that require it.  For gamma-lut, the license
that the hash identifies is:

https://hg.mozilla.org/mozilla-central/file/47611a305c44/third_party/rust/gamma-lut/LICENSE

For deque, the license that the hash identifies is:

https://hg.mozilla.org/mozilla-central/file/47611a305c44/third_party/rust/deque/LICENSE-MIT

Integrating any new licenses into about:license is left as separate work
for the time being.

UltraBlame original commit: 5bcb7b097c66663ec9e7dd713f7d8a7533927bb3