Bug 1344932 - Separate out the ESLint spidermonkey preprocessor to avoid breaking './mach eslint --fix' on the whole tree.' r=jryans
This doesn't allow --fix to run on the whole tree in one go, but does allow it to be run on individual directories. This
is due to an ESLint issue, where once the processor is loaded, it applies to the rest of the files.

MozReview-Commit-ID: Fqg5yK22Dhu

UltraBlame original commit: 27ee69dbbf5a30cfee62bfd966a186e649842135