Bug 1348982 - Make 'mach clang-format' format the current revision instead of the tip. r=gps
Currently the mach clang-format command is formating '-r tip^', which is
everything in the tip revision, plus changes. Often this does what you want,
but if you're using changeset evolution or history rewriting this doesn't work
so well.  If you `hg up $rev` to some other revision with the intention of
clang-formating that revision, you don't get that revision formated when you
run clang-format. You get the lines touched by the changeset with the "tip"
bookmark formatted. We really want to have clang-format format '-r .^' which
means "the current revision in the working dir, plus any uncommited changes".

This results in the expected behaviour, and means clang-format can be used to
re-format commits which are work-in-progress and/or being evolved.


MozReview-Commit-ID: iv82HENxiA

UltraBlame original commit: 6270b8fc98abb8377d650dc344c524d2f5f3228f