Bug 1348666: Don't nuke cross-compartment wrappers for ScriptSourceObjects. r=shu
When destroying add-on compartments, we generally nuke all wrappers in or out
of the compartment. However, when cloning a script into a compartment, we
store its source object as a wrapper into the original compartment, and the JS
runtime expects that wrapper to remain valid.

This change simply exempts all objects of that type from nuking.

MozReview-Commit-ID: L70QqkbuX3h

UltraBlame original commit: 5bd473699d777438dae3c43892732f83ca3a7163