Bug 1346247 - Bug 1346247 - Avoid race conditions when SW are updated - part 1 - PServiceWorkerUpdater actor, r=bkelly
In this patch I introduce a new actor: PServiceWorkerUpdater managed by
PServiceWorkerManager and running on PBackground. Each time a SoftUpdate() or
Update() operation starts, a PServiceWorkerUpdater is used for having a unique
lock for the update of a particular scope+originAttributes.
The implementation happens on the PBackground thread of the parent process,
where all these PServiceWorkerUpdater actors are taken in an array and, only
the first one for each scope+originAttributes is activated.

UltraBlame original commit: 7474a804fd1f18f3397f28bdba910f23582b933d