Bug 1328868 - Part 5 - Add a Java-side listener watching the new pref and the Android system font scale. r=sebastian
The state of the switch added in Part 4 is stored in our Android-side shared preferences. For this to have any actual effect on rendering, we now add a class that is initialised when Gecko starts up and listens to changes of that particular pref.

When it is turned on, we enable font inflation and add another listener for the system font scale, which then forwards the current font scale as well as any changes to Gecko, so mobile mode pages can be scaled correspondingly as well.

When the setting is turned back off again, the system font scale listener is stopped again and the Gecko font size settings reverted back to their default values.

MozReview-Commit-ID: GyffpZTQQX8

UltraBlame original commit: 96ad7f172a08dce4e59f0e7c5cb90c8274b0bcd3