Bug 1343589 - Rename animation observer test files r=hiro
As part of this patch series I'd like to add an animation observer test.
However, the current arrangement of test files is quite confusing. We have:

  test_animation_observers.html - Tests that run async only
  test_observers_for_sync_api.html - Tests that can be run synchronously

It's not at all obvious that these are related or that one is related to the
other. In this patch we rename the files so that their relationship is more
obvious. In a subsequent patch we'll add comments to further clarify the
distinction.

MozReview-Commit-ID: CSYL8wruWdK

UltraBlame original commit: 5f321e0d488ae2c89c239f0f3b4ebfc49fde9f7c