Bug 1265818 followup: touch CLOBBER after converting nsILayoutHistoryState.h to .idl. rs=mccr8
Without a clobber, the build may generate untracked copies of the generated .h
file in your source directory. This probably happens because your objdir may be
left with a stray symlink for this header, with the stray symlink pointing at
the now-missing version of the header in your source directory.  And when build
generates a .h file (from the new IDL file) in your objdir, it lands on top of
that symlink, which effectively recreates a version of the header in your source
directory.

MozReview-Commit-ID: Ew0aP6PM1Ms

UltraBlame original commit: 4e53b8baa32e585d11c4f591eba3df63ec4d7736