bug 1351475 - allow the windows ProxyDestroyed() function to called with a proxy with no wrapper.
unfortunately because creating a DocAccessibleParent is not atomic one can be
created by AllocPDocAccessible(), but then RecvPDocAccessibleConstructor() can
fail.  If that happens we will call actor destroyed on the new
DocAccessibleParent, however because the constructor messaged failed it will
not have a wrapper.  That means this assert does not necessarily hold.

UltraBlame original commit: 7700cdd26f31fe55ab42359fdec3df310f2ac626