Bug 1350437 - Install Mercurial 4.1.1+; r=dustin
We've been running Mercurial 3.9 in automation for a while. Mercurial
4.1 is out. It has the usual performance improvements and bug fixes,
making it an attractive upgrade. But what really makes it enticing
is support for zstandard compression over the wire protocol. This will
reduce server-side CPU load and make transfers between 4.1+ servers
faster.

So, let's upgrade to Mercurial 4.1.1.

The produced packages are built from the current tip of the
stable Mercurial branch, not exactly 4.1.1. Specifically, they
correspond to revision ed5b25874d99. I did this because there is
a patch in the stable branch that drastically improves performance
on repos with many heads. When 4.1.2 is released in a few days, we
can upgrade to it and do away with the one-off.

MozReview-Commit-ID: 6BPhVheHQXI

UltraBlame original commit: 7175a2fbf5047e556f7978c91cfd16602c7a4160