Bug 1353208 - Add check for attribute namespace ID when we decide if it should animate as a CSS property or not; r=dholbert
This seems like an existing bug. If the content specifies
attributeType="yer:opacity", we should not try to animate it as a CSS property.
This patch adds a namespace check before we try to animate as a CSS property.

MozReview-Commit-ID: 1LpBa23ddqX

UltraBlame original commit: 95e21766cd72bdda8da9bf8e8bd43909da3be861