crashtest for bug 1274083 r=jesup
This needed some fine tuning to produce the crash, and so is a very specific
test.  It is still of value because this specific situation is not otherwise
tested, and it provides input to fuzzers for potential similar situations.

UltraBlame original commit: 2dc4150c15c88a48a335d613fc88193866ff86b7