Bug 1351535 - Part 3: Only explicitly style newly inserted content if the parent is not in a display:none subtree. r=bholley
If we append children to an element inside a display:none tree and we
decide not to do lazy frame construction, StyleNewChildren will assert.
It's also unnecessary, anyway, since we want to keep the descendants
of the display:none element free of ElementData.

MozReview-Commit-ID: 19tuzAzAIDp

UltraBlame original commit: 649120253cc1ad746a8bf9bf6950833383a866ec