Bug 1355009 - Force update tests to only allow a single update. r=ato
There was never a need to run a multiple-update step in the past, and as
we agreed a while ago it is not something we want to do in the future.
It means that watershed releases will have to be tested by issuing
multiple update tests.

MozReview-Commit-ID: 7cmK3gEOkv1

UltraBlame original commit: d2380e05da13742a99bc8502e5612de4338bec66