Bug 1355471 - Test timeout field before value's typing; r=whimboo
The error message returned when unmarshalling the timeout configuration
object with invalid input is misleading, because it checks the typing
of the value before the field name.

This patch changes Marionette to run the type assertion for the value
after each case in the switch statement has been evaluated, ensuring
that the field is valid before asserting its value.

It also adds a few unit tests to verify this behaviour.

Fixes: https://github.com/mozilla/geckodriver/issues/633
MozReview-Commit-ID: LVjTyUacD0s

UltraBlame original commit: 64fa27c7e7279b63ac5170838d27ca2e44992074