Bug 1314979 Support cross compilation with MINGW in windows.configure r=glandium
MINGW builds do not need any of the checks that are performed in the
windows.configure file. Nor do they the D3D compiler DLL that is
needed for ANGLE, so we can skip that entire section in
old-configure.in.

MozReview-Commit-ID: DqufbgGoGy4

UltraBlame original commit: c943802a5256597fb7d68b324b2a19830b39864d