Bug 1353629 - PBlob refactoring - part 0 - IPCBlob, r=smaug
This first patch introduces the basic idea of the refactoring. Instead of
having a PBlob protocol, here we have a simple IPC struct: IPCBlob.
When a Blob is sent to a different process, we move an IPCBlob struct with
type, size, some specific File properties, and an inputStream as IPCStream.

UltraBlame original commit: 1cdd50772eecfbceb7f19017919b9ddbaa2cf5ae