Bug 1351678: Handle resolved conflicts correctly, r=kmag
The elements of the SyncResultObject#conflicts field are {local,
remote} pairs.  However, the elements of the SyncResultObject#resolved
field are just the resolutions, and trying to access "local" and
"remote" fields of those resolutions causes a crash. Add a test that
demonstrates the error, and then fix it.

MozReview-Commit-ID: LF0NBw5VKBC

UltraBlame original commit: 5053fcd89b010d99dee13ed41a4af9730bf6438e