Bug 1359028 - added encodeURIComponent and fixed css class typo; r=pbro
The patch of bug 1353005 removed by mistake a leading dot in a css class, and
the `encodeURIComponent` calls (so the # in the color definition was considered
the hash part in the url).
Also, it appears autoland rejected the last commit from Bug 1353005, so this
patch includes those changes as well.

MozReview-Commit-ID: 2aVW3hYHhSr

UltraBlame original commit: 5046bec1e8ea53cf799129d1e703553e71d56476