Bug 1357754 - Implement the traversal functions of WebRenderScrollDataWrapper. r=botond
This mainly implements the GetLastChild and GetPreviousSibling functions
in the WebRenderScrollDataWrapper, and adds helper functions that are
needed to make that possible. This allows the WebRenderScrollDataWrapper
to simulate a full "LayerMetrics" tree using the information in the
underlying WebRenderScrollData object.

MozReview-Commit-ID: K82Ud2gAo8K

UltraBlame original commit: 1f032699c9c56e3364757db13f25a05f688932df