Bug 1362212 - Move canPlayType telemetry to an idle service observer off main thread. r=gerald
Currently we call HTMLMediaElement.canPlayType() in a JS function called
shortly after startup in order to collect telemetry as to how many of our users
don't have functioning decoders.

Unfortunately, HTMLMediaElement.canPlayType() checks whether we can play a
codec by instantiating a decoder, and this requires us to load the system
decoding libraries from disk. This requires disk I/O, which can cause jank. We
have some BHR reports showing that canPlayType can hang for > 8 seconds to back
this up.

So move the collection of this telemetry to an idle service observer, so that
we only collect this when the user is idle, and do it on a non-main thread so
it is less likely to cause jank.


MozReview-Commit-ID: HJQawmRxz

UltraBlame original commit: 5f263d6fb16db08d0d7efc73aaccf53240d0b461