Bug 1362811 - Don't require REFTEST prefix in unstructured reftest analyzer, r=dbaron
This prefix can't be produced by the standard structured logging
formatters because it doesn't know that the tests are
reftests. Therefore the reftest harness has a hack to add this prefix,
and the unstructured analyzer doesn't work with wpt reftests. This can
simply be solved by making the prefix optional, but looking for the
"INFO - " prefix that mozharness adds, when present, to identify the
start of the TEST- string.

MozReview-Commit-ID: GxWVWOSkAUZ

UltraBlame original commit: c528edf75b2e506246c21602505164032fd69ebd