Bug 1359759 - 3 - Faster debounce for grid highlighting; r=gl
Now that we are not re-rendering the grid outline all the time, it
seems safe to change the throttling mechanism of the cell highlighting.
Instead, we use debouncing now, so that rapid movements won't lead to
useless highlighting. The cell will get highlighted only if the user
settles on a cell in the outline.
Also, we can make the timeout much shorter so that only really rapid
movements are ignored, and the rest leads to cell highlighting, resulting
in a smoother UX.

MozReview-Commit-ID: 5WxAQRaLJkC

UltraBlame original commit: bdc6784cdfd4b751daf6855251eae481f9ba53da