Bug 1363982 - Clean up fixtures for session creation in wdspec, a=testonly
[Cherry-picked from wpt]

This introduces two fixtures for creating a multi-test session and a
longlived session. For performance it is important that we don't
ordinarily create a new session for each test, but it's also important
that specific tests are able to create new sessions with arbitary
capabilities.

MozReview-Commit-ID: DNZCgmkrwvn

UltraBlame original commit: d68c202a67091f65e127e3b25fe2ca4db78b58ef