Bug 1325112 - 1. Remove GeckoInterface.isOfficial; r=nalexander
There are two callers of isOfficial() in GeckoThread. For purging the
startup cache, the code to add the extra argument is moved to
GeckoApplication/GeckoApp/GeckoService. For logging the arguments, the
"debug" flag is used instead of the "official" flag.

MozReview-Commit-ID: 9atEWY1MVVt

UltraBlame original commit: b4549c97046b466e05826bffa53f35b175ec88ba