Bug 1365660: Part 3 - Disable transparency (where necessary) for popups which contain remote content. r=mstange,bas
Transparency is not handled correctly for composited popups on any platform,
but works with varying degrees of success on some platforms, for some popups.
Oddly, out of the three main desktop platforms, Linux seems to handle it the
best, so long as we render the popup as opaque, and let the platform
compositor handle the transparency.

MozReview-Commit-ID: E8NQlToUQq3

UltraBlame original commit: cb4b09616e235659ac4b6570c99f5a771d2d9206