Bug 989373, part 6 - Rename |obj| to |thisObj| in PrepareObjectForLocation. r=kmag
The name obj is not very descriptive, and a later patch will add
another object to this method, so rename it to thisObj. thisObj is the
object we set properties on for a particular .jsm. With global
sharing, it will be different than the global.

MozReview-Commit-ID: 9TPqdbXKYXO

UltraBlame original commit: 67231a14bdc916eb8a0b673e0eeb34ada188a3dc