Bug 1358023: (1/2) Refactor the dependency between libpng and freetype2; f=jfkthame, r=glandium
This refactor is for exposing necessary symbols of libpng for freetype2 in a
general way. Currently the necessary symbols of libpng for freetype2 are
exposed only on Android. And whether these symbols are exposed solely depends
on the target platform checking. A better way to decide whether or not to
expose these symbols would be checking the dependency of freetype2 directly.

UltraBlame original commit: 5ef43ba36c8c519ed91d5136389cfa89d92f05a2