Bug 1366907 - Remove BufferDecoder's GMPCrashHelper. r=karlt
We no longer support decoding non-EME audio via GMPs, so it's not possible
for a GMP to be used for decoding content that passes through a WebAudio
BufferDecoder. So we don't need to keep the GMPCrashHelper around in the
WebAudio code. If EME content is being used, there will be a crash helper on
the MediaKeys object anyway.

MozReview-Commit-ID: BThxnTANSTD

UltraBlame original commit: 53561331ff27b51b11165babfcf29f809084cee0