Bug 1366180 - Fix title and parent links in resource:// listings. r=valentin
For a long time, opening a resource:// url that leads to a file list has
used a title of the form "Index of jar:file://..." where the
jar:file://... url is the actual location the resource:// url has been
resolved to in the omni.ja.

That same url is used to derive a link to the parent directory. Because
of security context restrictions, the resource://... page can't open a
link to jar:file://... .

So we use the original resource:// url to derive the parent directory
link, and while here, also fix the title.

UltraBlame original commit: 056af0b6adda57f3f3084df80ea494bd70ed0e89