Bug 1357917 - Only hide doorhanger when navtoolbox autohides r=Gijs
On OSX we want to show doorhangers when in fullscreen, since OSX
fullscreen doesn't hide the nav toolbox. This makes that change,
and also adds flip="slide" to the panel so that the arrow adjusts
correctly. Unfortunately there still seems to be a bit of a
problem with this where the doorhanger adjusts its position when
entering fullscreen but then waits a little bit (not sure what
triggers it) before updating the anchor arrow. This is tracked by
Bug 1368094.

MozReview-Commit-ID: 3dRLwgMjxIb

UltraBlame original commit: 559651eb7bf87d2ad10f4cdd75cb7fde6ffaacb1