Bug 1363760 - Part 3 - In tests, install SpecialPowers and mochijar as non-temporary addons r=ahal
This means they will be copied to $PROFILE/extensions, which the sandbox allows
access to; if they are installed as temporary addons, loading frame scripts in
the content process tries to read from wherever they happen to be on disk. This
breaks running tests with a packaged build once we have full read-restrictions
for the content process sandbox.

MozReview-Commit-ID: 7ZiiM9FMXfG

UltraBlame original commit: 71e0ff5a6c76a811fb7fcc0684578727754dab17