Bug 1369108 - 2. Refresh Android camera list when necessary; r=jesup
Currently, if permission is first denied, the list of cameras is empty.
However, if permission is later granted, the list stays empty because we
never try to refresh the list. This patch causes us to refresh the list
when necessary.

MozReview-Commit-ID: 5eodPCWVyaP

UltraBlame original commit: 54c93d5eb3089ed12163a7308edc003cdb41cf7a