Bug 1298307 - Change the cue's startTime/endTime to avoid the videoframe was not created. r=alwu
Because the nsVideoFrame's creation is not synchronous to "append to dom tree", there is a small gap
that the TextTrackManager wants to render cue but no frame. Change the testcase to not hit the period.

MozReview-Commit-ID: 9xhMjRJnoDR

UltraBlame original commit: 028a9e5e1355fb2a099f32200c09fb2d9b7e8ec4