Bug 1368264 - Include geckodriver in test archive r=jgraham
When the geckodriver binary exists, include it in the common test package
alongside the other binaries under dist/bin.

This approach works even if the geckodriver binary is not available, as
it indeed is not on selected platforms.

MozReview-Commit-ID: 4aUSpdHJtYG

UltraBlame original commit: 3f7e348f0ff218a49f8730cc514ae015072d09ad