Bug 1357082 - Retry synthesizeMouseAtCenter until it is performed. r=xidorn
This patch is for the intermittent bug 1357082 which has a failure in the test 'test_pointerlock-api.html' and is more frequent when it is run on Ubuntu 16.04. This failure is resolved by retrying the mouse movement until it actually occurs.

MozReview-Commit-ID: 9XC4zmWj8Gn

UltraBlame original commit: f8d48ac82e7b06c803558ff5086882f2e122783a