Bug 1372721 - Remove run-on-projects from source-test tasks; r=dustin
The default value for run-on-projects is "all." Many of the tasks
changed in this commit were defining "run-on-projects" to
integration+release, which is essentially all (notably missing "try"). I
can't think of a good reason to not schedule these lint tasks on all
repos by default. So just do that.

MozReview-Commit-ID: EowUGNzC7K3

UltraBlame original commit: c01b9173e0b51cadbbbfb9d01fcd5c959de752cf