Bug 1353593 - Part 1.1: Fix usage of char16ptr_t in a graphics logging message. r=milan
This message intends to print out a string but was inadvertantly converting
the string pointer to a hex value. Adding an implicit conversion of char16ptr_t
to |wchar*| breaks this, so we just update the code to actually convert and
print the string instead.

MozReview-Commit-ID: 90luEnoysX3

UltraBlame original commit: ca5ccd1d096eb818d593c7fa9b9332e51a3dd262