Bug 1370296 - Ensure rust libraries are fully in the compile graph. r=gps
This was not strictly necessary because they would appear as
dependencies of other things in the compile graph, but it makes it
easier to iterate the compile graph for all the node.

Incidentally, this makes the rust libraries on par with rust programs.

UltraBlame original commit: 052b55f072ac01ac33e3eb7bf6cd289b3da562bf