Bug 1372668 - [mozlog] Limit the size of data logged from error message when suite_start called multiple times, r=jgraham
We've seen a couple of enormous log files (200MB+). This largely happened due to a bug in the test harness that resulted
in suite_start being called over and over again. In each of these instances, mozlog logged something like:
log.error("Suite start called multiple times: {}".format(data))

The problem is that 'data' contained every single test id in the suite, which was *a lot*. Dumping all test ids in that
error message is not useful for debugging. This patch limits the size of the 'suite_start' data in the error message to
100 characters.

MozReview-Commit-ID: GnPizNrZ2QJ

UltraBlame original commit: f23dc18a0316fd3d2abf18cab056aa3ad2ef70c7