Bug 1371465 - Move MSVS_VERSION to moz.configure and properly define for vs2017; r=glandium
Before, MSVS was set in old-configure and could only be unset or
"2015." We move the definition of the variable to moz.configure
and support defining its value as "2017" when VS2017 is being used.

As part of this, I discovered that GYP barfs with a "2017" value.
This is likely a limitation of the legacy version of GYP we have
vendored. Rather than go down the rabbit hole of upgrading GYP,
I added code to convert the value to "2015." This preserves existing
behavior and unblocks us from setting MSVS_VERSION properly. A
warning is emitted to remind us to remove this hack once GYP is
upgraded.

After this commit, we now generate native VS2017 solutions and
projects when building with VS2017.

MozReview-Commit-ID: BvNJX3F8qCn

UltraBlame original commit: 398d9d49677547b1b9f106370858c925de8c1511