Bug 1373564 - Force enable update staging mode to prevent perma failures for update tests. r=automatedtester
Bug 1370576 disabled staging of application updates, which is not yet
supported in our firefox-ui update tests. As result there is a hang and
the restart button is never clicked on.

This is just a wallpaper fix and full support might still have to
be integrated if wanted.

MozReview-Commit-ID: 8LF1RKxnMEb

UltraBlame original commit: 21b600e847b9d4211070f0d21feee519387f8a54