Bug 1365772 - Refactor gfx blocklist tests to fail on unsupported platforms. r=kats
They currently silently pass and early-exit under several conditions, which is really confusing if you're trying to actually run them properly. They should either be marked as skipping or failing to hint to the developer that something strange is happening. It appears that skip-if simply doesn't function correctly when applied to these tests, so fail-if is used. fail-if is also useful because it forces anyone introducing a new platform to evaluate how these tests should handle it.

MozReview-Commit-ID: CPYdL242bGc

UltraBlame original commit: f978d13e38c715cbbb49702ef2577f9e724b6d8a