Bug 1369581 - Requesting an optional permission that does not cause a prompt should succeed, r=aswan
Currently, if an extension requests one or more optional permissions,
and those permissions do not cause a prompt to be displayed,
the framework will reject that request and not grant any permissions.
This should be the opposite in that we should grant permission to
those optional permissions event though no prompt is displayed.

MozReview-Commit-ID: 6SeyFSv92Lo

UltraBlame original commit: 2cf23132f3b1aba5ed6cfcd2b9cdaabb5cccc0f5