Bug 265894 - Part 8. Implement -moz-use-shadow-tree-root pseudo class. r=heycam
The spec said:
The generated instance of a ‘symbol’ that is the direct referenced element of a
‘use’ element must always have a computed value of inline for the display
property. In other words, it must be rendered whenever the host ‘use’ element is
rendered.

We use -moz-use-shadow-tree-root pseudo class to achieve this requirement.

MozReview-Commit-ID: 1ReJjmjxq6G

UltraBlame original commit: abeaf8ef5a3b2167cb107b50cce476e7a93d9286