Bug 1373568 - Retrieve message data in MessageContainer instead of in ConsoleOutput; r=bgrins.
Prior to this patch, we were retrieving the message data in the ConsoleOutput
render, for each message. This means that for each message addition, we were
retrieving the data for each message in the store.
This doesn't impact performance if there are not many messages in the store, but
if the store is "full", we might end up calling the function to get those messages
a lot. And since the messages are in an Immutable.OrderedMap, it can be very costly.

Moving the retrieval of the message data in ConsoleOutput allow us to only call this
function when a message is added or updated, which allow us to save some time on
the rendering path.

MozReview-Commit-ID: S4NSSW5Mvw

UltraBlame original commit: 74d375a6430c8c0142c7fadd659efd84707e4455