Bug 1375188 - Don't import TextEncoder from osfile.jsm in CoverageUtils.jsm. r=Gijs
This import is unnecessary, because the WebIDL interface for
TextEncoder has Exposed=System, so it is available on the JSM
global. I can revert some of the changes from bug 1348103 at the same
time.

I need to remove this import, because it will break with shared JSM
globals.

MozReview-Commit-ID: BKmaH8iTFnI

UltraBlame original commit: 6399592c5acae7d1445822e9de92947e1ba27fba