Bug 1371733 - Add tests for custom assertion errors; r=whimboo
The last commit discovered a bug that one assertion's error message was
not being propagated.  This adds extensive tests for this, for all the
remaining assertions to ensure this does not regress in the future.

MozReview-Commit-ID: EK91mUPeweG

UltraBlame original commit: fef1ad177d24a4071a71d4c63ae8b09349f1f7ce