Bug 1223277 - Return immediately when click command closes tab or window. r=ato
In some cases the click command can trigger the closing of the
currently selected tab or window. To not cause a hang when waiting
for a response from the removed framescript, the tab and window
closing events have to be observed. Also the command has to return
immediately.

MozReview-Commit-ID: 9WeXryrKEJr

UltraBlame original commit: 85c86dc8612239b3c099384e2a06fe832fdbe6cc