Bug 1354159 - Part 3 - Add a Bookmarks button to the Library panel that shows a subview with a list of most recent bookmarks. r=Gijs,mak
This patch changes the Recent Bookmarks flow in browser-places.js to use the new
PlacesPanelview to provide a rich view that support drag and drop and context
actions.
A new button 'Search Bookmarks' was added, which prefills the urlbar with '* ',
which is a shortcut for searching bookmarks using the Awesomebar.

MozReview-Commit-ID: 1XlO8HMKEJs

UltraBlame original commit: 28e8a3011a4cd3979929af8e7c52a1587746203b