Bug 1379482 - Upgrade mozrunner to 0.4.1; r=jgraham
Upgrades the mozrunner crate dependency by pinning it to 0.4.1.  With `./mach vendor rust' I could not find any other way to force a new version to be downloaded because cargo, as far as I understand, does not yet support using crates.io as a fallback.

This patch also pins the mozprofile and mozversion dependencies exactly, although this is strictly not needed yet because they we have the latest versions vendored in-tree.

MozReview-Commit-ID: 8emDKbiYd0S

UltraBlame original commit: 24577b96157af17860707c6637c8a848340c7777