Bug 1367940 - Ensure sccache inherits the make jobserver file descriptors. r=gps
Rustc 1.20 gained support for make jobserver. But when rustc is invoked
through sccache, sccache needs to have inherited the make jobserver file
descriptor to pass it down to rustc. But what will be running sccache
first during the build is not really deterministic. It might as well end
up being a script invoking the preprocessor that ends up doing it. And
we can't add the necessary + prefix to all commands that potentially can
end up invoking sccache.

So we manually start the sccache server at the beginning of the build,
through preflight.

Note this works on automation, but not for local builds, but few people
are using sccache locally for rust (yet). Let's figure that out in a
followup.

UltraBlame original commit: d56a606712f1ae1086f6dc94623dea03726a5d9b