Bug 1368434 - Fix race condition for slow frame script registrations. r=automatedtester
In some cases the registration of the frame script takes longer and as
result the page load events for DOMContentLoaded and pageshow are
missed. With that the current command waits forever and causes a page
load timeout error after 300s.

By checking the readyState of the document before the listeners are
getting attached, we can ensure to return immediately if the document
has already been finished loading.

MozReview-Commit-ID: 17f6jVz7sZZ

UltraBlame original commit: 4f240a4f635b71a641d45cd2c4d891600954edf3