Bug 1377216 - Discover version control info in configure; r=glandium
For reasons unknown to me, Windows CI is periodically failing to find
the Mercurial binary.

In addition, we've also reimplemented various VCS logic throughout
the build system. There is room to cut down on code complexity by
e.g. recording VCS info in configure instead of determining it
at run-time.

Also, for forensic purposes it is sometimes desirable to know which
VCS tool is in use by a build and which version of that tool is being
used.

This commit adds VCS type detection, binary searching, and version
resolution to configure.

substs now contains VCS_CHECKOUT_TYPE, HG, and GIT, which can be
consulted by downstream consumers.

If the Mercurial or Git versions could not be resolved, all variables
are not set. Otherwise, VCS_CHECKOUT_TYPE and one of HG or GIT is set.

If MOZ_AUTOMATION is set, we require that the VCS info be resolved.
This helps prevents weirdness in automation due to unexpected
environment configuration.

MozReview-Commit-ID: AMLy0Hfx5rD

UltraBlame original commit: ce73e01ddacb18603453c356e336084da3a61137