Bug 1377216 - Search for hg.exe before hg; r=glandium
Mercurial installs on Windows often provide both a native hg.exe
and a hg shell script. This allows them to work from both Windows
and UNIX-like shells.

While the which Python package searches for .exe and other common
executable extensions on Windows automatically, it doesn't prefer
these results to an extension-less value. This meant that "hg"
was resolved to the shell script. When we went to run it, it
failed because it isn't a valid Win32 application.

We work around the problem by preferring "hg.exe" over "hg"
for the binary name.

Ideally this would be platform specific. But the harm should be
minimal.

UltraBlame original commit: 344d6f11ef151e9bf6dbe5d5d4397f96df60ebe0