Bug 1355922 - Save to Pocket should have an associated animation. r=Gijs
The animations on the pocket-button and library-button are disabled if they are not in the nav-bar due to issues with getting overflow:hidden and position:absolute to work the same way outside of #nav-bar. This is on file as bug 1382894.

MozReview-Commit-ID: Lq20GuocMkg

UltraBlame original commit: 8c927b51ca411f35225282ece3d2271a89b8528d