Bug 1382862 - Remove GeckoViewFragment; r=droeh
GeckoViewFragment was more of a hack for GeckoApp-based custom tabs, and
if we want to provide a fragment implementation for GeckoView, I think
it'll look a lot different than today's GeckoViewFragment. So for now, I
think it's best to revert the changes that added GeckoViewFragment, and
that also simplifies GeckoView lifecycle management.

MozReview-Commit-ID: 5WzIZ8TyBDV

UltraBlame original commit: cb3966784befbe11354b79598926681cd4088072