Bug 1384181 - Ensure that the WebRenderScrollData has a single root layer node. r=jrmuizel
The semantics of the WebRenderScrollData structure is that the per-layer
structures form a tree with a single root node. When we build the data
structure from the display list, we are generating (for now) a flat
list. Therefore we need to synthesize a root node in order to make stuff
work as intended.

MozReview-Commit-ID: IDXyziBO7pk

UltraBlame original commit: b298cd21c0f1850caabd581b3d34b003d05384b8