Bug 1362786 - (1) Increase subprocess BUFFER_SIZE r=kmag
While 4KiB is a decently sized buffer, bumping it to 32 on my
machine reduced the time for getSymbolsFromNM from 12s to 8s. If
32 is too large for the typical use of this, then we can instead
just parameterize this so that we can specify as large a buffer
as we want for our use case.

MozReview-Commit-ID: 3wGpHzQ9uZ5

UltraBlame original commit: 242a12eae20e469b3d8e6cfdc0ae34f9a1752c5d