Bug 1362786 - (2) Run NMParser in worker r=kmag
Pulls out the NMParser work (parsing nm results and turning
them into an ArrayBuffer'd map of addresses to symbols) into
a worker.

For OSX we will still need to do some work to run c++filt in the
background, but this gets us most of the way there. Without a
Subprocess.jsm usable from a worker, we'll have to bounce data
back to the main thread in order to bounce it to the c++filt
worker.

MozReview-Commit-ID: LZi7J1qGpmh

UltraBlame original commit: 71b419e1788dd017f3f3fe6bf76ae04115698a66