Bug 1384396 - Add a depends_all utility function; r=nalexander
Often you only want to evaluate a function if all its dependencies
are true. Expressing this in a "when" can be difficult. So let's
add a convenience decorator for it.

The existing code for depends_if() was refactored to take an
evaluation function as its first argument. This prevents some
duplicate code and turns depends_if() and depends_all() into
one-liners.

MozReview-Commit-ID: Jbugvf0lioM

UltraBlame original commit: 0c2bc12f4ebe44428385745266d2fd158e0c3382