Bug 1380709 - Add mochitests and mocha tests to ensure entries are loaded as expected. r=bgrins
This adds several tests to make sure we can expand Sets and Maps entries, and that we manage
those properties in the store as expected.
We also add a stub for a Set object and take this as an opportunity to fix the stub name.

MozReview-Commit-ID: ChhIHKMPINF

UltraBlame original commit: 44cfd92f709cd8888549565a917d4cea6ee5d379