Bug 1386620 - dt-addon-prefs: move jsonview enabled pref outside of devtools addon;r=ochameau
DevTools preferences are loaded dynamically by calling DevtoolsPreferences.loadPrefs().
We can not preload them when the addon starts, otherwise this will slow down the startup
of Firefox.

But jsonview's converter-observer needs to check preferences to check if jsonview is
enabled very early. Moving devtools.jsonview.enabled to a separate preferences file
that is still processed by firefox fixes the issue.

The downside is that this pref will keep following m-c's release cycle and not the addon's.
But it is so generic it should not be a big issue.


MozReview-Commit-ID: HrD5IVe54Ks

UltraBlame original commit: e4c9887480c94272de57a3625bc738bcc865623d