Bug 1386052: Correct PageMetadata.HISTORY_GUID foreign key. r=Grisha
This change is fixing an issue unrelated to this bug.

It was using Visits.HISTORY_GUID, which is not in the table and thus doesn't
make sense for the first clause of the foreign key constraint. Visits & its
replacement have the same value so the behavior should not change.

MozReview-Commit-ID: 8EbJh586DN3

UltraBlame original commit: 91ac0a2b022ceff57b7a1666ab595d387bc9d2ce