Bug 1385699. P5 - fix AudioEndTime() and VideoEndTime(). r=cpearce
AudioEndTime() returns 0 if the media sink is not started. This is wrong
because the audio/video end time should be at least the current position
to ensure playback won't go backwards.

http://searchfox.org/mozilla-central/rev/bbc1c59e460a27b20929b56489e2e55438de81fa/dom/media/MediaDecoderStateMachine.cpp#2009-2012
This fix the case where we wrongly set playback position to 0 after seeking
to the end of resource.

MozReview-Commit-ID: 2VnTcqyIvoJ

UltraBlame original commit: 01a413050ab712732b9078528d78b3723bfa1e81