Bug 1387532 - Enforce that we don't schedule multiple chunks for the same test without passing buildbot the required chunking information. r=catlee
The issue here is that we had marked 'chunked' as false, but were still trying to use 8 chunks for reftest. Because of this we were also sending the unchunked buildbot name to BBB for each chunk we actually tried to run, on e10s and other variants.

MozReview-Commit-ID: Dc5npIq5sxr

UltraBlame original commit: 710c01ee0566f211d89fce28e50f0ce3bf051a2c