Bug 1381876 - Ensure Maximize Window is synchronous. r=automatedtester
Because the sizemodechange event is not synchronous on all platforms,
we additionally need to await the window's dimensions to change.

However, if the window manager does not have a notion of a maximised
state, this operation too could time out.  However, it is believed
that the additional wait will allow for the window to transition to the
required state in this scenario.

MozReview-Commit-ID: KeHJMKSJfjQ

UltraBlame original commit: ec8c6742603a676e21a86194cde43a594fbc2773