Bug 1356273: Part 1 - Set document readyState to uninitialized during document.open() r=smaug
Calling SetNewDocument() immediately triggers content-document-global-created
observers for the new window global. In the case of document.open() calls,
though, the document itself is in an intermediate state at that point, and
does not have the correct URL or DOM contents for the new global. Setting the
ready state to uninitialized lets listeners determine when that's the case.

MozReview-Commit-ID: FAkyvOnkFR2

UltraBlame original commit: c539ce828bd1b7cd4afbfea74c4819baa8e693f3