Bug 1328830 - [manifestparser] Check line continuation before looking for next key, r=jmaher
Currently manifestparser will only look for line continuations *after* looking for a key. This means
that line continuations cannot contain key separators. For example, this:

    [test]
    foo=
      bar=baz

gets treated as:

    {'name': 'test', 'foo': '', 'bar': 'baz'}

Here, bar=baz will be treated as a new key/value pair despite the indentation. This patch switches
the order around, so we look for a continuation first. Now, it is only treated as a continuation if
the indent is greater than the indent of the preceding key.

So this manifest:

    [test]
    foo=bar
      baz=fleem

is a continuation and results in:

    {'name': 'test', 'foo': 'bar\nbaz=fleem'}

But this manifest:

    [test]
      foo=bar
      baz=fleem

is not a continuation, and yields:

    {'name': 'test', 'foo': 'bar', 'baz': 'fleem'}

MozReview-Commit-ID: FAMP5TUIo9q

UltraBlame original commit: 8d64da85644e03c63a3cf9aa62b33af1975bd35a