Bug 1380081 - Part 10: Stop running BHR on beta, r=froydnj
These changes are going to increase the amount of data which we collect from BHR
a lot. It would be dangerous to run it on beta, especially considering how soon
the next merge is.

This should turn it off for 100% of beta users if I understand the logic
correctly.

MozReview-Commit-ID: 3HyEKWdXaqU

UltraBlame original commit: d3910f43605c9c130313d8e1d02063eb7b5ff760