Bug 1390264 - Only set up the pocket-animatable-box when animation should begin. r=Gijs
This wasn't noticed in normal windows because the navbar gets position:relative on windows that are showing the TabsToolbar, but in popup windows the navbar isn't relpos and as such some of the machinery that was being set up when not needed was visible. This machinery is now not set up until an animation is requested.

MozReview-Commit-ID: JUqKVYTxE8y

UltraBlame original commit: 1770d65a34281a5a8ee16b2febc9b017a5d1ed6b