Bug 1380512 - Fix intermittent on browser_webconsole_check_stubs_console_api; r=Honza.
This hopefully fix the frequent intermittent on this test.
In the cleanPacket function, we are overriding the actor
of arguments. Thus, some server-calls were using an unknown
actorId, which resulted in errors.
We now create a new object for the argument for the stub so
we don't override anything.

MozReview-Commit-ID: EcvVUqQS3Dt

UltraBlame original commit: 6da0a758637b82e939e46be99223e3f6d54df091