Bug 1387403 - Force quitting the browser if no session can be established. r=jgraham
In case the NewSession command fails due to errors returned
by Marionette, the browser has to be killed. This is necessary
because DeleteSession always requires an existent session, and
would fail in closing the browser. So the process would continue
to be alive.

MozReview-Commit-ID: 1llX4lPNYjN

UltraBlame original commit: 832139e8b9c630448bc677de88536ae2c75b82d2