Bug 1389435 - Explicitly use GCC 4.9 for hazard builds. r=me
The GCC plugin used for hazard builds was built for GCC 4.9. Until a new
plugin is built by a toolchain job for the "default" version of GCC we
build Firefox with, the builds using the plugin need to stay with the
version the plugin was built for.

UltraBlame original commit: 7de11f6fe7cf48221750ef23d06c05bd4e845f87