Bug 1358771 - Don't set the resolution on the main-thread from an APZ repaint request that was triggered by a main-thread update. r=tnikkel
Conceptually, a repaint request that arrives in APZCCallbackHelper with
!GetScrollOffsetUpdated() is a repaint request that the APZ is sending in
response to a main-thread update it received. The purpose of the repaint
request is mainly to update the displayport margins; the scroll position
and resolution should not have changed. Therefore actively trying to set
the scroll position or resolution as a result of these repaint requests
should at best be a no-op but often triggers unnecessary clobbering of things
that are in-flight, or extra reflows. This can result in buggy behaviour.
Bug 1286179 avoided doing this for the scroll position, and this patch does
the equivalent thing for zooming.

MozReview-Commit-ID: 91RbRP0ZHBe

UltraBlame original commit: 68d3c804b45a115f89c252a4e10b28feb713d9ba