Bug 1394391 - [tryselect] Split hg implementation of files_changed to a list, r=jhford
This fixes a regression from bug 1384593. I was going to add a test for this, but it would be
a better use of time to fix bug 1185599 and add a test there. We already have vcs tests in other
parts of the tree so consolidating them into a single module is likely easier than standing up
another one-off vcs tester.

MozReview-Commit-ID: E51Tb1qC9Wb

UltraBlame original commit: 3942e9db3da304e63113ab3695a4d3348b375412