Bug 1394662.  Make ::-moz-table-column a non-inheriting anonymous box.  r=heycam
In our UA sheet the only style we apply to ::-moz-table-column is "display: table-column".

Per spec, the only styles that apply to table columns are border props,
background props, 'width', and 'visibility'.

The only one of those that inherits is 'visibility'.  And the only relevant
value per spec is "collapse".  But an anonymous column can only be "visibility:
collapse" in Gecko right now if its colgroup is, and then the colgroup will get
collapsed away as a whole.  So it doesn't matter whether the column inherits the
visibility style.

In practice, we do something special for "hidden" on a table-column as well,
during display list building, which affects hit-testing.  Per
https://github.com/w3c/csswg-drafts/issues/1763 it's not clear that we should be
doing this at all, but for now we just keep our existing behavior and ensure
that the parent colgroup's visibility style is taken into account for anonymous
columns when building display lists.

UltraBlame original commit: 1e89ea2871e7f3a2d3158ea0e660199bd282efce