Bug 1395701 part 2.  Use a linked list, not a hashtable, for registering selection ranges on a node, so the registration will be faster.  r=ehsan
Inserting/removing things into a doubly-linked list is much faster than doing
the same with a hashtable.  Selection ranges register themselves on their common
ancestor, but all we do with that in non-debug code is iterate all the ranges
registered.  A doubly-linked list works fine for that.

This adds three words to every range for the LinkedListItem members, but that
should be OK.

UltraBlame original commit: dd99274926b87dee19d65e05d92247c6e76da5f6