Bug 1396292 - Part 2 - Explicitly run permissions callback on background thread where applicable. r=sebastian
Currently, this only concerns setImageAs, where not running on the UI thread after a permissions prompt triggers a network-related strict mode violation and interestingly enough results in a crash (as opposed to a logcat complaint) on release builds.

There's also a permissions check in the UpdateService which runs on a (but not *the*) background thread, but since we don't prompt in that case there's no action necessary.

MozReview-Commit-ID: KKxW96AyDWH

UltraBlame original commit: 1320c22e21796cb54f02d1712e28b78066fad6c0