Bug 1397460 - Don't throw for a failure in cpows_child.js. r=billm
When ok() is passed false, we send a message to the parent, which will
cause the test to fail. Throwing in this helper in the child just
makes the test hang for a while.

MozReview-Commit-ID: 4gwBACPYfDO

UltraBlame original commit: a5aa81a626da9e7cf14f6600ee8ac0ed9b6cee1f