Bug 1397232 - Use the hidden window to report media telemetry. r=florian
A normal window might not exist when this is called, for example on Mac where it's possible to close all windows but not the application. This has been true since bug 1362212 moved it to an idle observer.

MozReview-Commit-ID: EcL941XKTeP

UltraBlame original commit: 4f3ba0330ce0f3eb60c909e97581a5c6146ce7ac