Bug 1397088 - Adjust autofill records list height and allow country warning message wrapping. r=lchang
This patch fixes a light height problem in records to align with the visual spec. Another issue is that on Linux the country warning message in edit address dialog is pushed to the next line, where it should be wrapped.

MozReview-Commit-ID: FZiBVKWLeMU

UltraBlame original commit: f2e2ecd961cf5e242f301327308234d562af6ed3