Bug 1397717 - Using GenericPrinter for DEBUG-only C++ dump() APIs;r=nbp
We have a host of DEBUG-only dump()-style APIs that output either to
stderr or to a FILE*. Unfortunately, this means that we cannot use
these dumps e.g. for unit tests. This patch ports most of the dump()
APIs to use GenericPrinter, which is more flexible.

Some parts of the code have not been ported, in particular
TypeInference, which still uses stderr.

MozReview-Commit-ID: A5WGOPyIPTa

UltraBlame original commit: 95bb616125726c8bd4c300a1eb182fc8268c7141