Bug 1367765 - Implement scrollbar touch-dragging in APZ. r=rhunt
The general implementation strategy is to detect whether a touch
block is a scrollbar drag, and if so, convert the touch events
to mouse events before sending them to InputQueue. This is done
so that code for handling mouse-drags in InputQueue and APZC
can be re-used.

Content will receive the original touch events, not the synthesized
mouse events. This is done for two reasons:

  1) We don't really have a mechanism for APZ to send content
     synthesized events.

  2) Content already deals with touch events that drag a scroll
     thumb correctly.

MozReview-Commit-ID: 7cksE5FMGoS

UltraBlame original commit: e11695ed7958c642978b5d152f3ac196c8e9d431