Bug 1398974: Follow-up: Disconnect StreamFilters when closing extension context. r=me
If we don't do this explicitly, the channel is automatically disconnected when
it's GCed. However, if we start shutdown while a stream is being processed,
the stream may not be GCed before we shut down the parent process's message
loop. In that case, we get a shutdown crash because the StreamFilterParent's
data channel is still open when we shut down its message loop.

Explicitly disconnecting the StreamFilter when the context is closed prevents
this, since app shutdown is automatically blocked on extension shutdown, and
extension shutdown explicitly closes all extant contexts.

MozReview-Commit-ID: 5JPrSUooq1j

UltraBlame original commit: 4f6d9ea5a49e003309f5f92ff155accb91c2f474