Bug 1395973 - Remove index cache from nsContentIterator. r=masayuki
nsContentIterator used to maintain a stack of indices so that when it
finished iterating through a subtree it would know the position of the
next node. Maintaining this stack is expensive and unnecessary since we
have fast getters for next and previous siblings.

UltraBlame original commit: bc3d643c49739f2576a5f94eb53a13d3d3b4ebaf