Bug 1400072: Cache tilesSize in swapCursor. r=liuche
In this sense, it acts as a refresh function, binding all of our UI state at
the same time, helping to prevent bugs where the UI gets out of sync.

For example, this fixes this bug because now the tileSize gets updated when we
swap the cursor, not just when we the page is initially constructed.

MozReview-Commit-ID: 7V2gFyiOJ1R

UltraBlame original commit: 45f33d2b572218950d819c0341774d45a40c7bb5