Bug 1400289 - Add geckodriver release instructions. r=automatedtester
This is a liberal port of the old geckodriver release instructions
that are on the GitHub wiki:

	https://github.com/mozilla/geckodriver/wiki/Releasing

Because the project's canonical home has moved to mozilla-central,
it is better to keep all documentation in one place.  Documentation
close to the source code is the best.

I have improved the release instructions in a number of ways, in
particular by describing how to make a release to GitHub/Travis from
mozilla-central.

DONTBUILD

MozReview-Commit-ID: 6f87DLDM59E

UltraBlame original commit: 760c49597860adad4acad6b55829dc158d464480