Bug 1401733: Make StreamOverridablePageIconLayout caches static. r=liuche
Unfortunately, I can't reproduce so I'm unable to verify this fix works for
sure. However, I verified the behavior of the StreamOverridablePageIconLayout
remained the same after these changes.

As per my bug comments, using the new version of onCreateView may fix this
problem and is "more correct" than this fix but it's speculative and it's too
dangerous to make speculative fixes this close to the 57 merge. Instead, we do
this surefire fix and lose a little correctness.

The onCreateView fix moved to bug 1401779.

MozReview-Commit-ID: DSq9jEgz6gL

UltraBlame original commit: 7bb527a1748fad509712d02775c2637c2e3172aa