Bug 1394284 - add fallback discrete procedure for transform animation. r=boris,hiro
According to the spec, if one of the matrices for transform interpolation is
non-invertible, the used animation function must fall-back to a discrete
animation. However, in the current implementation, we always use an identity
matrix as a fallback for the non-invertible matrix.

Decompose2DMatrix and Decompose3DMatrix both return a boolean, but we just
never use it. So, in this patch, we use the returned boolean from the matrix
decomposition as a condition, and do the fallback discrete procedure for the
non-invertible matrices case.

MozReview-Commit-ID: E7i1a1MJOXN

UltraBlame original commit: 9044841344c9af963ed30293d8c5ecea0ee23845