Bug 1401613 - Don't require test type in wpt expectation manifest, r=maja_zf
The test type is already in the MANIFEST.json file and requiring it in
the expectation data doesn't make much sense. It isn't used execpt
during updates and so people often forget to add it. Therefore it
makes a lot of sense to just use the data from the main manifest.

MozReview-Commit-ID: HyOoN6T28qc

UltraBlame original commit: 9192c6646cec6809c0afcbed592457b182471814