Bug 1402728 - Use RefPtr<TrackBuffersManager> instead of raw pointers - r=jya
All in MediaSourceDemuxer:
- GetManager returns a RefPtr<TBM>. Probably unneeded as the only call stores
the pointer in a RefPtr, but I'm adding it for peace of mind, and it's very
low cost.
- AttachSourceBuffer takes a RefPtr& and explicitly stores the given TBM in a
RefPtr in the dispatched task, then DoAttachSourceBuffer takes a
RefPtr<TBM>&&. This should not be needed, as it should already be done
implicitly. But now it's more obvious, and a bit more optimized thanks to move
semantics.
- DetachSourceBuffer takes a RefPtr& and explicitly stores the given TBM in a
RefPtr in the dispatched task, then DoDetachSourceBuffer takes a
RefPtr<TBM>&&; And changed the loop to use RemoveElementsBy. Same as above,
probably not needed, but it's now obvious that we keep a RefPtr.

MozReview-Commit-ID: 9aOZtV7uS1P

UltraBlame original commit: 05b222ac8a540186f26caf7cfe514956399ecdfd