bug 1401796 - fix HSTS preload script to keep preexisting hosts if there was a connection error r=jcj DONTBUILD NPOTB
Bug 1255425 changed an 'of' to an 'in', which caused the code that would keep
preexisting entries on the preload list if there was a connection error to loop
over the indices in the array and not the values themselves. Thanks, JavaScript.

MozReview-Commit-ID: DvVWhpImp8n

UltraBlame original commit: 218757467ee8b8feec1fcf3b028d5397bb748bac