Bug 1296531 - Add MediaSegment::IsNull. r=jesup
This allows us to see if all chunks in a MediaSegment are null.
When this is true for the MediaSegment passed to NotifyQueuedChanges, we can
assume that the track's input is blocked for the given time.

MozReview-Commit-ID: zf7V7aiohg

UltraBlame original commit: 5cba097bf1eda93c6e7e0b1d7e57e9362377c7e6