Bug 1403041 - Don't re-enter libcubeb from the audio thread. r=kinetik
This delays initializing the channel count until actually initializing the
cubeb stream for an AudioCallbackDriver.

The constructor of an AudioCallbackDriver is usually called from the previous
driver's thread (be it a system thread or the audio callback, the exception
being when it's the first time a GraphDriver is being created), while the
AudioCallbackDriver::Init method is always called from a special thread that is
specially created to run potentially long and blocking operations, such as
creating an audio stream.

MozReview-Commit-ID: GFx85N8PlKX

UltraBlame original commit: 65ff7c0adadcb1236ad5ec604bdd8a6cb65fd0aa