Bug 1402170 - Use for...of instead of forEach for several tests in processing-a-keyframes-argument.html; r=hiro
for...of is generally preferred over forEach since it is a little easier to read
and allows using 'break' and 'continue'. Furthermore it is supported in all
major browsers. (It also makes wrapping one of the long lines in this file
easier.)

MozReview-Commit-ID: 1BuoW0QSxaG

UltraBlame original commit: 5951672afb9de5259431f7ee033e7f8ec7aff0a3