Bug 1405408 - Pre: Handle query string and fragments in URL parameters. r=dustin
This was simply oversight before.  I ran into this using the
taskcluster-proxy /bewit interface, which returns a URL of the form
https://domain.net/short/path/to.file?bewit="several thousand
characters", which leads to an IOError due to the long path.  Let's
assume that such query strings and fragments are transient; we should
drop these parts of the fetched URLs when writing to disk.

MozReview-Commit-ID: FMJHMp7a3rA

UltraBlame original commit: 831c287a4ae882823ec7411a00b05e88e69ebda2