Bug 1408638: Ensure accessible isn't defunct in Windows RootAccessibleWrap::accNavigate. r=MarcoZ
Bug 1407475 added support for accNavigate(NAVRELATION_EMBEDS) for e10s.
However, it's possible for a client to call this on the root accessible for a window which was since closed.
Therefore, we must check whether the root accessible is defunct before trying to retrieve the tab document.

MozReview-Commit-ID: 9iR6Kvzu5Mb

UltraBlame original commit: e36b5e90262d89363d3ad0fa78f5beb94706a532