Bug 1408795 - Have Eclipse CDT treat Objective-C files as C++. r=botond
Without this setting, eclipse will refuse to open Objective-C files (it will
defer to an external editor).  Adding *.mm to the file types that are treated
as C++ allows Eclipse to open them, and to provide some code assistance for
the bits of the files that it can understand.

MozReview-Commit-ID: ASeXesWxY4g

UltraBlame original commit: bf4442765b3e41a11d8fcff74ac7067e5d18069a