Bug 1409057 - Use new tooltool download for android sdk in emulator unit tests; r=jmaher
This version of the sdk download is based on the previous one but includes
a new version - 1.0.39 - of adb and eliminates some files not used by the
emulator (smaller, faster download). By including adb in the sdk download used
for the emulator, the tools sdk download can be eliminated.

UltraBlame original commit: 137fa515a573967392d02ca445bfcfaed3415f95