Bug 1372075 - Make android-* Gradle test suites tier 1. r=gbrown
This adds a new `tc-A` Tree Herder group, similar to the `A` Autophone
(or Android) group.  (I don't want to include a `g` for Gradle because
eventually there will be nothing that is _not_ Gradle.)

Per https://bugzilla.mozilla.org/show_bug.cgi?id=1371445#c31, the
sheriffs are satisfied with the test output.

As to the rest of
https://bugzilla.mozilla.org/show_bug.cgi?id=1372075#c0, the
documentation is in place, and |mach try fuzzy| has eclipsed
trychooser, so I think we should not update trychooser.

MozReview-Commit-ID: 2OWDEmGCd11

UltraBlame original commit: 9482fd0038bbd5431a7fe8352f45dfd3b5b79b7e