Bug 1388478 - Use job-defaults to simplify 'test' configs where appropriate, r=jmaher
This is mostly a refactor, moving shared defaults to the top of the file where
appropriate.

The only change this makes, is modifying a couple of the macosx talos tasks to
use the osx mozharness config. They were previously using the linux config
which was likely a bug.

MozReview-Commit-ID: 2c0muPZrHZ

UltraBlame original commit: 0316704b960c8cb071058dbe8ecc6b0322c9f994