bug 1407368 - rename Breakpad's src/common/memory.h. r=gsvelto
memory.h conflicts with a system header, so we have workarounds to
change include paths to work around this.

This is mostly a cherry-pick of this upstream commit:
https://chromium.googlesource.com/breakpad/breakpad/+/8bb3d55af734fd469213e0b88f5b5ad7ffec0abf

..but the patch was applied separately to toolkit/crashreporter/google-breakpad
and toolkit/crashreporter/breakpad-client since we've forked the latter,
and there's also one other fixup of a source file included.

MozReview-Commit-ID: HH92HZG7y9n

UltraBlame original commit: e53f2b4f821d8792793bfe0527287244de50edcd