Bug 1400256 - Add web element abstractions. r=whimboo
This patch introduces a series of web element abstraction types for
representing web element references.

Adds a series of new types for representing web element references
in Marionette: ChromeWebElement, ContentWebElement, ContentWebFrame,
and ContentWebWindow.  The last three are direct representations of
web element, web frame, and web window definitions described in the
Webdriver specification.  The first is a custom Marionette type as
we also support retrieving XUL elements from chrome space and must
be considered proprietary.

Each of the classes extend the WebElement abstract type, which is
the primary entry point when unmarshaling JSON input from the client.
Based on the characteristics of the JSON Object, one of the different
concrete types will be constructed.

The purpose of this change is to make marshaling of elements and
WindowProxies easier, both when we receive web element reference
objects from clients and when transporting them over IPC internally.

The WebElement.fromUUID function should be considered a temporary
workaround until we have fixed the current Marionette clients to send
web element reference JSON Objects as input, instead of plain {id:
<uuid>, …} fields.

MozReview-Commit-ID: FGcRq5H1Tzp

UltraBlame original commit: 135c0568ca51fb1887dcddd0c02a4042099774fd