Bug 1381427 - Result width should often be constrained by the width of the location bar. r=mak
This adds an end margin to the location bar search results that matches the end of the input field, unless doing this would be unbalanced with the start margin, in which case a symmetrical margin is chosen. This allows the location bar to be moved to the start of the navigation toolbar to reclaim space for results. This also handles gracefully other customization scenarios like having many icons on the right side of the location bar.

MozReview-Commit-ID: 9TwieL0lqHB

UltraBlame original commit: 20d45a0a9a7bc8ef8159823e07fd0b2380bfb39a