Bug 1387808 - Switch to using XUL spacer elements for the zoom and edit controls' leading space in the app menu to fix auto-sizing issues. r=jaws
When the locale has a slightly longer label for 'Zoom', the popup auto-sizing
code starts having trouble with the inline start padding of that label and doesn't
size the popup correctly anymore.
When I change this to using a spacer element, this issue no longer occurs and the
label may become as large as localizers may need, as long as it doesn't exceed
the max-width of a panelview - 30em, currently.

MozReview-Commit-ID: CHRheMqazrj

UltraBlame original commit: 62617c6d5534053189a8497fa4705d5b0466e771