Bug 1410366 - Improve socket handling in wait_for_port r=ato
As the method name describes the primary functionality of this method
should be to wait for the specific port to be reachable. Specific
application verification is done anyway in the 'newSession' command.

Also the temporarily created sockets have to be removed immediately
by the operating system. Otherwise those enter the TIME_WAIT state,
and will be removed about 90s later. This can cause a pile of orphaned
sockets.

MozReview-Commit-ID: BGy9ZOjhjT0

UltraBlame original commit: 43e45144dd9fffa826b7b4d8289f77ddeea172c6