Bug 1359908 Add FORTIFY_SOURCE to gcc and clang builds by default r=froydnj
This adds a section in both old-configure.in's for hardening flags.

In old-configure.in we add $HARDENING_CFLAGS (which are turned on by
--enable-hardening) and are defined in toolchain.configure (and
which does compiler detection there.) We then add non-optional
hardening flags, performing compiler detection here.

In js/src/old-configure.in we follow the same pattern, but omit
$HARDENING_CFLAGS because we don't apply the current lone flag
to js (doing so is Bug 1359905).

MozReview-Commit-ID: EFE0Pc7yZHa

UltraBlame original commit: 057166cb35fe9a0df57adf0778539d64f2e550d2