Bug 1359677 - P1: Add a new mochitest to verify the bypassing service worker. r=asuth
Since we decide to do this by setting the nsIChannel::LOAD_BYPASS_SERVICE_WORKER
to defaultLoadFlags in the docShell and the implement of the bypass service
worker flag is done, only having this test to verify the behavior. So that we
can simplely copy the code for setBypassSW() and getBypassSW() to dev-tools
server in the future when we decide the UI/UX and designs on dev-tools clients.

UltraBlame original commit: 97db331fc5ac194f00fe3dd16573f0f46ab6852c