Bug 1411081 - Use opt-level=1 for rustc in default configuration; r=glandium
As the data in the bug shows, the current default of opt-level=2 is
several minutes slower at compiling than opt-level=1. This slows down
builds significantly and the added benefits of running opt-level=2
for local development can't be justified for the common/default case.

This commit changes the default for local builds from opt-level=2 to
opt-level=1.

--enable-release (what we use for builds shipped to users) will imply
opt-level=2. --enable-optimize (the default) will use opt-level=1,
and --disable-optimize will use opt-level=0.

The RUSTC_OPT_LEVEL environment variable in mozconfigs can be used
to set an explicit opt-level level, regardless of what other
configure options are set. This includes the other potential values,
"s" and "z."

A side-effect of this change is that -Copt-level is now *always*
specified by the build system. Before, it was only specified if
the value was adjusted to 0 for --disable-optimize builds.

MozReview-Commit-ID: 67KX5qScnFc

UltraBlame original commit: 64a852e835b7469715fe13dc76b12458e12bb5cb