Bug 1410236 - Deny permission by default if doorhanger is dismissed; r=sebastian
Add a "defaultCallback" option to the doorhanger API that specifies a
callback to call when the doorhanger is dismissed without a button being
clicked. Use that to deny a permission in ContentPermissionPrompt.js
when the doorhanger is dismissed without explicit action.

MozReview-Commit-ID: 9kOAWirI4Ux

UltraBlame original commit: ef22d58e750f8b24ceac4c93065d137aa939c6d0