bug 1412690 - add fennec release bouncer sub. r=rail
This patch adds the first releasetask as a new kind. To support this, we
added a new `release-promotion` flag in the buildbot job. If this is
set, we use the new `bb_release_worker` function; otherwise we fall back
to the `bb_ci_worker` function (this is the old behavior, factored out
into a separate function).

We also added `build_number` and `release_promotion` attributes in the
task definition.  Finally, `build_number` now defaults to 1, allowing us
to create the task graph locally without forcing us to set
`BUILD_NUMBER` in the environment.

MozReview-Commit-ID: 8vNMHJemqAG

UltraBlame original commit: a7f43d6357e668a21101c822d9694ec178702b05