Bug 1411688 - Part 4: Make single-locale repacks agree with underlying Nightly builds. r=aki
There's nothing that makes sense in the existing setup; we're only not
getting bitten because the set of things that _do_ depend on all of
the flags that differ between the underlying Nightly builds and
single-locale repacks is small, and nobody has complained.  For
example, about:licenses probably does not include the Adjust SDK
license for single-locale repacks.

This patch series recompiles the Java code as part of each
single-locale repack, and that means the feature flags, etc, need to
be the same between the underlying compiled code (from the underlying
Nightly build) and the fresh Java compile.  This patch tries to
harmonize the two.

MozReview-Commit-ID: 230q7HuD1vV

UltraBlame original commit: 8ddf3257a8db317403f9be34e92a62569b4add06