Bug 1408919 - Rename and enable browser_webconsole_autocomplete_crossdomain_iframe.js; r=Honza.
This bug also renames the support file, and create a dedicated one for the child iframe.
An helper method is created to only wait for a single message with given text in it, as
well as returning an array of matched message in waitForMessages (we were returning the
result of setting the listener, which I doubt is useful for any test.

MozReview-Commit-ID: Gz7zqX825oR

UltraBlame original commit: ab8cb0873ccb66992d3fa34d7290125c910aee08