Bug 1408920 - Rename and enable browser_webconsole_autocomplete_in_debugger_stackframe.js; r=jdescottes.
We take this as an opportunity to add a jsterm helper function to set its value and
wait for completion, and use it in a test where we did the same thing.
This also copies the openDebugger helper from the old frontend head.js .
The test still uses the old debugger, since it will be fixed in a specific bug
for all the console tests.

MozReview-Commit-ID: AKB3CBY8GPv

UltraBlame original commit: 8823f33351025e25a949160bc23bae9fba9aac5f