Bug 1409570 - Ensure a pushed stream knows we've reached EOS. r=mcmanus
Previously, if a pushed stream was ended with a padding-only DATA frame
with the FIN bit set, and that stream didn't have a Content-Length,
there would be no way of knowing that the stream was finished. Now we
force-mark the push as complete if we hit a padding-only DATA frame with
the FIN bit set.

MozReview-Commit-ID: 7tk8x2FNgSj

UltraBlame original commit: 140a38df3d5639e8a15e2cb5f2fd35d844e44d67