Bug 1415283 - Increase timeout for tone events in RTCDTMFSender-helper.js; r=jgraham
With the current value of 150 milliseconds we get intermittent timeouts on our
test infrastructure. We are missing by only a small amount, e.g. 6 milliseconds
in Bug 1415283, but it seems safer to increase the timeout to 250 milliseconds so
that we don't end up revisiting this problem in the future.

This also fixes the error message to say milliseconds rather than seconds.

MozReview-Commit-ID: CqjTKZn33q8

UltraBlame original commit: dcf4f46ab589f77392b9a5e867c8cdc2fff99290