Bug 1108587 - Extending the grace period of AsyncShutdown and the nsTerminator for ASAN builds;r=gsvelto
Bug 1108587 seems to suggest that the grace period for nsTerminator is
too short for ASAN builds, which take much longer to shutdown. This
patch changes the grace period from 1 minute to 3 minutes, hoping that
this will be sufficient. Somewhere along the way, we also extend the
duration of AsyncShutdown, because that's the simplest way to do both
at once.

MozReview-Commit-ID: 28eWO5m6Wh3

UltraBlame original commit: 56e1ecb8bb58a8d26aa89367388ffa34911fe533