Bug 1416540 - Avoid unnecessary conversion to doubles in ActiveLayerTracker code. r=mstange
This code was written before the ScaleFactors() function was available on
the float-based gfx::Matrix, which I presume is why it was written to
convert the matrix to the double-based variant and store the scale as
doubles. We don't need to do that any more.

MozReview-Commit-ID: EoWLpny8g61

UltraBlame original commit: 9785c2745d8bded936d781bc9797b71a3dc33fef