Bug 1419426 - Implement browserSettings.contextMenuShowEvent, r=kmag
This new API exposes the ui.context_menus.after_mouseup preference to extensions
so mouse gesture add-ons can modify it.

Note that this is not supported on Android, and also calling it with a value of
"mousedown" on Windows is a no-op.

MozReview-Commit-ID: AkhRmAyzuSp

UltraBlame original commit: e8d475fed2068feacb781d712a667e07ac4e5b9c