Bug 1421242 - Fix netmonitor SEND_CUSTOM_REQUEST reducer; r=Honza.
The reducer code was still trying to use Immutable.Record.get, although
we switched the state to a native js object. This was throwing an error.

Adding assertion to the existing tests that was failing without the fix.

MozReview-Commit-ID: Klj94ApJr4Y

UltraBlame original commit: ee838fa888aca9f77a8503b5411372d8b56e3341