Bug 1420104 - rabaldr, clean up temp register managements and some ifdefs.  r=bbouvier
A few ifdefs that were on platform names were really supposed to be on
JS_PUNBOX64.

Changes temp register management so that functions in the platform
layer now allocate and return the temp registers when appropriate and
return invalid registers when not appropriate.  This cleans up a bit
of code in the platform-independent layer.

MozReview-Commit-ID: K69Bj1gUNbZ

UltraBlame original commit: 18d399f9bb570d683d955106af88c0afd3039591