Bug 1422915 - Be explicit about remote-active data when shutting down. r=whimboo
When Marionette starts and shuts down it emits the remote-active
system notification.  On starting the data is set to true (a boolean),
but on shutdown it is left undefined which defaults to null.

We should use false (boolean) explicitly.

MozReview-Commit-ID: Er0Wn2pZrOi

UltraBlame original commit: d61481e31a93819cb54e07525294c8b42b8f7e90