Bug 1417232 - Part 3: Extract magic Gradle configuration from |mach android *|. r=ted.mielczarek
This pushes all of the magic Gradle targets into gradle.configure,
which is the most accessible central place for them.  Some impact the
build, so they almost certainly need to be in moz.configure; and its
better to have one central place to update than many places.

Gradle has a notion of configurations.  Fennec determines many
configurations, only one of which is used in automation; and right
now, that one is "officialPhotonDebug".  Evolving the "one true"
configuration is frustrating, and this helps with that.  Post
Android-Gradle plugin 3.0+, we can do better and also extract the
configuration details for the :geckoview and :geckoview_example Gradle
projects, but right now they are ad-hoc and can't really be improved.

MozReview-Commit-ID: LdGE05zn2H1

UltraBlame original commit: 3acb2fd4adacc4d89e26375ab9ee1416da7d901a