Bug 1393817 Fix L10N check for MinGW build r=ted
exe_7z_archive.py runs during the MinGW build L10N check step, and
hardcodes 7z as the 7zip executable. This works on Windows, but not
Linux. We need to pass it the correct executable for 7zip, which is
located during configure.

However, repacks (repackage-winXX-nightly) don't do configure, and
don't have the tools to do configure. So we leave the default
command in the python script if one is not supplied.

MozReview-Commit-ID: B7GEKRYEJTD

UltraBlame original commit: 988a2837b8dc9c7e423ef7c0c8d91dbb6b1769df