Bug 1421701 - Chunk visits when notifying main thread r=mak
Since processing several hundred visits synchronously on the
main thread can be somewhat expensive, and since the main thread
might be idle while the storage worker is running, it makes sense
to chunk the messages into groups of 100.

We did have a concern that this changes the move of the underlying
array to a copy, which might bring some cost with it. For small
arrays, we simply move the underlying array to avoid the copy to
mitigate it. However, for large arrays I was unable to observe any
significant performance cost anyway, so I think we're in the clear.

MozReview-Commit-ID: 1hWSEyKw6pi

UltraBlame original commit: 15a13cdc34d5dbbf361ddfc6d246436935222f10