Bug 1398963 part 2: Make nsFlexContainerFrame label itself as legacy if it has -moz-box/-moz-inline-box display val. r=mats
To be clear, this is a "paving the way" patch.  At this point in the patch
series, it's not yet possible for us to generate a nsFlexContainerFrame that
has display:-moz-box.  (A later patch in this series will make that possible.)

This patch adds the mechanics to nsFlexContainerFrame instances so that they'll
label themselves appropriately (with NS_STATE_FLEX_IS_EMULATING_LEGACY_BOX)
once it *does* become possible for -moz-box to spawn a nsFlexContainerFrame.
Moreover, this patch updates the state bit's documentation to reflect its new
potential-usage.

MozReview-Commit-ID: ElApieVoTLf

UltraBlame original commit: 462e7346039f17b609f5af67d4fee7217253906d