Bug 1414204 Part1: Suppress input events when there is a dnd session. r=smaug.
There may be some pending input events in the queue of thread when content starts a dnd operation. Spec says that input events should be suppressed when there is a dnd operation. Add a flag in ESM and turn on/off when start/finish a dnd operation. Checking the flag in PresShell::HandleEvent because we may start a dnd operation with pointermove and we want to suppress the mousemove as well.

MozReview-Commit-ID: 43NZrA7SW4c

UltraBlame original commit: 61e28d94b9bdae9919431bc6176374a4b5cfa4e4