Bug 1426291 - Add Rust style mutex template which wraps data with the mutex that synchronizes it. r=jwwang
Rust's std::sync::Mutex has some nice properties. Its philosphy is to lock
data, rather than code.

It wraps data you're trying to make thread safe with a mutex, and in order to
get a reference to the wrapped data you need to lock the mutex and access it
through an intermediate layer. This is good, as the mutex that's protecting
access to the data is explicitly associated with the data, and it's impossible
to forget to take the lock before accessing the data.

This patch adds a similar mutex wrapper to Media Playback code. If it works
well, we can look at moving it into xpcom.

MozReview-Commit-ID: 4APAic6Fh8m

UltraBlame original commit: 4ae5f2caa47120e4247d7d09f11fcc4f77357cd8