Bug 1426553 - Autoconf-related changes to spidermonkey jobs. r=nalexander
On Debian, the autoconf binary is autoconf2.13 while it is autoconf-2.13
on Centos.

In make-source-package.sh, we need to run autoconf to generate the
old-configure to include in the package, so try both.

In hazard-analysis.sh, we actually don't need autoconf itself, so just
copy configure.in to configure.

UltraBlame original commit: 62aa4a2039b99cc7919fdbc594ce7785faf1015b