Bug 1414768 - Handle same-compartment wrappers in TypedArray methods. r=bz
CallTypedArrayMethodIfWrapped (and the CallNonGeneric machinery throughout
the engine) unwraps the `this` argument, but the other arguments are only
rewrapped for the target compartment.

The pattern being used before this patch to get the length of a TypedArray
or possible TypedArray wrapper is:

    callFunction(CallTypedArrayMethodIfWrapped, O, O, "TypedArrayLength")

The first O is the `this` value and the second is an argument.
If O is a cross-compartment wrapper, this works fine. The first O is unwrapped,
revealing the actual TypedArray object; the second O is rewrapped for that
TypedArray's compartment, producing the same TypedArray.

However, if O is a same-compartment wrapper, this doesn't work. The first O
is unwrapped, revealing the actual TypedArray object in the same compartment;
rewrapping the other O does nothing to it, since it is already an object in the
target compartment. Thus TypedArrayLength receives a `this` value that's an
unwrapped TypedArray, but an argument that is still a wrapper.

The fix is to have CallTypedArrayMethodIfWrapped targets only expect `this`
to be an unwrapped TypedArray.

UltraBlame original commit: 8bb6aeb63d5d3c5b9e2a084ff78b82f82b6a7f28