Bug 1409852 - nsContentUtils::ExtractErrorValues with nsAString;r=bz
nsContentUtils::ExtractErrorValues converts `aSourceSpecOut` from
UTF-16 to UTF-8, which is generally not what we want. This patch
introduces a new version of ExtractErrorValues that doesn't perform
the conversion. To keep the patch short and avoid rewriting existing
clients, the existing version of the function is left in place.

MozReview-Commit-ID: J2NQb5ZCfht

UltraBlame original commit: 99508707de1ed3de9010677e39416661f3d6847b