Bug 1420928 - BuildKeyframes takes animation name and timing function instead of StyleAnimation. r=boris
This is a prerequisite to use repeated values for the used value of the each
animation property instead of the computed values.

In a subsequent patch in this patch series, the timing function will be replaced
with the one which may be a repeated value in computed animation timing function values list.

MozReview-Commit-ID: GOKHE3MH0Ue

UltraBlame original commit: 7d0c3d617c4d50415e9a900f9412a2cea143fe4b