Bug 1426501 - Change nsIURIMutator to call set spec on cloned URI if available r=mayhemer
Calling SetSpec on an nsIURI object doesn't reinitialize the object, meaning
it's not equivalent with creating a new URI with the same spec. While this
might be counter-intuitive we want to preserve existing behaviour for
the moment.
This change makes BaseURIMutator::InitFromSpec call SetSpec on the existing
cloned URI if available. Otherwise it creates a new one.

MozReview-Commit-ID: LuHVRhBItiP

UltraBlame original commit: a3baf4893b8db0c291002b62570c5fefaeb0b3b9