Bug 1428014 - Fix intermittent on browser_webconsole_close_sidebar.js; r=bgrins.
The test was failing because it still had pending request to the server while shutting
down the toolbox. This request was probably the one to retrieve the properties of the
object we put in the sidebar.
By waiting the object properties to be fully loaded when opening the sidebar, we should
make the intermittent go away.

MozReview-Commit-ID: 7tY8hvbDVRp

UltraBlame original commit: a9c97c2aa4b9551ca53a3b22163c0d0ddf784b3d