Bug 1381463 - Reload to clear UA when exiting RDM. r=ochameau
When exiting RDM with a device applied, we clear the UA override, but we did not
force a refresh of the page.  This means that the customized UA was still
accessible via `navigator.userAgent` and the page might have a non-default
layout if it tests the UA value.

To ensure there's no chance of confusion after exiting RDM, force a reload if
clearing emulation values determines it is needed.

MozReview-Commit-ID: Fkji12Utmis

UltraBlame original commit: bf43b5b793d6cfd4fd09e34150575959a3ef4980