Make a CaptureTiledPaintState for each tile (bug 1425056, r=bas)
This makes it so that each tile of a paint gets a DrawTargetCapture and
its own buffer operations. Once this is done, each CaptureTiledPaintState
will be isolated from each other and able to be done in parallel.

MozReview-Commit-ID: BuBDXgjma4z

UltraBlame original commit: 1cbfd11a64c8e8c790eae0852fa3e492f40282c8