Bug 1429140 - Make signal handling work for --disable-ion builds. r=luke
Signal handling is improperly ifdeffed on JS_CODEGEN_NONE; the
appropriate conditions are whether the signal handling machinery knows
how to grab the machine state, and that is not JIT-dependent.  (The
Atomics machinery has the same constraint.)  So change the ifdefs
accordingly.

Also fix a bug in jsfun.cpp where it calls into the JIT even if the
JIT is not present; this causes crashes.

UltraBlame original commit: 16740cf93a77c1f8ba6d39cf5748e0180e7e95dc