Bug 1430309 - Remove classycle from the tree. r=froydnj
This was added in Bug 1096627 to enforce the baseline GeckoView
layering that we had at that time.  Now that GeckoView is a separate
Gradle project, that layering is automatically enforced.  It's time
for this to go.

MozReview-Commit-ID: Ly35QhgBdWM

UltraBlame original commit: 0b0cab4592d9e15e0ae68ded75838993c7bd9f7e