Bug 1430077 - Separate clearing session state from deregistering listeners. r=whimboo
With the forthcoming window tracking changes the message listeners of
the content frame script are left listening for the duration of the
Marionette lifetime, and not for the duration of the Marionette session.

To prepare for the window tracking refactoring, we will want to
remove message listeners separately from clearing the session state.
Functionally, this patch changes nothing in Marionette for the moment,
except we send two IPC messages to the frame script instead of one.

MozReview-Commit-ID: DwVBCpvk9V9

UltraBlame original commit: 659526d528ff2dc3291b2140d2190d73d16b11e4