Bug 1425541: Expose reader mode state on the browser accessible using a "reader" object attribute. r=eeejay,Gijs
This makes it easy for accessibility clients to retrieve the reader mode state programmatically.

There are three possibilities:
1. Reader mode is available for the current page (reader:available).
2. Reader mode is being used now (reader:active).
3. Reader is not available (the reader attribute is not present).

We do this by setting/removing the aria-reader attribute on the node.
This is not a real ARIA attribute, but it causes Gecko to expose it as an object attribute.

MozReview-Commit-ID: B38G3AYyBnS

UltraBlame original commit: 196006c746bde8b894252ec1310a4b5124e04c14