Bug 1430844: Make resize reflow sound re. viewport units. r=bz
In particular, we set the pres context visible area _before_ processing
restyles.

This causes inconsistencies when resolving viewport units. In particular, the
resulting style tree will have some units resolved in terms of the old size, and
some in terms of the new size, depending on whatever is dirty, because we don't
flush the pending media query changes.

Also, some sizes are resolved against the unconstrained size because of the
shrink-to-fit stuff. Fix this by flushing _before_ in this case, instead of
after, since we're going to set the size to an actual value later when reflowing
the root frame.

MozReview-Commit-ID: ExI5yTJCjGp

UltraBlame original commit: 7317de4672a82d6ddca0c79a36b717b627a77a2d