bug 1419756 - deactivate pytest cache plugin to stop it writing files in the srcdir. r=davehunt
The pytest cache plugin writes its cache in the srcdir, which means that it
shows up in `hg status`, which is annoying. Writing files to the srcdir is
generally bad practice anyway, so we disable this plugin to stop this
from happening.

MozReview-Commit-ID: HytLLMUtKlc

UltraBlame original commit: 99a87b01a21f5dd73520e05a461e65dadba90dd6