Bug 1429056 - Wrap llvm-dsymutil calls on automation. r=ted
We add a wrapper for llvm-dsymutil for macosx CI builds such that when
it crashes, we attempt to get a reduced test case and upload it as a
build artifact. This will allow to more easily report such crashes
upstream.

UltraBlame original commit: 2082870a97c2efd93d0dfecfe9aac4c052eb129a