Bug 1431474: Make AccessibleCaretManager flush notifications through the document. r=bz
We're crashing because we're trying to flush layout of a shell whose document is
in the bfcache, so when we try to get the shell from the document in the style
set we get null back.

This shouldn't generally happen, and all flushes should go through the document
instead to avoid bugs like this and also other bugs related to subdocuments, and
all the stuff that nsIDocument::FlushPendingNotifications handles but the
presshell version doesn't. That's bug 1431852.

MozReview-Commit-ID: 1IJ4DBa3DCO

UltraBlame original commit: 00ae6912038d66987b4b10022904a2e8bb7e970d