Bug 1432212 - Lazily load log service in Marionette. r=maja_zf
We do not have to load the log service in Marionette until we
actually use it.  This patch uses a lazy loader for the Log.jsm
module and a lazy getter for initialising the Marionette log repository.

This means we can get rid of MarionetteComponent#setupLogger.
The repository's log level is set at the same point in time as before.

MozReview-Commit-ID: K5PGqBB6JEO

UltraBlame original commit: a496e8ddbe5639c6a69536264999c05775d9939a