Bug 1432804 - Remove disabled attribute from Element Clear test. r=automatedtester
<select> is not a mutable form control element and it should not
be possible to call the WebDriver:ElementClear command on it.
We test this in test_select as part of element_clear.py in WPT,
but the test uses a <select disabled> element which triggers the
wrong code path in Marionette.

MozReview-Commit-ID: 4ybEdLY6OEo

UltraBlame original commit: 485d1b8f19485869f6ad704943009c0c5b093b97