Bug 1432813 - Remove warning on "./mach marionette-test". r=maja_zf
It is not possible to hook "./mach test" up to "./mach marionette test"
because the test flavours system does not allow delegation to subcommands.

Users find it annoying that using "./mach test" causes a warning
to be emitted.  Considering we will not be able to remove "./mach
marionette-test", this patch instead quenches the warning.

MozReview-Commit-ID: Ia0ukcN3JMJ

UltraBlame original commit: 8f503aeea177bcae52e66cf7409311874167e459