Bug 1432771 - Clear queues in dispatchMessagesClear; r=bgrins.
In the console, to batch messages addition, we use queues which are
then consumed once every 50ms. But if a message is pushed to a queue,
and dispatchMessagesClear is called before the 50ms interval, the message
would still be added to the output.
This is happening for example when calling the `clear` helper from jsterm.
Clearing the queues when calling dispatchMessagesClear fixes the issue.
A mochitest is added to make sure we don't regress this.

MozReview-Commit-ID: 2CI0yIGSrT

UltraBlame original commit: 37cd2d413696175be0a6ea9359feb1aa3c425dd0