Bug 1433497 - Include new Marionette prefs in Python client. r=whimboo
Some time ago the Marionette server renamed a couple of preferences to
be more in line with the naming scheme of other preferences in Gecko.
Unfortunately we forgot to update the Marionette Python client to
include these.

This patch does not remove the older preference names because there
is no need to unnecessarily break the client when it is used with
older Firefox versions, although in theory it is now safe to remove them.

MozReview-Commit-ID: 8gyFfgd2VkX

UltraBlame original commit: 9ac99a988ffa144a3bd29bf7d108811a1ea37c22