Bug 1427950 - Add re-render button to PaymentRequest debugging panel. r=jaws
This is useful to ensure that the UI state is properly represented in the store.
If the re-rendering causes visible changes then that means some state is not part
of the store or is not being rendered properly.

MozReview-Commit-ID: 3J5rKWCTMAj

UltraBlame original commit: 6a159c9b17738bb663f04a5be34af34c63af9c04