Bug 1432341 - Add reftest for column-rule broken across columns.  r=mattwoodrow
I tested locally that the reftest fails without the previous patch, and
passes with it.  The failure is because the column rules overflow (as
though they're drawn in each column for the entire extent of the inner
columnset rather than just the piece visible in that column, i.e.,
overflowing only below in the first column, above and below in the
second column, and only above in the third column).

MozReview-Commit-ID: 4bDcnYiXEQt

UltraBlame original commit: ded3e23e2442aedea6392ceb02fe4b76b4b47f8c