Bug 1434922 - Set a favicon for about:privatebrowsing. r=florian
This was initially removed when we added support for setting the favicon
ahead of parsing the page for newtab-like pages. Turns out we need it in
some scenarios, such as navigating back and forth or duplicating the tab.

MozReview-Commit-ID: HuMeEnqpLLB

UltraBlame original commit: 5f18850383870618cec8c9b734c00cc9a9b8ffb6