Bug 1435282 - Add a helper FlushPendingComposite function. r=sotaro
This extracts a code pattern that appears a couple of times in the code.
It occurs in CompositorBridgeParent as well but there's some extra stuff
involved there with the mOverrideComposeReadiness flag that I don't
understand so I'm leaving that as-is for now.

MozReview-Commit-ID: 2xqgaQZT4e1

UltraBlame original commit: ed13e2aac8f6e7990164f4a2e647f0b12b5fe655