Bug 1429593 - Part 2: Show that a WebExtension is managing the proxy config setting, r=jaws,mstriemer
Update the general page of about:preferences, as well as the Connection Settings panel, to show
when an extension is controlling proxy settings, and allow a user to disable the extension to
regain control.

This also updates some of the strings to use the phrase "is controlling" rather than "controls".

MozReview-Commit-ID: HKYPkg78IOK

UltraBlame original commit: 9dfaa0724a424dfa146159d388f8fbd0a70bce28