Bug 1434883 - Part 1 - Fix tests for the activated page action panel. r=Gijs
The promisePanelEvent function was unreliable because it did not raise an error if the provided panel did not exist, which caused one of the callers to ignore a missing panel silently. All the callers have now been updated based on whether they expect the panel to exist or not.

MozReview-Commit-ID: AGT4rHls4OB

UltraBlame original commit: 9abfd413e1d55f4ee8de2d62d6f9cdb87113b555