Bug 1435596.  Give cross-origin objects (windows and locations) a 'then' property in all cases.  r=bholley
For locations, it always returns undefined.  For windows, it returns undefined
unless there is a named subframe named "then", in which case that named
subframe is returned.

The idea is to be able to resolve promises with cross-origin objects.

MozReview-Commit-ID: HPyTvtwFdsG

UltraBlame original commit: 1f3a5b496acd2288cc8cf0c32af86cb35157ea4e