Bug 1436690 - Fix React 16 warning in NetworkMessage component; r=bgrins.
In this component we had an <a> element as a child of another <a> element,
which isn't valid. Furthermore, we didn't make use of this <a>, so switching
to a span brings us validity without any drawbacks.

MozReview-Commit-ID: 3EM8jkEtSKI

UltraBlame original commit: bb5aaa6697ec6169fcc8ba1acb35c10bbd33fb4e