Bug 1435569: Change CacheIRSpewer to allow less redundency in spewing. r=tcampbell
This patch adds an inner class to CacheIRSpewer that manages access to the
CacheIRSpewer and dramatically simplifies the consuming code.

Note: this changes the CacheIRSpewer to no longer use LockGuard; instead
the raw mutex is managed by CacheIRSpewer. This is because the RAII nature
of CacheIRSpewer::Guard prevents constructing the LockGuard if-and-only-if
the spewer is enabled.

UltraBlame original commit: 114d2a3202c0ddf706461952d63a14d35611f734