Bug 1438866 - Mark empty filenames not interesting. r=emilio
In some cases we have SourceLocation objections that are isMacroId()
rather than isFileId() and so don't have a filename at all. In other
cases the filename is something clang-internal like "<scratch>". In
both of these situations we don't want to output any analysis data.
However, the code previously was taking the empty filename and resolving
it relative to the source folder, and then tried to write to the
source folder or create a file corresponding to a subfolder. This
resulted in general badness. This patch makes sure we ignore analysis
from places that don't have a corresponding file so that we don't
have this problem.

MozReview-Commit-ID: 7WluygY9Uy8

UltraBlame original commit: d071f1a3ace473f5d5c292ae7a704da892165107