Bug 1438542 - Fix rendering of non-cached message issued from logStringMessage; r=miker.
We weren't doing the right transform for cached messages, and thus
we weren't creating the ConsoleMessage with the correct properties.
This is now fixed and a mocha test is added to make sure we don't
regress on this.

MozReview-Commit-ID: ACOCPF06GK2

UltraBlame original commit: 58cf8ba79e970d40cec74c67a82ea923cdb4d48d