Bug 1259168 - Resolves the table cell misplacement problem when row group width exceeds table's computed width in vertical-rl writing mode r=jfkthame
A table's used width may be different with its computed width. In the past, this
situation wasn't taken into account in process of table reflow, this patch
corrects the positioning of children of a table in this situation.

MozReview-Commit-ID: 4siu2ba0pqG

UltraBlame original commit: a7df35ca198189acd9d608773c24d628df905b39