Bug 1432232 - Implement prefix support in new console frontend; r=bgrins.
The prefix property was already on the packet, so all we had to
do was to support it. Which means:
 - retrieve the property in the packet and pass it to the message
 - when prefix exists, add a dom element with a custom style
 - search in prefix when filtering.

We add a couple of mocha tests with a stub packet to make sure we
handle those cases like we should.

MozReview-Commit-ID: KFkt685AdfE

UltraBlame original commit: d8ffc9e58ff1256cad9a92a04e23d8a4e2b8dade