Bug 1291387 - Reply with an error in GeckoSessionHandler when no listener registered r=jchen
This eliminates a race where the JS side thinks it has a listener, and
expects a reply, but will never get one because it was unregistered
while the message was in-flight. GeckoSessionHandler dispatches
on the Android UI thread, which is where listeners are set/unset, so
we do not need any synchronization.

MozReview-Commit-ID: 5W3hsQ1cmb7

UltraBlame original commit: 137940af492428e84d2acf13036dea8106443a4d