Bug 1439358 - Part 5 - Handle panel hiding during ViewShowing events. r=Gijs
The ViewHiding event is now dispatched consistently, regardless of whether the ViewShowing event is canceled or the panel is closed during the event. This is done by a new _openView helper, while the logic that is specific to each of the showMainView, showSubView, and goBack functions has been moved out of the _showView function.

MozReview-Commit-ID: 5WvW6THWbyb

UltraBlame original commit: b817cfb3367f9ee3a8bcb6ca8d7c3e7c68b24847