Bug 1440613: Flush pending media feature values changes before initializing the shell. r=dbaron
This is guaranteed not to run script because of:

  https://searchfox.org/mozilla-central/rev/47cb352984bac15c476dcd75f8360f902673cb98/layout/base/nsPresContext.cpp#2152

This doesn't have any effect on the top-level chrome window, but avoids a full
document restyle on XUL popups on startup.

With this patch, there are no more full document restyles due to media feature
changes on my machine on startup.

MozReview-Commit-ID: HLkhIsIIa2W

UltraBlame original commit: bd876f6cfc077e29607949c1d06bfceca0fcee03