Bug 1330529 - Part 1: Fix PDFium gtest by using absolute path to PDFium DLL. r=jimm
This gtest assumed the current working directory ($OBJDIR/_tests/gtest/) was in the DLL LoadLibrary path. That is no longer true with this bug's patches, so the gtest must specify the absolute path to the PDFium DLL to load it.

MozReview-Commit-ID: 5TXj6A9Tb9w

UltraBlame original commit: 875a35f50af245bb7c4429c07c36828c845ade14