Bug 1398845 - Handle null mBuffer in PaintedLayerComposite::IsOpaque r=mstange
In some cases we can end up with a PaintedLayerComposite still present
in the layer tree but with its mBuffer cleaned up. Whether this should
be occurring or not is still under discussion, but nevertheless a
PaintedLayerComposite without its mBuffer should not be considered
opaque, and treating it as opaque results in a black flash on OSX when
following links from the activity stream.

MozReview-Commit-ID: KklHEivfBWZ

UltraBlame original commit: fcde688c911e2cf602516a3a67097f0a29ae9354