Bug 1441974 - Avoid duplicate clicks in sidebar toggle. r=gl
After the React monkeypatch removal (bug 1420130), some events can be duplicated
because both privileged and non-privileged versions are dispatched.

To work around this, we can stop propagation for the event in cases like this
where it is safe to do so.

MozReview-Commit-ID: L6cHpuTw1D2

UltraBlame original commit: 853caeaad89b97ae57584eb1958850a3e44f89f9