Bug 1439657 Remove Time Precision Reduction from Console Timestamps r=baku
We currently reduce the precision of timestamps emitted by console.time
in the main context, but not in a Worker. Because web page content cannot
read the content of messages emitted to the console (right?), we don't
actually need to reduce the precision of this data.

MozReview-Commit-ID: EfpIEICy0tX

UltraBlame original commit: 828515186ebc425c586fccef80abadb147bac245