Bug 1441333: Part 4 - Add caller location to async API errors. r=zombie
When we report errors generated by async API operations, they are currently
not tied to any relevant extension source location. Instead, the locations in
the error reports point to the location in the framework code which reported
the error, which makes things extremely difficult to diagnose.

This change saves the extension caller location at the beginning of async API
operations, and adds it to the error reports genearted by that operation. This
should give extension developers a useful starting point for debugging them.

MozReview-Commit-ID: AkknkKEosFt

UltraBlame original commit: 47c2e679000ff75affdb8314a7ee83726411806e