Bug 1441333: Part 6 - Use caller location in error reports for StrongPromise errors. r=zombie
We currently report a useful location in error reports when extensions fail to
resolve a promise or call a response callback, but in some slightly
less-than-ideal ways. We currently generate a complete stack and parse its
string value (which is expensive), and then report the caller location as part
of the message, rather than as the error's location and stack.

This patch changes that behavior to store a single SavedStack frame, and to
properly report that as the location of the error.

MozReview-Commit-ID: Jmtf4C1O6pW

UltraBlame original commit: 930b44b7adc7e26e0b76cb46178489fc2e6704dc