Bug 1443592: Part 3: Change LiveSavedFrameCache::find to take a FramePtr, not a FrameIter. r=fitzgen
This obviates dynamic checks in find about the state of the iterator, since
they're covered by the fact that we were able to obtain a FramePtr at all.

This does mean that we have to pass the pc separately. But that's symmetrical
with the insert method, so it's okay.

MozReview-Commit-ID: FgsDjHB2il4

UltraBlame original commit: 269e77f6b1e5878062f9b4a5e12533cc20da01d0