Bug 1443948 - Remove NODE_FORCE_XBL_BINDINGS;r=bz
This was used to attach a binding to a cloned node before it got inserted
into the doc. This is no longer used in the browser chrome, so this patch
removes the feature to prevent future usage and simplify dom code.

MozReview-Commit-ID: KnkHWJ8oQig

UltraBlame original commit: dbf853820066253e476d7fbcee35da534f9cc538