Bug 1445478: Make pointerevents/test_bug1414336.html more reliable. r=kats
There can be something shuffling the iframe between the mouse event is sent and
the mouse event is received which makes us end up targeting the <body> instead
of the test target.

This can be reproduced with enough persistence either with and without the
patches from bug 1439875, at least on a headless build with rr chaos mode.

Move it to test_group_pointerevents in the apz tests to run it in a new window
and make it a bit more reliable.

MozReview-Commit-ID: BS6Es8iEmMY

UltraBlame original commit: a749119ba1e40d926418c1fd4ae380c31d640115