bug 1444532 - fix a leak in SHA256 in nsHttpConnectionInfo.cpp r=mayhemer
The original code (from bug 1200802) declared an XPCOM object as a static bare
pointer, which for future reference is probably never the right thing to do. It
might have worked if it was cleared before shutdown but it never was.

MozReview-Commit-ID: EMe7wgzm6zv

UltraBlame original commit: ad1e07a06363096efbfbf115cbc274c4ee9dcd40