Bug 1446200 - Add a helper method to encode Method flags. r=froydnj
Most of the classes in xpt.py define a helper method to encode the
values of the flags into a byte, but for some reason Method does
not. It'll be useful to have one for Method for the C++ backend I am
working on.

MozReview-Commit-ID: ESi1CnstbN2

UltraBlame original commit: 1a3a72fa6dc5423e2d3ceeec3c517b8bc462e91a