Bug 1443492: Flush the document, not the shell, in cross-doc getComputedStyle situations. r=dbaron
We flushed the style of the original doc, but not layout, and thus the iframe
resize wasn't noticed, and the style flush on the child presShell wasn't
sufficient.

Do a style flush on the child document instead, so that it flushes layout on the
parent document too if needed.

MozReview-Commit-ID: 5ZhBuxpKIUg

UltraBlame original commit: 9cbafe1855ccac5fd2f0c2e549e830a66a72f84f