Bug 1441287 - [mozcrash] check_for_crashes should always return count of crashes. r=ahal
Right now if no minidump file is present in the minidump folder,
the check_for_crashes method returns False. Whereby in all other
cases the number of crashes is returned.

To be consistent this method should always return a number, and
in case of no minidumps it should be 0.

MozReview-Commit-ID: 3DTgxn41TVn

UltraBlame original commit: 5082e91767c230418f71d3888346414b7f99d665