Bug 1447914 - Change FunctionCall::getAtomName() to appendName(). r=froydnj
getAtomName() is only used in one place -- FunctionCall::toString() -- where
the returned atom is appended to a string. This patch changes it to
appendName(), which does the appending itself. This avoids a bunch of manual
atom refcounting and the possibility of failure.

MozReview-Commit-ID: I75rO1Bv6Y8

UltraBlame original commit: 3145ee5bb85c8cf09b65d88e86e8aba2949efac4