Bug 1436040 - xslt processor: <?xslt-param name= value= ?> isn't recognized. r=bz.
The patch for bug 1387427 made us keep the source content in an array and we don't insert
it into the document anymore. The code for xslt-param and xslt-param-namespace should look
in the source content array instead of the document.

UltraBlame original commit: 21edf57115997ad529c68f0ac8bfb45e151f8329