Bug 1436523 - Make WebRTC mochitest harness more granular with fake device setup. r=achronop
Update head.js so that logic for fake audio and fake video is more granular.
Now code handles fake audio and fake video separately, and will set the fake
pref is either is needed. This allows for a single loopback device to be set
and fake used for the other. Previously both devices needed to be loopback or
fakes would be used.

MozReview-Commit-ID: K4blmPrSVeh

UltraBlame original commit: b6cccecb4a6b0f628312b775ee415254275c7067