Bug 1448900 - Avoid killing exited process. r=jgraham,whimboo
std::process::Child::kill() will return Err if the process has
already exited.  The assumption in bug 1443853 was that calling
::kill() would consistently return the std::process::ExitStatus
was the process already dead.

This patches the regression from bug 1443853 by employing
Child::try_wait() in a loop.  When the process gives some exit status,
this is return directly without relying on Child::kill() as before.

If the process has not exited and the timeout has elapsed, we kill
the process and return its return value.  If the process has not
exited but the timeout duration has not elapsed, we wait 100 ms as before.

MozReview-Commit-ID: 4VENbrKtcEh

UltraBlame original commit: 3d25026d07ef61dcb7951a9afd4c3cd6d4f8af22