Bug 1053898 - Update framework/selection to support slotted selections;r=gl
With shadow dom support, a node front can be represented twice in the
markup view, once as a "slotted" version inserted in a <slot> element
and once a regular element.

We should be able to know if the selection was made on the regular or
on the slotted version.

MozReview-Commit-ID: CUpREkL0TzL

UltraBlame original commit: 3b3833597bd5b9b399151f017f6d5e98198fc7d4