Bug 1404222 Part 9: Fix wpt reftest shape-image-025.html to make every frame of the animated GIF use a green box in the lower-left quadrant.
This test is designed to check that the float area is calculated from the
first frame of an animated GIF. The reference rendering is a solid green
square with the lower-left of the square supplied by the pixels in the GIF.
Frame 2 of the GIF has white pixels in that quadrant. If layout is completed
and onload is fired while frame 2 is showing, the test will fail even though
the float area has been correctly calculated. This change puts green in the
lower-left quadrant of all frames of the GIF so that the test is only testing
the float area and not also testing that onload fires while a certain frame
is showing.

Note that this change doesn't impact the thoroughness of how we're testing
shape-outside because it changes only the color of the pixels and not their
opacity, and the float area is calculated from the alpha of the pixels.

MozReview-Commit-ID: LTGlrgbxRyT

UltraBlame original commit: 491be7851cfc3e4959af6e22a1adf70d5f7bb168