Bug 1449532 - Part III, Polyfill Web Animation API features r=Gijs
The Animation and KeyframeEffect constructors and the finshed promise are not
enabled on release channel currently. The polyfill is added to make sure
we don't break on release.

When the feature ships, removing the polyfill should be as easy as
reverting this changeset.

MozReview-Commit-ID: 2EWN7hBN5tj

UltraBlame original commit: 64486670492f5c9cc2e890c5931284bb6a85d194