Bug 1443942 - Make redirect SJS' serve with headers to prevent Necko caching. r=jya
Try to prevent Necko from caching the results of our SJS media responses, as
some of the test that use it rely on the server being hit and serving a
redirect. Sometimes the tests which rely on hitting a redirect in an SJS
where timing out without this, as Necko would cache the response and not
hit the server, and so not hit the redirect.

Also include a noise parameter to increase the likelihood that the URL is
unique, to further reduce the chance that Necko caches the result.

MozReview-Commit-ID: 3cLEiDoh4HG

UltraBlame original commit: ee486ce2d6604682d4f6fa5a3264120829ba2cbe