Bug 1452597 - Remove debug_script functionality for injected scripts. r=maja_zf
The WebDriver:ExecuteScript and WebDriver:ExecuteAsyncScript commands
accepts a "debug_script" parameter that attaches an error handler
on the WindowProxy in the sandbox.

This used to be necessary because the error handler used to be
attached to the content window instead of the sandbox.

MozReview-Commit-ID: ImRVkC5T75O

UltraBlame original commit: 785c5c3cf010fe02f235a3084834154097508275