Bug 1454038 - 3. Reject sendRequestForResult promise on finalize; r=esawin
When a sendRequestForResult event callback is finalized, automatically
reject the promise (if it has not been resolved yet). This means if the
callback is never called, or if the communication link is disrupted, we
will end up with a rejected promise rather than a forever-pending
promise.

MozReview-Commit-ID: A1w19vwFxuT

UltraBlame original commit: baf6063071b523d291fb61e8974034f1767c4c26