Bug 1453869 part 7.  Make the DOMParser WebIDL constructor use a nullprincipal for the document if the caller is system-principal.  r=mrbkap
In our test suite, we only run into two calls to this constructor with a system
principal, and both are in test code.

After this, calling the WebIDL constructor from system code is _almost_
equivalent to creating by contract.  The one difference is that the resulting
DOMParser (and the documents it creates) will have its script handling object
set to the global the constructor came from instead of being null.

MozReview-Commit-ID: Fe2yMeqoYnB

UltraBlame original commit: 1ad69430054b58bedc5fe9e9561982c1524dc312