Bug 1432894 - Register marionette.contentListener pref. r=whimboo
marionette.contentListener is being used for silly purposes in
Marionette to tell us when it is "safe" to reuse a content frame.
It will be removed with https://bugzil.la/marionette-window-tracking
but we should nonetheless keep track of which preferences are used.

MozReview-Commit-ID: 5gMCMuYfLBa

UltraBlame original commit: f9b53fc2863e2cbb7b3699bffbf8c0cb8b4409c0