Bug 1456677: Make the blocklist service a JSM, with an XPCOM service stub. r=Gijs
Aside from making things easier for JS callers, this also makes it harder to
accidentally trigger an early load of the service, which can be expensive
during startup.

This also makes a slight change to nsPluginHost to initially preserve the
previous blocklist state when a plugin is updated, to avoid the risk of the
possible additioanl asynchrony unblocking a plugin that should stay blocked.

MozReview-Commit-ID: 4EvIGJ1Ke0Z

UltraBlame original commit: 35fe513112642f7d2f9c99da61538db193701a94