Bug 1454445 - Put windowtype/positioning attributes on the webconsole html tag to support opening as top-level;r=mossop
We currently never open this as a top level window (it gets wrapped in browserconsole.xul),
but we are working towards it. Putting these attributes on the <html> tag works similarly to
putting them on the <window> tag in the XUL document and makes it easier to make progress
on missing features.

MozReview-Commit-ID: x6seka5wyf

UltraBlame original commit: 2adfda24e3cd057e07f98e410453da7bd41ec489