Bug 1451576 - Set dir=auto on [Browse...] button of <input type=file> r=Ehsan
The label of the button comes from the UI; it's direction should be independent
of the web content directionality. As there is no access to :-moz-locale-dir()
selectors in HTML, we'll have to rely on auto-direction here.

The alternative (or, addition to the fix here) would be adding the value of
the dir attribute to HtmlForm.properties, allowing the localizers to set
the directionality explicitly. I however don't know if that's necessary.

MozReview-Commit-ID: 5NXeLtxLXVH

UltraBlame original commit: 09f7885a83ce43f44209d938e982316067446e5e