Bug 1455745: Remove more accessibility related diagnostic crashes on Nightly. r=aklotz
This condition is rare and does indicate a problem which breaks accessibility.
However, we aren't getting any closer to diagnosing this as a result of this crash, so it causes user pain without any gain to us.

MozReview-Commit-ID: GacAQDbhB0h

UltraBlame original commit: c36098a8e83f0238196d0bfbaad1f81ae84ae28e